% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{adjcolors}
\alias{adjcolors}
\title{Convert a color with arbitrary transparency to a fixed color}
\usage{
adjcolors(colors, alpha)
}
\arguments{
\item{colors}{}

\item{alpha}{}
}
\description{
Convert a color with arbitrary transparency to a fixed color
}
