% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{adjust.to.2pi}
\alias{adjust.to.2pi}
\title{Wrap Angle to [0, 2pi)}
\usage{
adjust.to.2pi(x)
}
\arguments{
\item{x}{Numeric. Angle(s) in radians.}
}
\value{
Numeric. Angle(s) mapped to \code{[0, 2pi)}.
}
\description{
Wrap Angle to [0, 2pi)
}
