% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateFeaturesOverAssays-deprecated.R
\name{aggregateFeaturesOverAssays-defunct}
\alias{aggregateFeaturesOverAssays-defunct}
\alias{aggregateFeaturesOverAssays}
\title{Aggregate features over multiple assays}
\usage{
aggregateFeaturesOverAssays(object, i, fcol, name, fun, ...)
}
\description{
The \code{aggregateFeaturesOverAssays} function is defunct. Please use the
\code{aggregateFeatures} method from the \code{QFeatures} package instead.
}
