% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{alpha.adjust}
\alias{alpha.adjust}
\title{Alpha Adjustment for Multiple Testing}
\usage{
alpha.adjust(alpha, k, method = "bonferroni")
}
\arguments{
\item{alpha}{Original alpha}

\item{k}{Number of tests}

\item{method}{"bonferroni" or "sidak"}
}
\value{
Adjusted alpha
}
\description{
Alpha Adjustment for Multiple Testing
}
