% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{amp.cut}
\alias{amp.cut}
\title{Amplitude Cutoff Adjustment}
\usage{
amp.cut(amp, a.TOJR)
}
\arguments{
\item{amp}{Vector indicating if amplitudes exceed cutoff}

\item{a.TOJR}{Current TOJR category}
}
\value{
Adjusted TOJR category
}
\description{
Amplitude Cutoff Adjustment
}
