% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-feature_annotation.R
\name{AnnotateFeatures}
\alias{AnnotateFeatures}
\title{AnnotateFeatures}
\usage{
AnnotateFeatures(
  srt,
  species = "Homo_sapiens",
  anno_TF = FALSE,
  merge_tf_by = c("symbol", "ensembl_id", "entrez_id"),
  anno_LR = FALSE,
  merge_lr_by = c("symbol", "ensembl_id", "entrez_id"),
  gtf = NULL,
  merge_gtf_by = "gene_name",
  attribute = c("gene_id", "gene_name", "gene_type"),
  assays = "RNA",
  overwrite = FALSE
)
}
\arguments{
\item{anno_TF}{Whether to add the transcription factor/cofactor annotation.}

\item{merge_tf_by}{"symbol", "ensembl_id" or "entrez_id"}

\item{anno_LR}{Whether to add the ligand/receptor annotation.}

\item{merge_lr_by}{"symbol", "ensembl_id" or "entrez_id"}

\item{merge_gtf_by}{}
}
\description{
AnnotateFeatures
}
\examples{
data("pancreas_sub")
pancreas_sub <- AnnotateFeatures(pancreas_sub, species = "Mus_musculus", anno_TF = TRUE, anno_LR = TRUE)
head(pancreas_sub[["RNA"]]@meta.features)

## Annotate features using a GTF file
# pancreas_sub <- AnnotateFeatures(pancreas_sub, gtf = "/data/reference/CellRanger/refdata-gex-mm10-2020-A/genes/genes.gtf")

}
