% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_matrix}
\alias{as_matrix}
\title{Attempts to turn a dgCMatrix into a dense matrix}
\usage{
as_matrix(matrix)
}
\arguments{
\item{matrix}{A dgCMatrix}
}
\description{
Attempts to turn a dgCMatrix into a dense matrix
}
