% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{blendcolors}
\alias{blendcolors}
\title{Blend colors}
\usage{
blendcolors(colors, mode = c("blend", "average", "screen", "multiply"))
}
\arguments{
\item{colors}{Color vectors.}

\item{mode}{Blend mode.}
}
\description{
Blend colors
}
\examples{
blend <- c("red", "green", blendcolors(c("red", "green"), mode = "blend"))
average <- c("red", "green", blendcolors(c("red", "green"), mode = "average"))
screen <- c("red", "green", blendcolors(c("red", "green"), mode = "screen"))
multiply <- c("red", "green", blendcolors(c("red", "green"), mode = "multiply"))
show_palettes(list("blend" = blend, "average" = average, "screen" = screen, "multiply" = multiply))

}
