% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalizes the characters
Making the first letter uppercase}
\usage{
capitalize(x, force_tolower = FALSE)
}
\arguments{
\item{x}{A vector of character strings to be capitalized.}

\item{force_tolower}{Whether to force the remaining letters to be lowercase.}
}
\description{
Capitalizes the characters
Making the first letter uppercase
}
