% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{CC_GenePrefetch}
\alias{CC_GenePrefetch}
\title{Prefetch cycle gene}
\usage{
CC_GenePrefetch(
  species = "Homo_sapiens",
  Ensembl_version = 103,
  mirror = NULL,
  max_tries = 5,
  use_cached_gene = TRUE
)
}
\arguments{
\item{species}{Latin names for animals,i.e., "Homo_sapiens", "Mus_musculus"}

\item{Ensembl_version}{Ensembl database version. If NULL, use the current release version.}

\item{mirror}{Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'.}

\item{max_tries}{The maximum number of attempts to connect with the BioMart service.}

\item{use_cached_gene}{Whether to use previously cached cell cycle gene conversion results for the species.}
}
\value{
A list of S-phase and G2M-phase genes.

\code{\link{GeneConvert}}
}
\description{
Based on the human cell cycle genes, the cell cycle genes of the corresponding species were captured by homologous gene conversion.
}
\examples{
ccgenes <- CC_GenePrefetch("Homo_sapiens")
str(ccgenes)
ccgenes <- CC_GenePrefetch("Mus_musculus")
str(ccgenes)
}
