% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{CellCorHeatmap}
\alias{CellCorHeatmap}
\title{CellCorHeatmap}
\usage{
CellCorHeatmap(
  srt_query,
  srt_ref = NULL,
  bulk_ref = NULL,
  query_group = NULL,
  ref_group = NULL,
  query_assay = NULL,
  ref_assay = NULL,
  query_reduction = NULL,
  ref_reduction = NULL,
  query_dims = ref_dims,
  ref_dims = 1:30,
  query_collapsing = !is.null(query_group),
  ref_collapsing = TRUE,
  features = NULL,
  features_type = c("HVF", "DE"),
  feature_source = "both",
  nfeatures = 2000,
  DEtest_param = list(max.cells.per.ident = 200, test.use = "wilcox"),
  DE_threshold = "p_val_adj < 0.05",
  distance_metric = "cosine",
  k = 30,
  filter_lowfreq = 0,
  prefix = "knnpredict",
  force = FALSE,
  rows = NULL,
  columns = NULL,
  cluster_columns = TRUE,
  cluster_rows = TRUE,
  nlabel = 3,
  label_cutoff = 0,
  label_by = "row",
  label_size = 10,
  gird_size = unit(8, "mm"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
CellCorHeatmap
}
