% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{CellDensityPlot}
\alias{CellDensityPlot}
\title{CellDensityPlot}
\usage{
CellDensityPlot(srt, along, group.by, split.by, decreasing = FALSE)
}
\description{
CellDensityPlot
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
pancreas_sub <- RunDynamicFeatures(pancreas_sub, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
CellDensityPlot(pancreas_sub, along = "Lineage1", group.by = "SubCellType")
CellDensityPlot(pancreas_sub, along = "Lineage2", group.by = "SubCellType")
}
