% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{CellDimPlot3D}
\alias{CellDimPlot3D}
\title{3D-Dimensional reduction plot for cell classification visualization.}
\usage{
CellDimPlot3D(
  srt,
  group.by,
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  palette = "Paired",
  palcolor = NULL,
  bg_color = "grey80",
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  lineages = NULL,
  lineage_palette = "Dark2",
  span = 0.75,
  arrow_reverse = FALSE,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{group.by}{Name of one or more metadata columns to group (color) cells by (for example, orig.ident).}

\item{reduction}{Which dimensionality reduction to use.}

\item{palette}{Name of palette to use. Default is "Paired".}

\item{bg_color}{Color value for NA points.}

\item{pt.size}{Point size for plotting.}

\item{cells.highlight}{A vector of names of cells to highlight.}

\item{cols.highlight}{A vector of colors to highlight the cells.}

\item{shape.highlight}{Shape of the cell to highlight. See \href{https://plotly.com/r/reference/scattergl/#scattergl-marker-symbol}{scattergl-marker-symbol}}

\item{sizes.highlight}{Size of highlighted cells.}

\item{width}{Width in pixels, defaults to automatic sizing.}

\item{height}{Height in pixels, defaults to automatic sizing.}
}
\description{
Plotting cell points on a reduced 3D space and coloring according to the groups of the cells.
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
CellDimPlot3D(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")

pancreas_sub <- Standard_SCP(pancreas_sub)
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")
CellDimPlot3D(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D", lineages = paste0("Lineage", 1:2))
}
