% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{check_DataType}
\alias{check_DataType}
\title{Check and report the type of data}
\usage{
check_DataType(srt, data = NULL, slot = "data", assay = NULL)
}
\arguments{
\item{srt}{A Seurat object}

\item{data}{Use a data instead of getting it from a Seurat object.}

\item{slot}{Specific slot to get data from.}

\item{assay}{Specific assay to get data from.}
}
\description{
Check and report the type of data
}
