% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_R}
\alias{check_R}
\title{Check and install R packages}
\usage{
check_R(
  packages,
  package_names = NULL,
  install_methods = c("BiocManager::install", "install.packages",
    "devtools::install_github"),
  lib = .libPaths()[1],
  force = FALSE
)
}
\arguments{
\item{packages}{Package to be installed. Package source can be CRAN, Bioconductor or Github, e.g. scmap, davidsjoberg/ggsankey.}

\item{package_names}{The name of the package that corresponds to the \code{packages} parameter, used to check if the package is already installed.
By default, the package name is extracted according to the \code{packages} parameter.}

\item{install_methods}{Functions used to install R packages.}

\item{lib}{The location of the library directories where to install the packages.}

\item{force}{Whether to force the installation of packages. Default is \code{FALSE}.}
}
\description{
Check and install R packages
}
