% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{check_srtMerge}
\alias{check_srtMerge}
\title{Check and preprocess a seurat object}
\usage{
check_srtMerge(
  srtMerge,
  batch = "orig.ident",
  assay = "RNA",
  do_normalization = NULL,
  normalization_method = "LogNormalize",
  do_HVF_finding = TRUE,
  HVF_source = "separate",
  HVF_method = "vst",
  nHVF = 2000,
  HVF_intersect = FALSE,
  HVF_min_intersection = 1,
  HVF = NULL,
  vars_to_regress = NULL,
  seed = 11,
  ...
)
}
\arguments{
\item{srtMerge}{}

\item{HVF_min_intersection}{}
}
\description{
Check and preprocess a seurat object
}
