% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{choose.abs.min}
\alias{choose.abs.min}
\title{Choose the minimum absolute value from each row}
\usage{
choose.abs.min(mat = matrix(stats::rnorm(21), ncol = 3))
}
\description{
Choose the minimum absolute value from each row
}
