% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{circular_concentration}
\alias{circular_concentration}
\title{Circular concentration (1 - circular variance)}
\usage{
circular_concentration(angles, period = 24)
}
\arguments{
\item{angles}{Angles (in hours)}

\item{period}{Period (default 24)}
}
\value{
Concentration (0 = uniform, 1 = all same)
}
\description{
Circular concentration (1 - circular variance)
}
