% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{circular_mean}
\alias{circular_mean}
\title{Circular mean calculation}
\usage{
circular_mean(angles, period = 24)
}
\arguments{
\item{angles}{Angles (in hours)}

\item{period}{Period (default 24)}
}
\value{
Mean angle (in hours)
}
\description{
Circular mean calculation
}
