% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{ClassDimPlot3D}
\alias{ClassDimPlot3D}
\title{3D-Dimensional reduction plot for cell classification visualization.}
\usage{
ClassDimPlot3D(
  srt,
  group.by = "orig.ident",
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  palette = "Paired",
  pal_color = NULL,
  bg_color = "grey80",
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  lineages = NULL,
  lineage_palette = "Dark2",
  span = 0.75,
  arrow_reverse = FALSE,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{group.by}{Name of one or more metadata columns to group (color) cells by (for example, orig.ident).}

\item{reduction}{Which dimensionality reduction to use.}

\item{palette}{Name of palette to use. Default is "Paired".}

\item{bg_color}{Color value for NA points.}

\item{pt.size}{Point size for plotting.}

\item{cells.highlight}{A vector of names of cells to highlight.}

\item{cols.highlight}{A vector of colors to highlight the cells.}

\item{shape.highlight}{Shape of the cell to highlight. See \href{https://plotly.com/r/reference/scattergl/#scattergl-marker-symbol}{scattergl-marker-symbol}}

\item{sizes.highlight}{Size of highlighted cells.}

\item{width}{Width in pixels, defaults to automatic sizing.}

\item{height}{Height in pixels, defaults to automatic sizing.}

\item{force}{}
}
\description{
Plotting cell points on a reduced 3D space and coloring according to the groups of the cells.
}
\examples{
data(pancreas1k)
pancreas1k <- Standard_SCP(pancreas1k)
ClassDimPlot3D(pancreas1k, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")

pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")
ClassDimPlot3D(pancreas1k, group.by = "SubCellType", reduction = "StandardpcaUMAP3D", lineages = paste0("Lineage", 1:2))
}
