% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_methods}
\alias{compare_methods}
\title{Compare Multiple Methods}
\usage{
compare_methods(power_results_list)
}
\arguments{
\item{power_results_list}{List of results from multiple methods}
}
\value{
Comparison summary table
}
\description{
Compare power, FDR, and FDC across methods
}
