% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_UVW.R
\name{compute_UVW}
\alias{compute_UVW}
\title{Internal function: compute U, V, W in Appendix A}
\usage{
compute_UVW(Q, Y)
}
\arguments{
\item{Q}{inverse of the covariance matrix}

\item{Y}{a vector with \eqn{n} data values.}
}
\value{
a list of U, V, W values
}
\description{
Internal function: compute U, V, W in Appendix A
}
