% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{compute_velocity_on_grid}
\alias{compute_velocity_on_grid}
\title{Compute velocity on grid
The original python code is on https://github.com/theislab/scvelo/blob/master/scvelo/plotting/velocity_embedding_grid.py}
\usage{
compute_velocity_on_grid(
  X_emb,
  V_emb,
  density = NULL,
  smooth = NULL,
  n_neighbors = NULL,
  min_mass = NULL,
  scale = 1,
  adjust_for_stream = FALSE,
  cutoff_perc = NULL
)
}
\description{
Compute velocity on grid
The original python code is on https://github.com/theislab/scvelo/blob/master/scvelo/plotting/velocity_embedding_grid.py
}
