% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{computeMedianCV_SCoPE2}
\alias{computeMedianCV_SCoPE2}
\title{(Deprecated) Compute the median coefficient of variation (CV) per
cell}
\usage{
computeMedianCV_SCoPE2(object, i, peptideCol, proteinCol, batchCol)
}
\arguments{
\item{object}{NULL}

\item{i}{NULL}

\item{peptideCol}{NULL}

\item{proteinCol}{NULL}

\item{batchCol}{NULL}
}
\description{
This function is deprecated and should no longer be used. To
reproduce the SCoPE2 script, you can now use \code{medianCVperCell}
with the following arguments:
}
\details{
\itemize{
\item \code{norm = "SCoPE2"}
\item \code{nobs = 6}
}

Make sure to provide the peptide data from separate assays so that
the normalization factors are computed per batch.
}
