% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conda_python}
\alias{conda_python}
\title{Find the path to Python associated with a conda environment}
\usage{
conda_python(envname = NULL, conda = "auto", all = FALSE)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{all}{Boolean; report all instances of Python found?}
}
\description{
Find the path to Python associated with a conda environment
}
