% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_pred.R
\name{conformal_pred}
\alias{conformal_pred}
\title{Spatial Conformal Prediction Intervals}
\usage{
conformal_pred(s0, s, Y, theta, eta = Inf, m = NULL, alpha = 0.05)
}
\arguments{
\item{s0}{prediction location}

\item{s}{an \eqn{n \times d}{n x d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{theta}{spatial covariance parameters as in \code{\link{mat_cov}}}

\item{eta}{numerical value of the kernel bandwidth for the weight schema in conformal prediction. Defauls to \eqn{Inf} meaning equal weight on surrounding \eqn{m} points.}

\item{m}{an postive integer representing the number of nearest locations used for prediction. Depends on eta.}

\item{alpha}{significance level. Defaults to 0.05.}
}
\value{
A vector of lower and upper bounds of the conformal prediction interval.
}
\description{
This function provides the conformal prediction interval for spatial location \code{s0} given obserations \code{s, Y}.
}
\examples{
N = 40; n = N^2
S = seq(0,1,length=N)
s <- expand.grid(S,S)
d <- as.matrix(dist(s))

theta        = c(0,3,0.1,0.7)
names(theta) = c("Nugget","PartialSill","Range","Smoothness")
C <- mat_cov(d,theta)
X <- t(chol(C))\%*\%rnorm(n)
Y <- X^3 + rnorm(n)

# Estimate spatial covariance parameters
bins     <- seq(0.01,0.2,0.01)
thetaHat <- get_theta(s,Y,dists=bins,plot_fitted=FALSE)
Q        <- solve(mat_cov(d,thetaHat))

s0       <- c(0.5, 0.5)
conformal_pred(s0,s,Y,thetaHat,m=100,eta=0.1)
}
