% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copyAssay}
\alias{copyAssay}
\title{Duplicate an assay in a `Features` object}
\usage{
copyAssay(obj, i, name = "newAssay", link = FALSE)
}
\arguments{
\item{obj}{A `Features` object}

\item{i}{A `numeric(1)` or `character(1)` indicating which assay to transfer 
the `colData` to.}

\item{name}{A `character(1)` naming the new assay. Note that the function 
will fail if of the names in `name` is already present.}

\item{link}{A `logical(1)` indicating whether to add a one-to-one assayLink
from assay `i` and to the copied assay}
}
\value{
A `Features` object
}
\description{
The function copies an assay within the supplied `Features` object, 
eventually creating a one-to-one link between the input assay and the copied 
assay.
}
\examples{
TODO
}
