% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{CP_OneGroup}
\alias{CP_OneGroup}
\title{Single Group Rhythmicity Analysis (Internal)}
\usage{
CP_OneGroup(x1, period = 24, alpha = 0.05, CI = FALSE, p.adjust.method = "BH")
}
\arguments{
\item{x1}{Group 1 data (list with data, time, gname components)}

\item{period}{Period (default 24)}

\item{alpha}{Significance threshold}

\item{CI}{Calculate confidence intervals?}

\item{p.adjust.method}{P-value adjustment method}
}
\value{
List with rhythmicity results
}
\description{
Single Group Rhythmicity Analysis (Internal)
}
