% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-app.R
\name{CreateDataFile}
\alias{CreateDataFile}
\title{CreateDataFile}
\usage{
CreateDataFile(
  srt,
  DataFile,
  name = NULL,
  assays = "RNA",
  slots = "data",
  compression_level = 6,
  overwrite = FALSE
)
}
\arguments{
\item{srt}{}

\item{overwrite}{}
}
\description{
CreateDataFile
}
