% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{CSS_integrate}
\alias{CSS_integrate}
\title{CSS_integrate}
\usage{
CSS_integrate(
  srtMerge = NULL,
  batch = "orig.ident",
  append = TRUE,
  srtList = NULL,
  do_normalization = NULL,
  normalization_method = "logCPM",
  do_HVF_finding = TRUE,
  HVF_source = "separate",
  HVF_method = "vst",
  nHVF = 2000,
  HVF = NULL,
  do_scaling = TRUE,
  vars_to_regress = NULL,
  regression_model = "linear",
  liner_reduction = "pca",
  liner_reduction_dims = 100,
  liner_reduction_dims_use = NULL,
  liner_reduction_distance = "euclidean",
  force_liner_reduction = FALSE,
  nonliner_reduction = "umap",
  nonliner_reduction_dims = c(2, 3),
  nonliner_reduction_distance = "cosine",
  force_nonliner_reduction = TRUE,
  do_cluster_finding = TRUE,
  cluster_algorithm = "louvain",
  cluster_resolution = 0.6,
  cluster_reorder = TRUE,
  CSS_params = list(),
  seed = 11
)
}
\arguments{
\item{srtMerge}{}

\item{CSS_param}{}
}
\description{
CSS_integrate
}
