% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_DoubletDetection}
\alias{db_DoubletDetection}
\title{Run doublet-calling with DoubletDetection}
\usage{
db_DoubletDetection(srt, assay = "RNA", db_rate = ncol(srt)/1000 * 0.01, ...)
}
\description{
Run doublet-calling with DoubletDetection
}
\examples{
data("pancreas1k")
pancreas1k <- db_DoubletDetection(pancreas1k)
ClassDimPlot(pancreas1k, reduction = "umap", group.by = "db.DoubletDetection_class")
ExpDimPlot(pancreas1k, reduction = "umap", features = "db.DoubletDetection_score")
}
