% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_scDblFinder}
\alias{db_scDblFinder}
\title{Run doublet-calling with scDblFinder}
\usage{
db_scDblFinder(srt, assay = "RNA", db_rate = ncol(srt)/1000 * 0.01, ...)
}
\description{
Run doublet-calling with scDblFinder
}
\examples{
if (interactive()) {
  data("pancreas1k")
  pancreas1k <- db_scDblFinder(pancreas1k)
  ClassDimPlot(pancreas1k, reduction = "umap", group.by = "db.scDblFinder_class")
  ExpDimPlot(pancreas1k, reduction = "umap", features = "db.scDblFinder_score")
}
}
