% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_scds}
\alias{db_scds}
\title{Run doublet-calling with scds}
\usage{
db_scds(
  srt,
  assay = "RNA",
  db_rate = ncol(srt)/1000 * 0.01,
  method = "hybrid",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Run doublet-calling with scds
}
\examples{
data("pancreas1k")
pancreas1k <- db_scds(pancreas1k, method = "hybrid")
ClassDimPlot(pancreas1k, reduction = "umap", group.by = "db.scds_hybrid_class")
ExpDimPlot(pancreas1k, reduction = "umap", features = "db.scds_hybrid_score")
}
