% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_Scrublet}
\alias{db_Scrublet}
\title{Run doublet-calling with Scrublet}
\usage{
db_Scrublet(srt, assay = "RNA", db_rate = ncol(srt)/1000 * 0.01, ...)
}
\description{
Run doublet-calling with Scrublet
}
\examples{
if (interactive()) {
  data("pancreas1k")
  pancreas1k <- db_Scrublet(pancreas1k)
  ClassDimPlot(pancreas1k, reduction = "umap", group.by = "db.Scrublet_class")
  ExpDimPlot(pancreas1k, reduction = "umap", features = "db.Scrublet_score")
}
}
