% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{DCP_sim_data}
\alias{DCP_sim_data}
\title{Simulate Data for Rhythmicity Comparison}
\usage{
DCP_sim_data(
  ngene = 1000,
  nsample = 30,
  A1 = c(1, 3),
  A2 = c(1, 3),
  phase1 = c(0, 2 * pi),
  phase2 = c(0, 2 * pi),
  M1 = c(4, 6),
  M2 = c(4, 6),
  sigma1 = 1,
  sigma2 = 1
)
}
\arguments{
\item{ngene}{Number of genes}

\item{nsample}{Number of samples per group}

\item{A1}{Amplitude range for group 1 (c(min, max))}

\item{A2}{Amplitude range for group 2}

\item{phase1}{Phase range for group 1}

\item{phase2}{Phase range for group 2}

\item{M1}{MESOR range for group 1}

\item{M2}{MESOR range for group 2}

\item{sigma1}{Noise level for group 1}

\item{sigma2}{Noise level for group 2}
}
\value{
List of two elements, each is a list with data, time, gname components
}
\description{
Simple two-group study simulation for illustration
}
\examples{
x = DCP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)

}
