% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DefaultReduction}
\alias{DefaultReduction}
\title{Find the default reduction name in the srt object.}
\usage{
DefaultReduction(srt, pattern = NULL, min_dim = 2, max_distance = 0.1)
}
\arguments{
\item{srt}{A Seurat object.}

\item{pattern}{Name of the pattern to search for.}

\item{min_dim}{Minimum dimension threshold.}
}
\value{
Default reduction name.
}
\description{
Find the default reduction name in the srt object.
}
