% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect_DCP}
\alias{detect_DCP}
\title{DCP single-cohort rhythmicity detection}
\usage{
detect_DCP(expr, times, period = 24)
}
\arguments{
\item{expr}{Gene x sample expression matrix}

\item{times}{Numeric time vector (length = ncol(expr))}

\item{period}{Period (default 24)}
}
\value{
Numeric vector of p-values, length nrow(expr)
}
\description{
DCP single-cohort rhythmicity detection
}
