% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect_DODR}
\alias{detect_DODR}
\title{DODR two-group differential oscillation detection}
\usage{
detect_DODR(expr1, times1, expr2, times2, period = 24)
}
\arguments{
\item{expr1, expr2}{Gene x sample matrices}

\item{times1, times2}{Numeric time vectors}

\item{period}{Period (default 24)}
}
\value{
list(pval_DR, pval_DP=NA, pval_DM=NA)
}
\description{
Wraps DODR::dodr(). DR only.
}
