% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divideByReference}
\alias{divideByReference}
\title{Divide assay columns by a reference column}
\usage{
divideByReference(obj, i, colDataCol, samplePattern, refPattern)
}
\arguments{
\item{obj}{A `Features` object}

\item{i}{A `numeric()` or `character()` vector indicating from which assays 
the `rowData` should be taken.}

\item{colDataCol}{A `character(1)` indicating the variable to take from 
`colData(obj)` that gives the sample annotation.}

\item{samplePattern}{A `character(1)` pattern that matches the sample 
encoding in `colDataCol`}

\item{refPattern}{A `character(1)` pattern that matches the carrier 
encoding in `colDataCol`. Only one match per assay is allowed, otherwise
only the first match is taken}
}
\value{
A `Features` object
}
\description{
The function divides the sample columns by a reference column. The sample 
and reference columns are defined based on the provided `colDataCol` 
variable and on regular expression matching.
}
\details{
The supplied assay(s) are replaced with the values computed after reference
division.
}
\examples{
TODO
}
