% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_UVW.R
\name{.compute_UVW2}
\alias{.compute_UVW2}
\title{compute U, V, W}
\usage{
.compute_UVW2(Q, Y)
}
\arguments{
\item{Q}{inverse of the covariance matrix}

\item{Y}{a vector with \eqn{n} data values.}
}
\value{
a list of U, V, W values
}
\description{
compute U, V, W in Mao et al. (2020) Appendix A when delta_i = (Y_i - Yhat_i)^2
}
\keyword{internal}
