% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR.R
\name{.define.bins}
\alias{.define.bins}
\title{Define Bins}
\usage{
.define.bins(max.dist, uvec = "default", breaks = "default", nugget.tolerance)
}
\arguments{
\item{nugget.tolerance}{}
}
\value{

}
\description{
Define Bins
}
\keyword{internal}
