% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_contour.R
\name{.fast_plausibility_contour}
\alias{.fast_plausibility_contour}
\title{Fast plausibility_contour calculation when pred_fun = Krige_fun and (dfun = "residual2" or "std_residual2")}
\usage{
.fast_plausibility_contour(Y_aug, w_aug, Q, std = TRUE)
}
\arguments{
\item{Q}{an \eqn{n+1 \times n+1} inverse of estimated covariance matrix}

\item{s0}{prediction location}

\item{s}{an \eqn{n \times d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{w}{weights for the non-conformity measures.}

\item{alpha}{significance level. Defaults to 0.05.}
}
\value{
A vector of lower and upper bounds of the conformal prediction interval.
}
\description{
Internal function provides the conformal prediction interval
for spatial location \code{s0} given partial obserations \code{s1,...,sm, Y1,.., Ym}
when the square of the standard kriging residuals are used as the non-conformality measures,
and weights are provided.
}
\keyword{internal}
