% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{.generate_Y_cand}
\alias{.generate_Y_cand}
\title{Generate candidate Y values}
\usage{
.generate_Y_cand(pred_fun, dfun, precision)
}
\arguments{
\item{pred_fun}{spatial prediction function with inputs being \eqn{s0, s, Y} and ouputs being predicted \code{Y(s0)}
(and its standard error). Defaults to \code{\link{krige_pred}} representing Kriging prediction.}

\item{dfun}{non-conformity measure with four options.
In which, \code{"residual2"} (default) represents squared residual,
\code{"std_residual2"} represents standardized squared residual,
\code{"abs_residual"} represents absolute residual,
and \code{"std_abs_residual"} represents standardized absolute residual.}

\item{precision}{a positive scalar represents how dense the candidates for \code{Y(s)} are. Defaults to \code{NULL}.}
}
\value{
a vector of candidate Y values
}
\description{
Generate candidate Y values
}
\keyword{internal}
