% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{.prime}
\alias{.prime}
\title{Transform the input parameters}
\usage{
.prime(s0, s, Y, global, eta, m, dfun)
}
\arguments{
\item{s0}{prediction location, a numeric vector with \code{length = 2}.}

\item{s}{an \eqn{n \times 2}{n x 2} \code{matrix} or \code{data.frame} with two coordinates of \eqn{n} locations.}

\item{Y}{a vector with \eqn{n} values corresponding to \code{Y(s)}.}

\item{global}{logical; if \code{TRUE} , \code{scp} function returns the result of global spatial conformal prediction (GSCP);
if \code{FALSE}, \code{scp} function returns the result of local spatial conformal prediction (LSCP)
and users need to specify \code{eta}. Defaults to \code{TRUE}.}

\item{eta}{kernel bandwidth for weight schema, a positve scalar with smaller value meaning more localized procedure.
Defauls to \code{Inf}, which puts equal weight on surrounding \eqn{m} points.}

\item{m}{an postive integer representing the number of nearest locations to use for prediction.
Default depands on \code{eta}.}

\item{dfun}{non-conformity measure with four options.
In which, \code{"residual2"} (default) represents squared residual,
\code{"std_residual2"} represents standardized squared residual,
\code{"abs_residual"} represents absolute residual,
and \code{"std_abs_residual"} represents standardized absolute residual.}
}
\description{
An internal function to transform the input parameters
}
\keyword{internal}
