% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_contour.R
\name{.Q_plausibility_contour}
\alias{.Q_plausibility_contour}
\title{Need inverse covariance matrix to calculate plausibility_contour}
\usage{
.Q_plausibility_contour(Q, Y_cand, Y_aug, w_aug, pred_fun, T_dfun)
}
\arguments{
\item{std}{}
}
\value{

}
\description{
Need inverse covariance matrix to calculate plausibility_contour
}
\keyword{internal}
