% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download}
\alias{download}
\title{Download File from the Internet}
\usage{
download(
  url,
  destfile,
  methods = c("auto", "wget", "libcurl", "curl", "wininet", "internal"),
  quiet = FALSE,
  ...,
  max_tries = 2
)
}
\arguments{
\item{url}{a \code{\link{character}} string (or longer vector e.g.,
    for the \code{"libcurl"} method) naming the URL of a resource to be
    downloaded.}

\item{destfile}{a character string (or vector, see the \code{url}
    argument) with the file path where the downloaded file is to be
    saved.  Tilde-expansion is performed.}

\item{methods}{Methods to be used for downloading files. The default is to try different download methods in turn until the download is successfully completed.}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and
    the progress bar.}

\item{...}{Other arguments passed to \code{\link[utils]{download.file}}}

\item{max_tries}{Number of tries for each download method.}
}
\description{
Download File from the Internet
}
