% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{drop_data}
\alias{drop_data}
\title{Drop all data in the plot (only one observation is kept)}
\usage{
drop_data(p)
}
\description{
Drop all data in the plot (only one observation is kept)
}
\examples{
library(ggplot2)
p <- ggplot(data = mtcars, aes(x = mpg, y = wt, colour = cyl)) +
  geom_point() +
  scale_x_continuous(limits = c(10, 30)) +
  scale_y_continuous(limits = c(1, 6))
object.size(p)

p_drop <- drop_data(p)
object.size(p_drop)

p / p_drop

}
