% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DynamicHeatmap}
\alias{DynamicHeatmap}
\title{Heatmap plot for dynamic features along lineages}
\usage{
DynamicHeatmap(
  srt,
  lineages,
  features = NULL,
  feature_from = lineages,
  use_fitted = FALSE,
  border = TRUE,
  min_expcells = 20,
  r.sq = 0.2,
  dev.expl = 0.2,
  padjust = 0.05,
  cell_density = 1,
  order_by = c("peaktime", "valleytime"),
  slot = "counts",
  assay = "RNA",
  exp_method = c("zscore", "raw", "fc", "log2fc", "log1p"),
  lib_normalize = TRUE,
  libsize = NULL,
  family = NULL,
  cluster_rows = FALSE,
  cluster_row_slices = FALSE,
  show_row_names = FALSE,
  show_column_names = FALSE,
  row_title_rot = 0,
  column_title_rot = 0,
  feature_split = NULL,
  feature_split_by = NULL,
  n_split = NULL,
  split_method = c("mfuzz", "kmeans", "kmeans-peaktime", "hclust", "hclust-peaktime"),
  decreasing = FALSE,
  fuzzification = NULL,
  show_fuzzification = FALSE,
  anno_terms = FALSE,
  anno_keys = FALSE,
  anno_features = FALSE,
  terms_width = unit(4, "in"),
  terms_fontsize = 8,
  keys_width = unit(2, "in"),
  keys_fontsize = c(6, 10),
  features_width = unit(2, "in"),
  features_fontsize = c(6, 10),
  IDtype = "symbol",
  species = "Homo_sapiens",
  db_update = FALSE,
  db_version = "latest",
  convert_species = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  db = "GO_BP",
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  universe = NULL,
  GO_simplify = FALSE,
  GO_simplify_padjustCutoff = 0.2,
  simplify_method = "Rel",
  simplify_similarityCutoff = 0.7,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  topTerm = 5,
  show_termid = FALSE,
  topWord = 20,
  min_word_length = 3,
  exclude_words = c("cell", "cellular", "dna", "rna", "protein", "development",
    "organization", "system", "regulation", "positive", "negative", "response",
    "process"),
  nlabel = 20,
  features_label = NULL,
  label_size = 10,
  label_color = "black",
  pseudotime_label = NULL,
  pseudotime_label_color = "black",
  pseudotime_label_linetype = 2,
  pseudotime_label_linewidth = 3,
  heatmap_palette = "RdBu",
  heatmap_palcolor = NULL,
  pseudotime_palette = "cividis",
  pseudotime_palcolor = NULL,
  feature_split_palette = "jama",
  feature_split_palcolor = NULL,
  cell_annotation = NULL,
  cell_palette = "Paired",
  cell_palcolor = NULL,
  cell_annotation_params = list(),
  feature_annotation = NULL,
  feature_palette = "Dark2",
  feature_palcolor = NULL,
  feature_annotation_params = list(),
  reverse_ht = NULL,
  use_raster = NULL,
  raster_device = "png",
  height = NULL,
  width = NULL,
  units = "inch",
  seed = 11,
  ht_params = list()
)
}
\arguments{
\item{seed}{}
}
\description{
Heatmap plot for dynamic features along lineages
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
pancreas_sub <- RunDynamicFeatures(pancreas_sub, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
ht_result1 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = "Lineage1",
  cell_annotation = "SubCellType",
  n_split = 5, split_method = "kmeans-peaktime"
)
ht_result1$plot

ht_result2 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = "Lineage1",
  features = c("Sox9", "Neurod2", "Isl1", "Rbp4", "S_score", "G2M_score"),
  cell_annotation = "SubCellType"
)
ht_result2$plot

ht_result3 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_split = 5, split_method = "kmeans",
  cluster_rows = TRUE
)
ht_result3$plot

pancreas_sub <- AnnotateFeatures(pancreas_sub, species = "Mus_musculus", db = c("TF", "SP"))
ht_result3 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  use_fitted = TRUE, heatmap_palette = "viridis",
  cell_annotation = "SubCellType",
  n_split = 5, split_method = "mfuzz",
  reverse_ht = "Lineage1",
  feature_annotation = c("TF", "SP"),
  feature_palcolor = list(c("gold", "steelblue"), c("forestgreen"))
)
ht_result3$plot

ht_result5 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_split = 5, reverse_ht = "Lineage1",
  species = "Mus_musculus", db = "GO_BP",
  anno_terms = TRUE, anno_keys = TRUE, anno_features = TRUE,
  pseudotime_label = max(pancreas_sub$Lineage1[pancreas_sub$SubCellType == "Pre-endocrine"], na.rm = TRUE)
)
ht_result5$plot

}
