% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DynamicPlot}
\alias{DynamicPlot}
\title{Plot for dynamic features along lineages.}
\usage{
DynamicPlot(
  srt,
  features,
  lineages,
  slot = "counts",
  assay = "RNA",
  family = NULL,
  exp_method = c("log1p", "zscore", "log2fc", "raw"),
  lib_normalize = TRUE,
  libsize = NULL,
  order.by = "pseudotime",
  group.by = NULL,
  compare_lineages = TRUE,
  compare_features = FALSE,
  add_line = TRUE,
  add_interval = TRUE,
  line.size = 1,
  line_palette = "Dark2",
  line_palcolor = NULL,
  add_point = TRUE,
  pt.size = 1,
  point_palette = "Paired",
  point_palcolor = NULL,
  add_rug = TRUE,
  add_density = TRUE,
  aspect.ratio = 0.5,
  legend.position = "right",
  legend.direction = "vertical",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  align = "hv",
  axis = "lr"
)
}
\arguments{
\item{axis}{}
}
\description{
Plot for dynamic features along lineages.
}
\examples{
data("pancreas1k")
pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "UMAP")
DynamicPlot(
  srt = pancreas1k,
  lineages = c("Lineage1", "Lineage2"),
  features = c("Nnat", "Irx1"),
  group.by = "SubCellType",
  compare_lineages = TRUE,
  compare_features = FALSE
)
}
