% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{EnrichmentPlot}
\alias{EnrichmentPlot}
\title{EnrichmentPlot}
\usage{
EnrichmentPlot(
  x,
  enrichment = "GO_BP",
  plot_type = "bar",
  topTerm = 6,
  topGene = 100,
  text_size_scale = 1,
  character_width = 50,
  line_height = 0.7,
  panel_fix = FALSE,
  panel_height_scale = 0.25 * text_size_scale,
  panel_width = 2.5,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  palette = "Spectral",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  align = "hv",
  axis = "lr"
)
}
\arguments{
\item{x}{A enrichemnt result generated by the \link{RunEnrichment} function.}

\item{plot_type}{Type of plot to be visualized.}

\item{topTerm}{Number of terms to plot if \code{plot_type="bar"} or \code{plot_type="lollipop"}.}

\item{topGene}{Number of genes to plot if \code{plot_type="wordcloud"}.}

\item{pvalueCutoff}{pvalueCutoff}

\item{padjustCutoff}{padjustCutoff}

\item{palette}{palette}

\item{combine}{combine}

\item{nrow}{nrow}

\item{ncol}{ncol}

\item{byrow}{byrow}

\item{Enrichment}{Enrichment column name in \code{x}.}

\item{Groups}{Groups column name in \code{x}.}
}
\description{
EnrichmentPlot
}
\examples{
if (interactive()) {
  data("pancreas1k")
  library(dplyr)
  pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType")
  de_filter <- filter(pancreas1k@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05)
  res <- RunEnrichment(geneID = de_filter$gene, geneID_groups = de_filter$group1, enrichment = "GO_BP", species = "Mus_musculus")
  EnrichmentPlot(res$enrichment, plot_type = "bar")
  EnrichmentPlot(res$enrichment, plot_type = "bar", character_width = 30, text_size_scale = 0.8)

  EnrichmentPlot(filter(res$enrichment, Groups \%in\% c("Ductal", "Endocrine")), plot_type = "lollipop", ncol = 1)
  EnrichmentPlot(filter(res$enrichment, Groups \%in\% c("Ductal", "Endocrine")), plot_type = "wordcloud")
}
}
