% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env_exist}
\alias{env_exist}
\title{Check if a conda environment exists}
\usage{
env_exist(conda = "auto", envname = NULL, envs_dir = NULL)
}
\arguments{
\item{conda}{The path to a conda executable. Use \code{"auto"} to allow SCP to automatically find an appropriate conda binary.}

\item{envname}{The name of a conda environment.}

\item{envs_dir}{Directories in which conda environments are located.}
}
\description{
Check if a conda environment exists
}
