% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Env_requirements}
\alias{Env_requirements}
\title{Env_requirements function}
\usage{
Env_requirements(version = "3.8-1")
}
\arguments{
\item{version}{A character vector specifying the version of the environment (default is "3.8-1").}
}
\value{
A list of requirements for the specified version.
}
\description{
This function provides the SCP python environment requirements for a specific version.
}
\details{
The function returns a list of requirements including the required Python version
         and a list of packages with their corresponding versions.
}
\examples{
# Get requirements for version "3.8-1"
Env_requirements("3.8-1")

}
