% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exist_Python_pkg}
\alias{exist_Python_pkg}
\title{Check if the python package exists in the environment}
\usage{
exist_Python_pkg(pkg, envname = "SCP", show_pkg_info = TRUE)
}
\arguments{
\item{pkg}{Python package name.}

\item{envname}{The name of, or path to, a Python virtual environment.}

\item{show_pkg_info}{Whether to display information about installed packages.}
}
\description{
Check if the python package exists in the environment
}
