% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exist_Python_pkgs}
\alias{exist_Python_pkgs}
\title{Check if the python package exists in the environment}
\usage{
exist_Python_pkgs(packages, envname = NULL, conda = "auto")
}
\arguments{
\item{packages}{A character vector, indicating package names which should be installed or removed. Use \code{⁠<package>==<version>}⁠ to request the installation of a specific version of a package.}

\item{envname}{The name of a conda environment.}

\item{conda}{The path to a conda executable. Use \code{"auto"} to allow SCP to automatically find an appropriate conda binary.}
}
\description{
Check if the python package exists in the environment
}
