% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{ExpDotPlot}
\alias{ExpDotPlot}
\title{ExpDotPlot}
\usage{
ExpDotPlot(
  srt,
  genes = NULL,
  gene_groups = NULL,
  cell_group_by = NULL,
  aggregate_fun = mean,
  slot = "counts",
  assay = "RNA",
  exp_method = c("zscore", "raw", "log2fc", "log1p"),
  lib_normalize = FALSE,
  lib_size = NULL,
  heatmap_palette = "YlOrRd",
  gene_palette = "Paired",
  cell_palette = "Paired",
  add_reticle = FALSE,
  cluster_rows = TRUE,
  cluster_columns = FALSE,
  grid_size = 0.4
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{genes}{A vector of gene names to plot.}

\item{gene_groups}{A vector of group names for genes.}

\item{cell_group_by}{Columns used to calculate cell expression. One heatmap per column name.}

\item{assay}{Assay used to calculate the expression.}

\item{exp_method}{Method used to calculate cell expression.}

\item{heatmap_palette}{Heatmap expression palette.}

\item{gene_palette}{Gene group palette.}

\item{cell_palette}{Column palette.}

\item{cluster_rows}{Whether to cluster rows.}

\item{cluster_columns}{Whether to cluster columns.}

\item{grid_size}{size for each dot.}
}
\description{
ExpDotPlot
}
\examples{
library(dplyr)
data(pancreas1k)
pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType")
de_filter <- filter(pancreas1k@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
de_top <- de_filter \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)
ExpDotPlot(pancreas1k, genes = de_top$gene, gene_groups = de_top$group1, cell_group_by = "CellType")
}
