% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{ExpHeatmap}
\alias{ExpHeatmap}
\title{ExpHeatmap}
\usage{
ExpHeatmap(
  srt,
  features = NULL,
  feature_split = NULL,
  cluster_features = FALSE,
  cell_split_by = "orig.ident",
  cluster_cells = FALSE,
  max_cells = 100,
  slot = "counts",
  assay = "RNA",
  exp_method = c("zscore", "raw", "log2fc", "log1p"),
  n_split = NULL,
  heatmap_split_by = cell_split_by[1],
  split_method = c("kmeans", "hclust", "mfuzz"),
  row_title_size = 12,
  row_title_rot = 90,
  decreasing = TRUE,
  lib_normalize = TRUE,
  libsize = NULL,
  anno_terms = FALSE,
  anno_keys = FALSE,
  anno_features = FALSE,
  IDtype = "symbol",
  species = "Homo_sapiens",
  db_update = FALSE,
  db_version = "latest",
  convert_species = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  db = "GO_BP",
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  universe = NULL,
  GO_simplify = FALSE,
  GO_simplify_padjustCutoff = 0.2,
  simplify_method = "Rel",
  simplify_similarityCutoff = 0.7,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  topTerm = 5,
  show_id = TRUE,
  topWord = 20,
  min_word_length = 3,
  exclude_words = c("cell", "cellular", "dna", "rna", "protein", "development",
    "organization", "system", "regulation", "positive", "negative", "response",
    "process"),
  anno_width = unit(c(4, 2, 2), "in"),
  anno_size = c(6, 10),
  nlabel = 20,
  features_label = NULL,
  label_size = 10,
  label_color = "black",
  heatmap_palette = "RdBu",
  cell_split_palette = "Paired",
  feature_split_palette = "jama",
  cell_annotation = NULL,
  cell_palette = "Paired",
  cell_palcolor = NULL,
  feature_annotation = NULL,
  feature_palette = "Dark2",
  feature_palcolor = NULL,
  use_raster = NULL,
  height = NULL,
  width = NULL,
  units = "inch",
  seed = 11
)
}
\arguments{
\item{seed}{}
}
\description{
ExpHeatmap
}
\examples{
library(dplyr)
data(pancreas_sub)
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType")
de_filter <- filter(pancreas_sub@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
ht1 <- ExpHeatmap(
  srt = pancreas_sub, features = de_filter$gene, cell_split_by = "CellType", cluster_features = TRUE
)
ht1$plot

pancreas_sub <- AnnotateFeatures(pancreas_sub, species = "Mus_musculus", anno_TF = TRUE)
ht2 <- ExpHeatmap(
  srt = pancreas_sub, features = de_filter$gene, n_split = 4, cell_split_by = "CellType",
  feature_annotation = c("TF", "TF_cofactors"), feature_palcolor = list(c("black", "transparent"))
)
ht2$plot

ht3 <- ExpHeatmap(
  srt = pancreas_sub, features = de_filter$gene, feature_split = de_filter$group1, cell_split_by = "CellType",
  species = "Mus_musculus", anno_terms = TRUE, anno_keys = TRUE, anno_features = TRUE
)
ht3$plot

}
