% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{ExpHeatmapPlot}
\alias{ExpHeatmapPlot}
\title{ExpHeatmapPlot}
\usage{
ExpHeatmapPlot(
  srt,
  genes = NULL,
  gene_groups = NULL,
  cluster_genes = TRUE,
  cell_group_by = "orig.ident",
  cluster_cells = TRUE,
  max_cells = 100,
  slot = "counts",
  assay = "RNA",
  exp_method = c("zscore", "raw", "log2fc", "log1p"),
  lib_normalize = FALSE,
  lib_size = NULL,
  use_raster = NULL,
  heatmap_palette = "RdBu",
  gene_palette = "Paired",
  cell_palette = "Paired"
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{genes}{A vector of gene names to plot.}

\item{gene_groups}{A vector of group names for genes.}

\item{cell_group_by}{Columns used to calculate cell expression. One heatmap per column name.}

\item{assay}{Assay used to calculate the expression.}

\item{exp_method}{Method used to calculate cell expression.}

\item{heatmap_palette}{Heatmap expression palette.}

\item{gene_palette}{Gene group palette.}

\item{cell_palette}{Column palette.}

\item{ncell_sample}{Number of cells used to sample representative cell populations.}
}
\description{
ExpHeatmapPlot
}
\examples{
library(dplyr)
data(pancreas1k)
pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType")
de_filter <- filter(pancreas1k@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
ExpHeatmapPlot(srt = pancreas1k, genes = de_filter$gene, gene_groups = de_filter$group1, cell_group_by = "CellType")
}
