% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{ExpStatPlot}
\alias{ExpStatPlot}
\title{Violin plot for cell expression.}
\usage{
ExpStatPlot(
  srt,
  features = NULL,
  group.by = NULL,
  split.by = NULL,
  bg.by = NULL,
  fill.by = c("group", "feature", "expression"),
  cells = NULL,
  slot = "data",
  assay = "RNA",
  keep_empty = FALSE,
  stat_single = FALSE,
  plot_type = c("violin", "box", "bar", "histogram", "density", "ridgeline", "horizon"),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  bg_palette = "Paired",
  bg_palcolor = NULL,
  bg_apha = 0.5,
  add_box = TRUE,
  box_fill = "black",
  box_width = 0.1,
  add_point = FALSE,
  pt.color = "black",
  pt.size = NULL,
  pt.alpha = 1,
  jitter.width = 1,
  cells.highlight = NULL,
  cols.highlight = "red",
  sizes.highlight = 1,
  alpha.highlight = 1,
  calculate_coexp = FALSE,
  compare_features = FALSE,
  y.nbreaks = 4,
  y.min = NULL,
  y.max = NULL,
  same.y.lims = FALSE,
  y.trans = "identity",
  sort = FALSE,
  adjust = 1,
  split.plot = FALSE,
  stack = FALSE,
  flip = FALSE,
  comparisons = list(),
  ref_group = NULL,
  pairwise_method = "wilcox.test",
  multiplegroup_comparisons = FALSE,
  multiple_method = "kruskal.test",
  theme_use = "theme_scp",
  aspect.ratio = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = group.by,
  ylab = "Expression level",
  legend.position = "right",
  legend.direction = "vertical",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  align = "hv",
  axis = "lr",
  force = FALSE
)
}
\arguments{
\item{y.min}{}
}
\description{
Violin plot for cell expression.
}
\examples{
library(dplyr)
data("pancreas_sub")
ExpStatPlot(pancreas_sub, features = c("G2M_score", "S_score"), group.by = "SubCellType")
ExpStatPlot(pancreas_sub, features = c("G2M_score", "S_score"), group.by = "SubCellType", split.by = "Phase")
ExpStatPlot(pancreas_sub, features = c("G2M_score", "S_score"), group.by = "SubCellType", fill.by = "expression", palette = "Blues")
ExpStatPlot(pancreas_sub, features = c("Rbp4", "Pyy"), group.by = "SubCellType", comparisons = list(c("Alpha", "Beta"), c("Alpha", "Delta")), multiplegroup_comparisons = TRUE)
ExpStatPlot(pancreas_sub, features = c("Neurog3", "Fev"), group.by = "SubCellType", bg.by = "CellType", stack = TRUE)
ExpStatPlot(pancreas_sub,
  features = c(
    "Sox9", "Anxa2", "Bicc1", # Ductal
    "Neurog3", "Hes6", # EPs
    "Fev", "Neurod1", # Pre-endocrine
    "Rbp4", "Pyy", # Endocrine
    "Ins1", "Gcg", "Sst", "Ghrl" # Beta, Alpha, Delta, Epsilon
  ),
  group.by = "SubCellType", bg.by = "CellType", stack = TRUE, flip = TRUE
)
ExpStatPlot(pancreas_sub,
  features = c(
    "Sox9", "Anxa2", "Bicc1", # Ductal
    "Neurog3", "Hes6", # EPs
    "Fev", "Neurod1", # Pre-endocrine
    "Rbp4", "Pyy", # Endocrine
    "Ins1", "Gcg", "Sst", "Ghrl" # Beta, Alpha, Delta, Epsilon
  ), fill.by = "expression", palette = "Reds",
  group.by = "SubCellType", bg.by = "CellType", stack = TRUE, flip = TRUE
) \%>\% panel_fix_single(width = 10, height = 5)
}
