% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{false_discovery_cost}
\alias{false_discovery_cost}
\title{False Discovery Cost (FDC)}
\usage{
false_discovery_cost(TP, FP)
}
\arguments{
\item{TP}{True positives}

\item{FP}{False positives}
}
\value{
FDC (FP / TP)
}
\description{
Calculate the cost of each true discovery, measured as the
number of false discoveries per true discovery.
}
