% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_scp.R
\name{fast_scp}
\alias{fast_scp}
\title{Fast Spatial Conformal Prediction Intervals}
\usage{
fast_scp(s0, s, Y, thetaHat, alpha = 0.05)
}
\arguments{
\item{s0}{prediction location}

\item{s}{an \eqn{n \times d}{n x d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{thetaHat}{estimated Matern covariance parameters as in \code{\link{mat_cov}}}

\item{alpha}{significance level. Defaults to 0.05.}

\item{w}{weights for the non-conformity measures.}
}
\value{
A vector of lower and upper bounds of the conformal prediction interval.
}
\description{
Internal function provides the conformal prediction interval
for spatial location \code{s0} given partial obserations \code{s1,...,sm, Y1,.., Ym}
when the square of the standard kriging residuals are used as the non-conformality measures,
and weights are provided.
}
