% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{fastMNN_integrate}
\alias{fastMNN_integrate}
\title{fastMNN_integrate}
\usage{
fastMNN_integrate(
  srtMerge = NULL,
  batch = "orig.ident",
  append = TRUE,
  srtList = NULL,
  do_normalization = NULL,
  normalization_method = "logCPM",
  do_HVF_finding = TRUE,
  HVF_source = "separate",
  HVF_method = "vst",
  nHVF = 2000,
  HVF = NULL,
  fastMNN_dims_use = NULL,
  nonliner_reduction = "umap",
  nonliner_reduction_dims = c(2, 3),
  nonliner_reduction_distance = "cosine",
  force_nonliner_reduction = TRUE,
  do_cluster_finding = TRUE,
  cluster_algorithm = "louvain",
  cluster_resolution = 0.6,
  cluster_reorder = TRUE,
  fastMNN_params = list(),
  seed = 11
)
}
\arguments{
\item{srtMerge}{}

\item{...}{}
}
\description{
fastMNN_integrate
}
