% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-app.R
\name{FetchH5}
\alias{FetchH5}
\title{Fetch data from the hdf5 file}
\usage{
FetchH5(
  DataFile,
  MetaFile,
  name = NULL,
  features = NULL,
  slot = NULL,
  assay = NULL,
  metanames = NULL,
  reduction = NULL
)
}
\arguments{
\item{reduction}{}
}
\description{
Fetch data from the hdf5 file
}
\examples{
if (interactive()) {
  data("pancreas_sub")
  pancreas_sub <- Standard_SCP(pancreas_sub)
  PrepareSCExplorer(pancreas_sub, base_dir = "./SCExplorer")
  FetchH5("./SCExplorer/Data.hdf5", "./SCExplorer/Meta.hdf5", features = "Isl1", reduction = "UMAP")
}
}
