% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{format_for_DCP}
\alias{format_for_DCP}
\title{Bridge: Convert simulation output to DCP format}
\usage{
format_for_DCP(expr_matrix, times, gene_names = NULL)
}
\arguments{
\item{expr_matrix}{Gene x sample matrix}

\item{times}{Time vector}

\item{gene_names}{Gene name vector}
}
\value{
DCP-formatted list
}
\description{
Bridge: Convert simulation output to DCP format
}
\keyword{internal}
