% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{forwardStop}
\alias{forwardStop}
\title{Forward Stopping Rule}
\usage{
forwardStop(pv, alpha = 0.1)
}
\arguments{
\item{pv}{P-values}

\item{alpha}{Threshold}
}
\value{
List with stop and p
}
\description{
Forward Stopping Rule
}
