% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{GeneConvert}
\alias{GeneConvert}
\title{Gene ID conversion function using biomart}
\usage{
GeneConvert(
  geneID,
  geneID_from_IDtype = "symbol",
  geneID_to_IDtype = "entrez_id",
  species_from = "Homo_sapiens",
  species_to = NULL,
  Ensembl_version = 103,
  try_times = 5,
  mirror = NULL
)
}
\arguments{
\item{geneID}{A vector of the geneID character.}

\item{geneID_from_IDtype}{ID type of the geneID. e.g. "symbol", "ensembl_id", "entrez_id"}

\item{geneID_to_IDtype}{ID type to convert to. e.g. "symbol", "ensembl_id", "entrez_id"}

\item{species_from}{Latin names for animals of the input geneID.}

\item{species_to}{Latin names for animals of the output geneID. e.g. "Homo_sapiens","Mus_musculus"}

\item{Ensembl_version}{Ensembl database version. If NULL, use the current release version.}

\item{try_times}{Times when trying to connect with the biomart service.}

\item{mirror}{Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'.}
}
\value{
A list.
}
\description{
This function can convert different gene ID types within one species or bewteen two species using biomart service.
}
\examples{
if (interactive()) {
  res <- GeneConvert(
    geneID = c("CDK1", "MKI67", "TOP2A", "AURKA", "CTCF"), geneID_from_IDtype = "symbol", geneID_to_IDtype = "entrez_id",
    species_from = "Homo_sapiens", species_to = "Mus_musculus", Ensembl_version = 103
  )
  str(res)
}
}
