% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsankey.R
\name{geom_alluvial_label}
\alias{geom_alluvial_label}
\alias{geom_alluvial_text}
\title{geom_alluvial_label}
\usage{
geom_alluvial_text(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  space = 0,
  width = 0.1,
  inherit.aes = TRUE,
  ...
)

geom_alluvial_label(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  space = 0,
  width = 0.1,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{provide you own mapping. both x and y need to be numeric.}

\item{data}{provide you own data}

\item{position}{change position}

\item{na.rm}{remove missing values}

\item{show.legend}{show legend in plot}

\item{space}{space between nodes in the y-direction}

\item{width}{width of nodes}

\item{inherit.aes}{should the geom inherits aestethics}

\item{...}{other arguments to be passed to the geo}
}
\value{
ggplot layer
}
\description{
Creates centered labels or text in nodes of your alluvial plot. Needs to have the exact same aestethics as the call to `geom_alluvial` to work.
}
\details{
Other important arguments is; `space` which proves the space between nodes in the y-direction; `shift` which shifts nodes in the y-direction.
}
