% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsankey.R
\name{geom_sankey}
\alias{geom_sankey}
\title{geom_sankey

Creates a sankey plot which visualize flows between nodes. Each observation needs to have a `x` aesthetic as well as a `next_x` column which declares where that observation should flow.
Also each observation should have a `node` and a `next_node` aesthetic which provide information about which group in the y-direction. By default each row of the data frame is counted to calculate the size of flows. A manual flow value can be added with the `value` aesthetic.}
\usage{
geom_sankey(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  space = NULL,
  type = "sankey",
  width = 0.1,
  smooth = 8,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{provide you own mapping. both x and y need to be numeric.}

\item{data}{provide you own data}

\item{position}{change position}

\item{na.rm}{remove missing values}

\item{show.legend}{show legend in plot}

\item{space}{space between nodes in the y-direction}

\item{type}{either 'sankey' or 'alluvial'}

\item{width}{width of nodes}

\item{smooth}{how much smooth should the curve have? More means steeper curve.}

\item{inherit.aes}{should the geom inherits aestethics}

\item{...}{other arguments to be passed to the geom}
}
\value{
ggplot layer
}
\description{
geom_sankey

Creates a sankey plot which visualize flows between nodes. Each observation needs to have a `x` aesthetic as well as a `next_x` column which declares where that observation should flow.
Also each observation should have a `node` and a `next_node` aesthetic which provide information about which group in the y-direction. By default each row of the data frame is counted to calculate the size of flows. A manual flow value can be added with the `value` aesthetic.
}
\section{Aesthetics}{

geom_sankey understand the following aesthetics (required aesthetics are in
bold):

- **x0**
- **y0**
- **a**
- **b**
- **angle**
- m1
- m2
- color
- fill
- size
- linetype
- alpha
- lineend
}

