% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsankey.R
\name{geom_sankey_bump}
\alias{geom_sankey_bump}
\title{geom_sankey_bump}
\usage{
geom_sankey_bump(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  smooth = 8,
  type = "sankey",
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{provide you own mapping. both x and y need to be numeric.}

\item{data}{provide you own data}

\item{position}{change position}

\item{na.rm}{remove missing values}

\item{show.legend}{show legend in plot}

\item{smooth}{how much smooth should the curve have? More means steeper curve.}

\item{type}{either 'sankey' or 'alluvial'}

\item{inherit.aes}{should the geom inherits aestethics}

\item{...}{other arguments to be passed to the geo}
}
\value{
ggplot layer
}
\description{
Creates an alluvial plot which visualize flows between nodes. Each observation needs to have a `x` aesthetic as well as a `next_x` column which declares where that observation should flow.
Also each observation should have a `node` and a `next_node` aesthetic which provide information about which group in the y-direction.
}
\details{
Other important arguments is; `space` which proves the space between nodes in the y-direction; `shift` which shifts nodes in the y-direction.
}
