% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theta.R
\name{get_theta}
\alias{get_theta}
\title{Estimate theta via variogram fitting}
\usage{
get_theta(s, Y, dists = NULL, max_range = NULL)
}
\arguments{
\item{s}{spatial location}

\item{Y}{data}

\item{dists}{breakpoints for bins}

\item{max_range}{maximum range}
}
\value{
Matern covariance parameters
}
\description{
Estimate theta via variogram fitting
}
\keyword{internal}
