% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{get_vars}
\alias{get_vars}
\title{Get used vars in a ggplot object}
\usage{
get_vars(p, reverse, verbose = FALSE)
}
\arguments{
\item{p}{A \code{ggplot} object.}

\item{reverse}{If \code{TRUE} then will return unused vars.}

\item{verbose}{Whether to print messages.}
}
\description{
Get used vars in a ggplot object
}
