% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ifnb_sub}
\alias{ifnb_sub}
\title{A subsetted version of 'ifnb' datasets}
\format{
A \code{Seurat} object.
}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\description{
Human PBMC control/IFNB-stimulated dataset
}
\examples{
\dontrun{
if (interactive()) {
  if (!require("SeuratData", quietly = TRUE)) {
    devtools::install_github("satijalab/seurat-data")
  }
  library(SeuratData)
  library(Seurat)
  suppressWarnings(InstallData("ifnb"))
  data("ifnb")
  set.seed(11)
  cells_sub <- unlist(lapply(split(colnames(ifnb), ifnb$stim), function(x) sample(x, size = 1000)))
  ifnb_sub <- subset(ifnb, cells = cells_sub)
  ifnb_sub <- ifnb_sub[rowSums(ifnb_sub@assays$RNA@counts) > 0, ]
  ifnb_sub <- UpdateSeuratObject(ifnb_sub)
  # usethis::use_data(ifnb_sub, compress = "xz")
}
}
}
\concept{data}
