% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infIsNA}
\alias{infIsNA}
\title{Remove infinite data}
\usage{
infIsNA(obj, i)
}
\arguments{
\item{obj}{A `Features` object}

\item{i}{A `numeric()` or `character()` vector indicating from which assays 
the `rowData` should be taken.}
}
\value{
A `Features` object
}
\description{
This function coinsiders any infinite value as missing data. So, any value
in assay `i` that return `TRUE` to `is.infinite` will be replaced by `NA`.
}
\examples{
TODO
}
