% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installed_Python_pkgs}
\alias{installed_Python_pkgs}
\title{Show all the python packages in the environment}
\usage{
installed_Python_pkgs(envname = NULL, conda = "auto")
}
\arguments{
\item{envname}{The name of a conda environment.}

\item{conda}{The path to a conda executable. Use \code{"auto"} to allow SCP to automatically find an appropriate conda binary.}
}
\description{
Show all the python packages in the environment
}
