% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interval_score}
\alias{interval_score}
\title{Calculate interval score}
\usage{
interval_score(l, u, x, alpha = 0.05)
}
\arguments{
\item{l}{interval lower bound}

\item{u}{interval upper bound}

\item{x}{data}

\item{alpha}{significance level. Defaults to 0.05.}
}
\value{
a value of interval score
}
\description{
Calculate interval score
}
\examples{
eps = 0.05
x = rnorm(1000)
u = qnorm(1-eps/2)
l = qnorm(eps/2)
interval_score(l, u, x, alpha = eps)

}
\references{
\url{https://www.stat.washington.edu/raftery/Research/PDF/Gneiting2007jasa.pdf}, section 6.2
}
\keyword{internal}
