% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invoke}
\alias{invoke}
\title{Invoke a function with a list of arguments}
\usage{
invoke(.fn, .args = list(), ..., .env = caller_env())
}
\arguments{
\item{.fn}{A function, or function name as a string.}

\item{.args}{A list of arguments.}

\item{...}{Other arguments passed to the function.}

\item{.env}{Environment in which to evaluate the call. This will be most useful if .fn is a string, or the function has side-effects.}
}
\description{
Invoke a function with a list of arguments
}
