% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{isOutlier}
\alias{isOutlier}
\title{Detect outliers using MAD(Median and Median Absolute Deviation) method}
\usage{
isOutlier(
  x,
  nmads = 2.5,
  constant = 1.4826,
  type = c("both", "lower", "higher")
)
}
\arguments{
\item{type}{}
}
\description{
Detect outliers using MAD(Median and Median Absolute Deviation) method
}
