% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_pred.R
\name{krige_pred}
\alias{krige_pred}
\title{Kriging Prediction Function}
\usage{
krige_pred(
  s0,
  s,
  Y,
  alpha = 0.05,
  thetaHat = NULL,
  interval = FALSE,
  return_sd = FALSE
)
}
\arguments{
\item{s0}{prediction location}

\item{s}{an \eqn{n \times d}{n x d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{alpha}{significance level. Defaults to 0.05.}

\item{thetaHat}{spatial covariance parameters as in \code{\link{mat_cov}}.
If not given, emperical variogram is used to estimate \code{thetaHat}.}

\item{interval}{logical; if TRUE, \code{Krige_pred} function returns prediction interval;
if FALSE, \code{Krige_pred} function returns point prediction interval. Defaults to FALSE.}

\item{return_sd}{logical; if TRUE, \code{Krige_pred} function returns
standard deviation along with the point prediction. Defaults to FALSE.}
}
\value{
a value of point prediction if \code{interval} is FALSE
or a vector of lower and upper bounds of the Kriging prediction intervalif \eqn{interval} is TRUE.
}
\description{
Kriging Prediction Function
}
\examples{
N = 21; n = N^2
S = seq(0,1,length=N)
s = expand.grid(S,S)
d = as.matrix(dist(s))

theta        = c(0,3,0.1,0.7)
names(theta) = c("Nugget","PartialSill","Range","Smoothness")
C = mat_cov(d,theta)
X = t(chol(C))\%*\%rnorm(n)
Y = X^3 + rnorm(n)

s0 = c(0.5, 0.5)
krige_pred(s0,s,Y)
krige_pred(s0,s,Y,interval=TRUE)
}
