% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lifemap_cell}
\alias{lifemap_cell}
\alias{lifemap_compartment}
\alias{lifemap_organ}
\title{Embryonic Development Database from LifeMap Discovery}
\source{
\url{https://discovery.lifemapsc.com/in-vivo-development/cellular}
}
\usage{
lifemap_cell

lifemap_compartment

lifemap_organ
}
\description{
Embryonic Development Database from LifeMap Discovery
}
\concept{data}
