% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{ListDB}
\alias{ListDB}
\title{ListDB}
\usage{
ListDB(species = "Homo_sapiens", db = NULL)
}
\arguments{
\item{db}{}
}
\description{
ListDB
}
\examples{
ListDB(species = "Homo_sapiens")
ListDB(species = "Mus_musculus")

}
