% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{ListEnrichmentDB}
\alias{ListEnrichmentDB}
\title{ListEnrichmentDB}
\usage{
ListEnrichmentDB(
  species = "Homo_sapiens",
  enrichment = c("GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway", "Reactome",
    "ProteinComplex", "PFAM", "Chromosome")
)
}
\arguments{
\item{enrichment}{}
}
\description{
ListEnrichmentDB
}
