% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{LR_diff}
\alias{LR_diff}
\title{Likelihood Ratio Test for Differential Parameters}
\usage{
LR_diff(time1, y1, time2, y2, period, FN = TRUE, type = "amplitude")
}
\arguments{
\item{time1}{Time points for condition 1}

\item{y1}{Expression values for condition 1}

\item{time2}{Time points for condition 2}

\item{y2}{Expression values for condition 2}

\item{period}{Period}

\item{FN}{Finite sample correction}

\item{type}{Parameter type: "amplitude", "phase", "basal"}
}
\value{
List with p-value and parameter estimates
}
\description{
Likelihood Ratio Test for Differential Parameters
}
