% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lscp.R
\name{lscp}
\alias{lscp}
\title{Local Spatial Conformal Prediction (LSCP) At a Single Input Location}
\usage{
lscp(
  s0,
  s,
  Y,
  eta = Inf,
  m = NULL,
  pred_fun = krige_pred,
  theta = NULL,
  alpha = 0.05,
  dfun = "std_residual2",
  precision = 0.01
)
}
\arguments{
\item{s0}{prediction location}

\item{s}{an \eqn{n \times d}{n x d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{eta}{numerical value of the kernel bandwidth for the weight schema in conformal prediction. Defauls to \eqn{Inf} meaning equal weight on surrounding \eqn{m} points.}

\item{m}{an postive integer representing the number of nearest locations used for prediction. Depends on eta.}

\item{alpha}{significance level. Defaults to 0.05.}

\item{dfun}{non-conformity measure}

\item{precision}{Defaults to 0.01.}
}
\value{

}
\description{
This function provides the local conformal prediction interval for spatial location \code{s0} given obserations \code{s, Y}.
}
