% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsankey.R
\name{make_long}
\alias{make_long}
\title{make_long}
\usage{
make_long(.df, ..., value = NULL)
}
\arguments{
\item{.df}{a data frame}

\item{...}{unquoted columnnames of df that you want to include in the plot.}

\item{value}{if each row have a weight this weight could be kept by providing column name of weight.}
}
\value{
a longer data frame
}
\description{
Prepares a 'wide' data frame into a format that `geom_sankey` or `geom_alluvial` understands. Useful to show flows between dimensions in dataset.
}
