% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_cov.R
\name{mat_cov}
\alias{mat_cov}
\title{Matern covariance function}
\usage{
mat_cov(d, theta)
}
\arguments{
\item{d}{a numeric distance, a vector of distances, or a distance matrix}

\item{theta}{Matern covariance parameters}
}
\value{
a numeric covariance, a vector of covariances, or a covariance matrix in the same size of \eqn{d}
}
\description{
Matern covariance function
}
