% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergeRows}
\alias{MergeRows}
\title{Merge Rows of a Matrix by Groupings}
\usage{
MergeRows(x, groupings)
}
\arguments{
\item{x}{A matrix.}

\item{groupings}{A vector specifying the groupings for merging rows.}
}
\description{
MergeRows is a function that sum the values of rows based on specified groupings.
}
