% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{MNN_integrate}
\alias{MNN_integrate}
\title{MNN_integrate}
\usage{
MNN_integrate(
  srtMerge = NULL,
  batch = "orig.ident",
  append = TRUE,
  srtList = NULL,
  do_normalization = NULL,
  normalization_method = "logCPM",
  do_HVF_finding = TRUE,
  HVF_source = "separate",
  HVF_method = "vst",
  nHVF = 2000,
  HVF = NULL,
  do_scaling = TRUE,
  vars_to_regress = NULL,
  regression_model = "linear",
  liner_reduction = "pca",
  liner_reduction_dims = 100,
  liner_reduction_dims_use = NULL,
  force_liner_reduction = FALSE,
  nonliner_reduction = "umap",
  nonliner_reduction_dims = c(2, 3),
  force_nonliner_reduction = TRUE,
  do_cluster_finding = TRUE,
  cluster_algorithm = "louvain",
  cluster_resolution = 0.6,
  cluster_reorder = TRUE,
  mnnCorrect_params = list(),
  seed = 11
)
}
\arguments{
\item{srtMerge}{}

\item{...}{}
}
\description{
MNN_integrate
}
