% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mq}
\alias{mq}
\title{MaxQuant example data}
\format{an MSnSet with MS intensities for 1000 peptides in 11 TMT channels}
\source{
The original data can be downloaded from the
\href{https://scope2.slavovlab.net/docs/data}{Slavov Lab} website, in the
\href{https://scope2.slavovlab.net/docs/data}{Data repository 1}.
}
\usage{
data("mq")
}
\description{
Subset of the MaxQuant evidence file from the Specht et al. 2019
preprint.
}
\references{
Specht, Harrison, Edward Emmott, Toni Koller, and Nikolai Slavov.
2019. “High-Throughput Single-Cell Proteomics Quantifies the Emergence of
Macrophage Heterogeneity.” bioRxiv (\href{https://doi.org/10.1101/665307}
{DOI}).
}
\keyword{datasets}
