% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{PAGAPlot}
\alias{PAGAPlot}
\title{PAGA plot}
\usage{
PAGAPlot(
  srt,
  paga = srt@misc$paga,
  reduction = NULL,
  dims = c(1, 2),
  show_transition = FALSE,
  node_palette = "Paired",
  node_size = 4,
  node_alpha = 1,
  node_highlight = NULL,
  node_highlight_color = "red",
  label = FALSE,
  label.size = 3.5,
  label.fg = "white",
  label.bg = "black",
  label.bg.r = 0.1,
  label_insitu = FALSE,
  label_repel = FALSE,
  label_repulsion = 20,
  label_point_size = 1,
  label_point_color = "black",
  label_segment_color = "black",
  edge_threshold = 0.01,
  use_triangular = c("upper", "lower", "both"),
  edge_line = c("straight", "curved"),
  edge_line_curvature = 0.3,
  edge_line_angle = 90,
  edge_size = c(0.2, 1),
  edge_color = "grey40",
  edge_alpha = 0.5,
  edge_shorten = 0,
  edge_offset = 0,
  edge_highlight = NULL,
  edge_highlight_color = "red",
  transition_threshold = 0.01,
  transition_line = c("straight", "curved"),
  transition_line_curvature = 0.3,
  transition_line_angle = 90,
  transition_size = c(0.2, 1),
  transition_color = "black",
  transition_alpha = 1,
  transition_arrow_type = "closed",
  transition_arrow_angle = 20,
  transition_arrow_length = unit(0.02, "npc"),
  transition_shorten = 0.05,
  transition_offset = 0,
  transition_highlight = NULL,
  transition_highlight_color = "red",
  theme_use = "theme_scp",
  aspect.ratio = 1,
  title = "PAGA",
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  lab_cex = 1,
  xlen_npc = 0.15,
  ylen_npc = 0.15,
  legend.position = "right",
  legend.direction = "vertical",
  return_layer = FALSE
)
}
\description{
PAGA plot
}
\examples{
data("pancreas1k")
pancreas1k <- RunPAGA(srt = pancreas1k, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP", return_seurat = TRUE)
PAGAPlot(pancreas1k)
PAGAPlot(pancreas1k, reduction = "PCA")
PAGAPlot(pancreas1k, label = TRUE)
PAGAPlot(pancreas1k, label = TRUE, label_insitu = TRUE)
PAGAPlot(pancreas1k, label = TRUE, label_insitu = TRUE, label_repel = TRUE)
PAGAPlot(pancreas1k, edge_line = "curved")
PAGAPlot(pancreas1k, node_size = "GroupSize")
PAGAPlot(pancreas1k, node_highlight = "Ductal")
PAGAPlot(pancreas1k, edge_highlight = paste("Pre-endocrine", "Delta", sep = "-"))
PAGAPlot(pancreas1k, edge_highlight = paste("Pre-endocrine", levels(pancreas1k$SubCellType), sep = "-"))
PAGAPlot(pancreas1k, use_triangular = "both", edge_highlight = paste("Pre-endocrine", levels(pancreas1k$SubCellType), sep = "-"))

pancreas1k <- RunSCVELO(srt = pancreas1k, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP", return_seurat = TRUE)
PAGAPlot(pancreas1k, show_transition = TRUE)
}
