% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{panc8_sub}
\alias{panc8_sub}
\title{A subsetted version of human 'panc8' datasets}
\format{
A \code{Seurat} object.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE81076} \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi} \url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-5061/} \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi} \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi}
}
\description{
Human pancreatic islet cell datasets produced across four technologies, CelSeq (GSE81076) CelSeq2 (GSE85241), Fluidigm C1 (GSE86469), and SMART-Seq2 (E-MTAB-5061) from \href{https://github.com/satijalab/seurat-data}{SeuratData} package. For each data set in `panc8`, 200 cells were downsampled to form the \code{panc8_sub} dataset.
}
\examples{
\dontrun{
if (interactive()) {
  data("pancreas_sub")
  if (!require("SeuratData", quietly = TRUE)) {
    devtools::install_github("satijalab/seurat-data")
  }
  library(SeuratData)
  library(Seurat)
  suppressWarnings(InstallData("panc8"))
  data("panc8")
  set.seed(11)
  cells_sub <- unlist(lapply(split(colnames(panc8), panc8$dataset), function(x) sample(x, size = 200)))
  panc8_sub <- subset(panc8, cells = cells_sub)
  panc8_sub <- panc8_sub[rowSums(panc8_sub@assays$RNA@counts) > 0, ]
  panc8_sub <- panc8_sub[toupper(rownames(panc8_sub)) \%in\% toupper(rownames(pancreas_sub)), ]
  panc8_sub <- UpdateSeuratObject(panc8_sub)
  # usethis::use_data(panc8_sub, compress = "xz")
}
}
}
\concept{data}
