% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{panel_fix_single}
\alias{panel_fix_single}
\title{Set the panel width/height for the single plot.}
\usage{
panel_fix_single(
  x,
  panel_index = NULL,
  respect = NULL,
  width = NULL,
  height = NULL,
  margin = 0.5,
  units = "in",
  raster = FALSE,
  dpi = 300
)
}
\arguments{
\item{x}{}

\item{dpi}{}
}
\description{
Set the panel width/height for the single plot.
}
