% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility.R
\name{plausibility}
\alias{plausibility}
\title{calculate plausibility for \code{Y0}}
\usage{
plausibility(
  Y0,
  s0,
  s,
  Y,
  global = TRUE,
  eta = Inf,
  m = NULL,
  pred_fun = krige_pred,
  dfun = c("residual2", "abs_residual", "std_residual2", "std_abs_residual")
)
}
\arguments{
\item{Y0}{a scalar or a vector}

\item{s0}{prediction location, a numeric vector with \code{length = 2}.}

\item{s}{an \eqn{n \times 2}{n x 2} \code{matrix} or \code{data.frame} with two coordinates of \eqn{n} locations.}

\item{Y}{a vector with \eqn{n} values corresponding to \code{Y(s)}.}

\item{global}{logical; if \code{TRUE} , \code{scp} function returns the result of global spatial conformal prediction (GSCP);
if \code{FALSE}, \code{scp} function returns the result of local spatial conformal prediction (LSCP) and users need to
specify \code{eta < Inf} or \code{m} \eqn{\leq} \code{n}. Defaults to \code{TRUE}.}

\item{eta}{kernel bandwidth for weight schema, a positve scalar with smaller value meaning more localized procedure.
Defauls to \code{Inf}, which puts equal weight on surrounding \code{m} points.}

\item{m}{an postive integer representing the number of nearest locations to use for prediction.
Default to \code{NULL}. If \code{global = TRUE}, \code{m = n};
if \code{global = FALSE} and \code{m} is not specified, \code{m} would be determined by \code{eta}.}

\item{pred_fun}{spatial prediction function with inputs being \code{s0, s, Y} and ouputs being predicted \code{Y(s0)}
(and its standard error). Defaults to \code{\link{krige_pred}}.}

\item{dfun}{non-conformity measure with four options.
In which, \code{"residual2"} (default) represents squared residual,
\code{"std_residual2"} represents standardized squared residual,
\code{"abs_residual"} represents absolute residual,
and \code{"std_abs_residual"} represents standardized absolute residual.}
}
\value{
The output is a scalar or a vector with plausibility values for \code{Y0}. The numbers are between 0 and 1.
}
\description{
This function provides the plausibility of \code{Y(s0)} being \code{Y0},
given observations \code{s} and \code{Y},
using spatial conformal prediction algorithms.
}
\examples{
## To predict Y(s0), where s0 = c(0.5,0.5), using sample data
## What's the plausibility if Y(s0) = 0? Y(s0) = 1.5?

#?sample_data
s0 = c(0.5,0.5)
s  = sample_data$s
Y  = sample_data$Y

# plausibility for Y(s0) = 0: 0.8744795
plausibility(Y0=0,s0=s0,s=s,Y=Y)

# plausibility for Y(s0) = 1.5: 0.4669839
plausibility(Y0=1.5,s0=s0,s=s,Y=Y)

# plausibility for a sequence of Y0's
plausibility(Y0=seq(0,1,0.1),s0=s0,s=s,Y=Y)

}
\references{
to be entered
}
\seealso{
\code{\link{plausibility_contour}}
}
\author{
Huiying Mao, \email{hmao@samsi.info}, Brian Reich \email{bjreich@ncsu.edu}
}
