% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_contour.R
\name{plausibility_contour}
\alias{plausibility_contour}
\title{plausibility_contour calculation}
\usage{
plausibility_contour(Y_cand = NULL, s0, s, Y, global, eta, m, pred_fun, dfun)
}
\arguments{
\item{dfun}{}
}
\value{

}
\description{
plausibility_contour calculation
}
