% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_contour.R
\name{plot.plausibility_contour}
\alias{plot.plausibility_contour}
\title{Plot a plausibility_contour object}
\usage{
\method{plot}{plausibility_contour}(df, ...)
}
\arguments{
\item{df}{a plausibility_contour object, returned by \code{\link{plausibility_contour}}}

\item{...}{}
}
\description{
Plot a plausibility_contour object
}
\keyword{internal}
