% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_power_curves}
\alias{plot_power_curves}
\title{Plot Power Curves}
\usage{
plot_power_curves(power_results, target_power = 0.8)
}
\arguments{
\item{power_results}{Results from power_single_condition}

\item{target_power}{Target power level (default 0.8)}
}
\value{
ggplot object
}
\description{
Visualize power vs sample size for different methods
}
