% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_stratified_power}
\alias{plot_stratified_power}
\title{Plot Stratified Power}
\usage{
plot_stratified_power(stratified_results)
}
\arguments{
\item{stratified_results}{Results from stratified_power}
}
\value{
ggplot object
}
\description{
Bar plot of power by effect size stratum
}
