% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_two_group_heatmap}
\alias{plot_two_group_heatmap}
\title{Plot Two-Group Power Heatmap}
\usage{
plot_two_group_heatmap(power_matrix, n_A_range, n_B_range)
}
\arguments{
\item{power_matrix}{Power matrix from power_two_group}

\item{n_A_range}{n_A values}

\item{n_B_range}{n_B values}
}
\value{
ggplot object
}
\description{
Heatmap of power for different (n_A, n_B) combinations
}
