% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dr_power.R
\name{plotDensityResults}
\alias{plotDensityResults}
\title{Generate Sampling Density Plot}
\usage{
plotDensityResults(results_file, output_file = NULL)
}
\arguments{
\item{results_file}{Path to saved density results}

\item{output_file}{Path to save PDF figure}
}
\value{
Invisible density results
}
\description{
Generate Sampling Density Plot
}
