% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{PrepareEnrichmentDB}
\alias{PrepareEnrichmentDB}
\title{Prepare the database for enrichment analysis}
\usage{
PrepareEnrichmentDB(
  species = "Homo_sapiens",
  enrichment = c("GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway", "Reactome",
    "ProteinComplex", "PFAM", "Chromosome"),
  db_IDtypes = c("symbol", "entrez_id", "ensembl_id"),
  db_update = FALSE,
  Ensembl_version = 103,
  mirror = NULL
)
}
\arguments{
\item{species}{species}

\item{enrichment}{Enrichment database name.}

\item{db_update}{Whether update the database.}
}
\value{
A list containing the database.
}
\description{
Prepare the database for enrichment analysis
}
\examples{
if (interactive()) {
  db_list <- PrepareEnrichmentDB(species = "Homo_sapiens")
}
}
