% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PrepareVirtualEnv}
\alias{PrepareVirtualEnv}
\title{Prepare SCP python virtual environment}
\usage{
PrepareVirtualEnv(
  python = NULL,
  pipy_mirror = "https://pypi.org/simple/",
  remove_old = FALSE
)
}
\arguments{
\item{python}{Python path which is used to create the virtual environment.}

\item{pipy_mirror}{pipy mirrors. Default is "https://pypi.org/simple/". Options can be "https://pypi.tuna.tsinghua.edu.cn/simple", "http://mirrors.aliyun.com/pypi/simple/", "https://pypi.mirrors.ustc.edu.cn/simple/", etc.}

\item{remove_old}{Whether to remove the old SCP virtual environment.}
}
\description{
Prepare SCP python virtual environment
}
