% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{print.CircadianBioOptions}
\alias{print.CircadianBioOptions}
\title{Print a CircadianBioOptions object}
\usage{
\method{print}{CircadianBioOptions}(x, ...)
}
\arguments{
\item{x}{A \code{CircadianBioOptions} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a CircadianBioOptions object
}
