% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{print.CircadianBootstrapOptions}
\alias{print.CircadianBootstrapOptions}
\title{Print a CircadianBootstrapOptions object}
\usage{
\method{print}{CircadianBootstrapOptions}(x, ...)
}
\arguments{
\item{x}{A \code{CircadianBootstrapOptions} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a CircadianBootstrapOptions object
}
