% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{print.CircadianDesignOptions}
\alias{print.CircadianDesignOptions}
\title{Print a CircadianDesignOptions object}
\usage{
\method{print}{CircadianDesignOptions}(x, ...)
}
\arguments{
\item{x}{A \code{CircadianDesignOptions} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a CircadianDesignOptions object
}
