% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{print.SCPDiffResult}
\alias{print.SCPDiffResult}
\title{Print an SCPDiffResult object}
\usage{
\method{print}{SCPDiffResult}(x, ...)
}
\arguments{
\item{x}{An \code{SCPDiffResult} from \code{runDifferentialPower()}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an SCPDiffResult object
}
