% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{print.SCPRecommendResult}
\alias{print.SCPRecommendResult}
\title{Print an SCPRecommendResult object}
\usage{
\method{print}{SCPRecommendResult}(x, ...)
}
\arguments{
\item{x}{An \code{SCPRecommendResult} from \code{recommendDesign()}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an SCPRecommendResult object
}
