% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{print.SCPSingleResult}
\alias{print.SCPSingleResult}
\title{Print an SCPSingleResult object}
\usage{
\method{print}{SCPSingleResult}(x, ...)
}
\arguments{
\item{x}{An \code{SCPSingleResult} from \code{runSingleCohortGrid()}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an SCPSingleResult object
}
