% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{ProjectionPlot}
\alias{ProjectionPlot}
\title{Projection Plot}
\usage{
ProjectionPlot(
  srt_query,
  srt_ref,
  query_group = "orig.ident",
  ref_group = "orig.ident",
  query_reduction = "ref.umap",
  ref_reduction = srt_query[[query_reduction]]@misc[["reduction.model"]] \%||\% NULL,
  query_param = list(palette = "Dark2"),
  ref_param = list(palette = "Paired"),
  xlim = NULL,
  ylim = NULL,
  pt.size = 0.8,
  stroke.highlight = 0.5
)
}
\description{
Projection Plot
}
