% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{rayleigh_test_circular}
\alias{rayleigh_test_circular}
\title{Rayleigh test for uniform circular distribution}
\usage{
rayleigh_test_circular(angles, period = 24)
}
\arguments{
\item{angles}{Angles (in hours)}

\item{period}{Period (default 24)}
}
\value{
List with statistic and p-value
}
\description{
Rayleigh test for uniform circular distribution
}
