% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSCP.R
\name{readSCPfromDIANN}
\alias{readSCPfromDIANN}
\title{Read DIA-NN output as a QFeatures objects for single-cell
proteomics data}
\usage{
readSCPfromDIANN(
  colData,
  reportData,
  extractedData = NULL,
  ecol = "MS1.Area",
  multiplexing = "none",
  ...
)
}
\arguments{
\item{colData}{A data.frame or any object that can be coerced to a
data.frame. colData is expected to contains all the sample
annotations. We require the table to contain a column called
\code{File.Name} that links to the \code{File.Name} in the DIA-NN report
table. If \code{multiplexing = "mTRAQ"}, we require a second column
called \code{Label} that links the label to the sample (the labels
identified by DIA-NN can be retrieved from \verb{Modified Sequence}
column in the report table).}

\item{reportData}{A data.frame or any object that can be coerced
to a data.frame that contains the data from the \code{Report.tsv}
file generated by DIA-NN.}

\item{extractedData}{A data.frame or any object that can be coerced
to a data.frame that contains the data from the \verb{*_ms1_extracted.tsv}
file generated by DIA-NN. This argument is optional and is
only applicable for mulitplixed experiments}

\item{ecol}{A \code{character(1)} indicating which column in
\code{reportData} contains the quantitative information.}

\item{multiplexing}{A \code{character(1)} indicating the type of
multiplexing used in the experiment. Provide \code{"none"} if the
experiment is label-free (default). Available options are:
\code{"mTRAQ"}.}

\item{...}{Further arguments passed to \code{readSCP()}}
}
\value{
An instance of class QFeatures. The expression data of
each acquisition run is stored in a separate assay as a
SingleCellExperiment object.
}
\description{
This function takes the output tables from DIA-NN and converts them
into a QFeatures object using the scp framework.
}
