% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSingleCellExperiment}
\alias{readSingleCellExperiment}
\title{Read SingleCellExperiment from tabular data}
\usage{
readSingleCellExperiment(table, ecol, fnames, ...)
}
\arguments{
\item{table}{File or object holding the quantitative data. Can be
either a \code{character(1)} with the path to a text-based
spreadsheet (comma-separated values by default, but see \code{...})
or an object that can be coerced to a \code{data.frame}. It is
advised not to encode characters as factors.}

\item{ecol}{A \code{numeric} indicating the indices of the columns to
be used as assay values. Can also be a \code{character}
indicating the names of the columns. Caution must be taken if
the column names are composed of special characters like \code{(}
or \code{-} that will be converted to a \code{.} by the \code{read.csv}
function. If \code{ecol} does not match, the error message will
dislpay the column names as seen by the \code{read.csv} function.}

\item{fnames}{An optional \code{character(1)} or \code{numeric(1)}
indicating the column to be used as row names.}

\item{...}{Further arguments that can be passed on to \link{read.csv}
except \code{stringsAsFactors}, which is always \code{FALSE}.}
}
\value{
An instance of class \link{SingleCellExperiment}.
}
\description{
Convert tabular data from a spreadsheet or a \code{data.frame} into a
\code{SingleCellExperiment} object.
}
\seealso{
The code relies on \link[Features:readFeatures]{Features::readSummarizedExperiment}.
}
\author{
Laurent Gatto, Christophe Vanderaa
}
