% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RenameClusters}
\alias{RenameClusters}
\title{Rename clusters for the Seurat object}
\usage{
RenameClusters(
  srt,
  group.by,
  newnames = NULL,
  nameslist = list(),
  createnew = NULL
)
}
\arguments{
\item{srt}{}
}
\description{
Rename clusters for the Seurat object
}
\examples{
data("pancreas_sub")
levels(pancreas_sub@meta.data[["SubCellType"]])

# Rename all clusters
pancreas_sub <- RenameClusters(pancreas_sub, group.by = "SubCellType", newnames = letters[1:8], createnew = "newclusters")
pancreas_sub@meta.data[["newclusters"]] <- factor(pancreas_sub@meta.data[["newclusters"]], levels = letters[1:8])
ClassDimPlot(pancreas_sub, "newclusters")

# Rename specified clusters
pancreas_sub <- RenameClusters(pancreas_sub, group.by = "SubCellType", newnames = c("Alpha" = "a", "Beta" = "b"), createnew = "newclusters")
ClassDimPlot(pancreas_sub, "newclusters")

# Merge and rename clusters
pancreas_sub <- RenameClusters(pancreas_sub, group.by = "SubCellType", nameslist = list("EndocrineClusters" = c("Alpha", "Beta", "Epsilon", "Delta")), createnew = "newclusters")
ClassDimPlot(pancreas_sub, "newclusters")

}
