% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RenameFeatures}
\alias{RenameFeatures}
\title{Rename features for the Seurat object}
\usage{
RenameFeatures(srt, newnames = NULL, assays = NULL)
}
\arguments{
\item{srt}{A Seurat object.}

\item{newnames}{A vector with the same length of features in Seurat object, or characters named with old features.}

\item{assays}{Assays to rename.}
}
\description{
Rename features for the Seurat object
}
\examples{
data("panc8_sub")
head(rownames(panc8_sub))
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(capitalize(rownames(panc8_sub), force_tolower = TRUE))
panc8_rename <- RenameFeatures(panc8_sub, newnames = genenames)
head(rownames(panc8_rename))

}
