% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowDataToDF}
\alias{rowDataToDF}
\title{Extract the \code{rowData} of a \code{QFeatures} object to a
\code{DataFrame}}
\usage{
rowDataToDF(object, i, vars)
}
\arguments{
\item{object}{A \code{QFeatures} object}

\item{i}{A \code{numeric()} or \code{character()} vector indicating from
which assays the \code{rowData} should be taken.}

\item{vars}{A \code{character()} vector indicating which variables from
the \code{rowData} should be extracted.}
}
\value{
A \code{DataFrame} object with the \code{rowData} row-binded over
the required assays.
}
\description{
This function is deprecated. You should rather use
[QFeatures::rbindRowData\verb{]}
}
