% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{RunCellQC}
\alias{RunCellQC}
\title{Run cell-level quality control for single cell RNA-seq data.}
\usage{
RunCellQC(
  srt,
  assay = "RNA",
  qc_metrics = c("doublets", "outlier", "umi", "gene", "mito", "ribo", "ribo_mito_ratio",
    "species"),
  return_filtered = FALSE,
  db_method = "scDblFinder",
  db_rate = ncol(srt)/1000 * 0.01,
  outlier_cutoff = c("log10_nCount:lower:2.5", "log10_nCount:higher:5",
    "log10_nFeature:lower:2.5", "log10_nFeature:higher:5", "featurecount_dist:lower:2.5"),
  outlier_n = 1,
  UMI_threshold = 3000,
  gene_threshold = 1000,
  mito_threshold = 20,
  mito_pattern = c("MT-", "Mt-", "mt-"),
  mito_gene = NULL,
  ribo_threshold = 50,
  ribo_pattern = c("RP[SL]\\\\d+\\\\w{0,1}\\\\d*$", "Rp[sl]\\\\d+\\\\w{0,1}\\\\d*$",
    "rp[sl]\\\\d+\\\\w{0,1}\\\\d*$"),
  ribo_gene = NULL,
  ribo_mito_ratio_range = c(1, Inf),
  species = NULL,
  species_gene_prefix = NULL,
  species_percent = 95,
  seed = 11
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{assay}{Assay to use.}

\item{qc_metrics}{QC metrics applied.}

\item{return_filtered}{Whether to return a cell-filtered \code{Seurat} object.}

\item{db_method}{Doublet-calling methods used. Can be one of \code{scDblFinder}, \code{Scrublet}, \code{DoubletDetection}, \code{scds_cxds}, \code{scds_bcds}, \code{scds_hybrid}}

\item{db_rate}{The expected doublet rate. By default this is assumed to be 1\% per thousand cells captured (so 4\% among 4000 thousand cells), which is appropriate for 10x datasets.}

\item{outlier_cutoff}{MAD outlier metrics. See \link[scuttle]{isOutlier}.}

\item{outlier_n}{Minimum number of outlier metrics that meet the conditions for determining outlier cells. Default is 1.}

\item{UMI_threshold}{UMI number threshold. Cells that exceed this threshold will be considered as kept. Default is 3000.}

\item{gene_threshold}{Gene number threshold. Cells that exceed this threshold will be considered as kept. Default is 1000.}

\item{mito_threshold}{Percentage of UMI counts of mitochondrial genes. Cells that exceed this threshold will be considered as discarded. Default is 20.}

\item{mito_pattern}{Regex patterns to match the mitochondrial genes.}

\item{mito_gene}{A defined mitochondrial genes. If features provided, will ignore the \code{mito_pattern} matching.}

\item{ribo_threshold}{Percentage of UMI counts of ribosomal genes. Cells that exceed this threshold will be considered as discarded. Default is 50.}

\item{ribo_pattern}{Regex patterns to match the ribosomal genes.}

\item{ribo_gene}{A defined ribosomal genes. If features provided, will ignore the \code{ribo_pattern} matching.}

\item{species}{Species used as the suffix of the QC metrics. The first is the species of interest. Default is \code{NULL}.}

\item{species_gene_prefix}{Species gene prefix used to calculate QC metrics for each species. Default is \code{NULL}.}

\item{species_percent}{Percentage of UMI counts of the first species. Cells that exceed this threshold will be considered as kept. Default is 95.}

\item{seed}{Set a random seed. Default is 11.}
}
\value{
Returns Seurat object with the QC results stored in the meta.data slot.
}
\description{
In CellQC, doublet-calling will be run first and then reject abnormal cell data based on median absolute deviation (MAD) outliers.
After doublet-calling and outlier filtering, CellQC will perform general QC (gene count, UMI count, etc.)
and reject cell data for non-species of interest based on the proportion and number of UMIs for the species.
}
\note{
General quality control usually uses metrics such as gene count, UMI count, etc.
In addition, ribo content and mito content can be used as QC indicators:
mito content can be used as an indication of apoptosis,
with a general threshold of 20% and less than 10% mito content in high quality cells;
ribo content is cell type dependent, with some cell types having even more than 50% ribo content;
however, ribo content in single cell data can also indicate whether the empty droplets,
and the ribo content in empty droplets is usually high and the mito content is low instead.
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunCellQC(pancreas_sub)
}
