% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RunDimReduction}
\alias{RunDimReduction}
\title{RunDimReduction}
\usage{
RunDimReduction(
  srt,
  prefix = "",
  features = NULL,
  assay = NULL,
  slot = "data",
  liner_reduction = NULL,
  liner_reduction_dims = 100,
  liner_reduction_params = list(),
  force_liner_reduction = FALSE,
  nonliner_reduction = NULL,
  nonliner_reduction_dims = 2,
  reduction_use = NULL,
  reduction_dims = NULL,
  neighbor_use = NULL,
  graph_use = NULL,
  distance_use = NULL,
  nonliner_reduction_params = list(),
  force_nonliner_reduction = TRUE,
  verbose = TRUE,
  seed = 11
)
}
\arguments{
\item{liner_reduction}{"pca", "ica", "nmf", "mds", "glmpca"}

\item{nonliner_reduction}{"umap", "umap-naive",  "tsne",  "dm",   "phate", "pacmap", "trimap", "largevis"}

\item{seed}{}
}
\description{
RunDimReduction
}
