% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunDM}
\alias{RunDM}
\alias{RunDM.dist}
\alias{RunDM.matrix}
\alias{RunDM.Seurat}
\title{Run DiffusionMap(DM)}
\usage{
RunDM(object, ...)

\method{RunDM}{dist}(
  object,
  exprs,
  ndcs = 2,
  sigma = "local",
  k = NULL,
  slot = "scale.data",
  assay = NULL,
  reduction.key = "DM_",
  seed.use = 42,
  verbose = TRUE,
  ...
)

\method{RunDM}{matrix}(
  object,
  ndcs = 2,
  sigma = "local",
  k = NULL,
  knn_method = "find_knn",
  dist.method = "euclidean",
  assay = NULL,
  slot = "scale.data",
  reduction.key = "DM_",
  seed.use = 42,
  verbose = TRUE,
  ...
)

\method{RunDM}{Seurat}(
  object,
  reduction = "pca",
  dims = 1:30,
  assay = NULL,
  slot = "scale.data",
  features = NULL,
  ndcs = 2,
  sigma = "local",
  k = NULL,
  knn_method = "find_knn",
  dist.method = "euclidean",
  reduction.name = "dm",
  reduction.key = "DM_",
  seed.use = 42,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to \link[destiny]{DiffusionMap}.}

\item{ndcs}{Total Number of DM to compute and store (2 by default).}

\item{slot}{Name of slot dm is being run on.}

\item{assay}{Name of Assay dm is being run on.}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. "DM_" by default.}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}

\item{verbose}{Show a progressbar and other progress information}

\item{reduction.name}{dimensional reduction name, 'dm' by default}
}
\value{
Returns Seurat object with the DM calculation stored in the reductions slot.
}
\description{
Run a DM dimensionality reduction.
}
\concept{dimensional_reduction}
