% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{RunDoubletCalling}
\alias{RunDoubletCalling}
\title{Run doublet-calling for single cell RNA-seq data.}
\usage{
RunDoubletCalling(
  srt,
  assay = "RNA",
  db_method = "scDblFinder",
  db_rate = ncol(srt)/1000 * 0.01,
  ...
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{assay}{The name of the assay to be used for doublet-calling. Default is "RNA".}

\item{db_method}{Doublet-calling methods used. Can be one of \code{scDblFinder}, \code{Scrublet}, \code{DoubletDetection}, \code{scds_cxds}, \code{scds_bcds}, \code{scds_hybrid}}

\item{db_rate}{The expected doublet rate. By default this is assumed to be 1\% per thousand cells captured (so 4\% among 4000 thousand cells), which is appropriate for 10x datasets.}

\item{...}{Arguments passed to the corresponding doublet-calling method.}
}
\value{
Returns Seurat object with the doublet prediction results and prediction scores stored in the meta.data slot.
}
\description{
Identification of heterotypic (or neotypic) doublets in single-cell RNAseq data.
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunDoubletCalling(pancreas_sub, db_method = "scDblFinder")
CellDimPlot(pancreas_sub, reduction = "umap", group.by = "db.scDblFinder_class")
FeatureDimPlot(pancreas_sub, reduction = "umap", features = "db.scDblFinder_score")
}
