% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDynamicEnrichment}
\alias{RunDynamicEnrichment}
\title{RunDynamicEnrichment}
\usage{
RunDynamicEnrichment(
  srt,
  lineages,
  score_method = "AUCell",
  ncore = 1,
  slot = "data",
  assay = "RNA",
  min_expcells = 20,
  r.sq = 0.2,
  dev.expl = 0.2,
  padjust = 0.05,
  geneID = NULL,
  IDtype = "symbol",
  species = "Homo_sapiens",
  enrichment = "GO_BP",
  db_update = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE,
  seed = 11
)
}
\description{
RunDynamicEnrichment
}
\examples{
if (interactive()) {
  data("pancreas1k")
  pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "UMAP")
  pancreas1k <- RunDynamicFeatures(pancreas1k, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
  ht_result <- DynamicHeatmap(
    srt = pancreas1k,
    lineages = c("Lineage1", "Lineage2"),
    cell_annotation = "SubCellType",
    n_cluster = 6, reverse_ht = 1, use_raster = FALSE
  )
  ht_result$plot

  pancreas1k <- RunDynamicEnrichment(
    srt = pancreas1k,
    lineages = c("Lineage1", "Lineage2"),
    score_method = "AUCell",
    enrichment = "GO_BP",
    species = "Mus_musculus"
  )
  ht_result <- DynamicHeatmap(
    srt = pancreas1k, assay = "GO_BP", use_fitted = TRUE,
    lineages = c("Lineage1_GO_BP", "Lineage2_GO_BP"),
    cell_annotation = "SubCellType",
    n_cluster = 6, reverse_ht = 1,
    height = 5, width = 7, use_raster = FALSE
  )
  ht_result$plot
}
}
