% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDynamicFeatures}
\alias{RunDynamicFeatures}
\title{RunDynamicFeatures}
\usage{
RunDynamicFeatures(
  srt,
  lineages,
  features = NULL,
  suffix = lineages,
  n_candidates = 1000,
  minfreq = 5,
  family = NULL,
  slot = "counts",
  assay = "RNA",
  libsize = NULL,
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE,
  seed = 11
)
}
\arguments{
\item{seed}{}
}
\description{
RunDynamicFeatures
}
\examples{
data("pancreas1k")
pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "UMAP")
pancreas1k <- RunDynamicFeatures(pancreas1k, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
ht_result <- DynamicHeatmap(
  srt = pancreas1k,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_cluster = 6, reverse_ht = "Lineage1",
  height = 5, width = 7, use_raster = FALSE
)
ht_result$plot

p <- DynamicPlot(
  srt = pancreas1k,
  lineages = c("Lineage1", "Lineage2"),
  features = c("Nnat", "Irx1"),
  group.by = "SubCellType",
  compare_lineages = TRUE,
  compare_features = FALSE
)
p
}
