% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunFR}
\alias{RunFR}
\alias{RunFR.Seurat}
\alias{RunFR.default}
\title{Create the force-directed layout developed by Fruchterman and Reingold.}
\usage{
RunFR(object, ...)

\method{RunFR}{Seurat}(
  object,
  reduction = NULL,
  dims = NULL,
  features = NULL,
  assay = DefaultAssay(object = object),
  slot = "data",
  graph = NULL,
  neighbor = NULL,
  k.param = 20,
  ndim = 2,
  niter = 500,
  seed.use = 42L,
  verbose = TRUE,
  reduction.name = "FR",
  reduction.key = "FR_",
  ...
)

\method{RunFR}{default}(
  object,
  ndim = 2,
  niter = 500,
  reduction.key = "FR_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}

\item{n_components}{}
}
\description{
Create the force-directed layout developed by Fruchterman and Reingold.
}
\concept{dimensional_reduction}
