% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunGSEA}
\alias{RunGSEA}
\title{Perform the enrichment analysis(GSEA) on the genes}
\usage{
RunGSEA(
  geneID = NULL,
  geneScore = NULL,
  geneID_groups = NULL,
  IDtype = "symbol",
  result_IDtype = "symbol",
  species = "Homo_sapiens",
  enrichment = "GO_BP",
  db_IDtype = "symbol",
  db_update = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  scoreType = "std",
  GO_simplify = FALSE,
  GO_simplify_padjustCutoff = 0.05,
  simplify_method = "Rel",
  simplify_similarityCutoff = 0.7,
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE
)
}
\arguments{
\item{geneID}{geneID}

\item{geneID_groups}{geneID_groups}

\item{IDtype}{IDtype}

\item{result_IDtype}{result_IDtype}

\item{species}{species}

\item{enrichment}{enrichment}

\item{db_update}{db_update}

\item{TERM2GENE}{TERM2GENE}

\item{TERM2NAME}{TERM2NAME}

\item{minGSSize}{minGSSize}

\item{maxGSSize}{maxGSSize}

\item{GO_simplify}{GO_simplify}

\item{GO_simplify_padjustCutoff}{GO_simplify_padjustCutoff}

\item{simplify_method}{simplify_method}

\item{simplify_similarityCutoff}{simplify_similarityCutoff}

\item{universe}{universe}

\item{cl}{cores}
}
\description{
Perform the enrichment analysis(GSEA) on the genes
}
\examples{
if (interactive()) {
  data("pancreas1k")
  library(dplyr)
  pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType", only.pos = FALSE, fc.threshold = 1)
  de_filter <- filter(pancreas1k@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05)
  res <- RunGSEA(geneID = de_filter$gene, geneScore = de_filter$avg_log2FC, geneID_groups = de_filter$group1, species = "Mus_musculus")
  GSEAPlot(x = res$results[[1]], geneSetID = res$results[[1]]@result$ID[1])
}
}
