% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunHarmony2}
\alias{RunHarmony2}
\title{Harmony single cell integration}
\usage{
RunHarmony2(object, ...)
}
\arguments{
\item{object}{An Seurat object}

\item{...}{other parameters}
}
\description{
Run Harmony algorithm with Seurat pipelines.
}
