% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunKNNMap}
\alias{RunKNNMap}
\title{Single-cell reference mapping with KNN method}
\usage{
RunKNNMap(
  srt_query,
  srt_ref,
  ref_umap = NULL,
  force = FALSE,
  ref_group = NULL,
  features = NULL,
  nfeatures = 2000,
  query_reduction = NULL,
  ref_reduction = NULL,
  query_dims = 1:30,
  ref_dims = 1:30,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{vote_fun}{}
}
\description{
Single-cell reference mapping with KNN method
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)

data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(stringr::str_to_title(rownames(panc8_sub)))
panc8_rename <- RenameFeatures(panc8_sub, newnames = genenames)
head(rownames(panc8_rename))
panc8_rename <- Integration_SCP(panc8_rename, batch = "tech", integration_method = "Seurat")
ClassDimPlot(panc8_rename, "celltype")

# Projection
pancreas_sub <- RunKNNMap(srt_query = pancreas_sub, srt_ref = panc8_rename, ref_umap = "SeuratUMAP2D")
ProjectionPlot(srt_query = pancreas_sub, srt_ref = panc8_rename, query_group = "SubCellType", ref_group = "celltype")

}
