% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cell_annotation.R
\name{RunKNNPredict}
\alias{RunKNNPredict}
\title{Annotate single cells using similarity.}
\usage{
RunKNNPredict(
  srt_query,
  srt_ref = NULL,
  bulk_ref = NULL,
  query_group = NULL,
  ref_group = NULL,
  query_assay = NULL,
  ref_assay = NULL,
  query_reduction = NULL,
  ref_reduction = NULL,
  query_dims = ref_dims,
  ref_dims = 1:30,
  query_collapsing = !is.null(query_group),
  ref_collapsing = TRUE,
  return_full_distance_matrix = FALSE,
  features = NULL,
  features_type = c("HVF", "DE"),
  feature_source = "both",
  nfeatures = 2000,
  DEtest_param = list(max.cells.per.ident = 200, test.use = "wilcox"),
  DE_threshold = "p_val_adj < 0.05",
  nn_method = NULL,
  distance_metric = "cosine",
  k = 30,
  filter_lowfreq = 0,
  prefix = "knnpredict",
  force = FALSE
)
}
\arguments{
\item{srt_query}{A seurat object.}

\item{bulk_ref}{A cell atlas matrix, e.g., SCP::ref_scHCL}

\item{query_assay}{\code{assay} used in the srt_query}

\item{DE_threshold}{Threshold used to filter the DE features. Default is "p_val < 0.05". If using "roc" test, \code{DE_threshold} should be needs to be reassigned. e.g. "power > 0.5"}

\item{k}{Number of predictions to return.}

\item{prefix}{}

\item{force}{Whether to force to annotate when data type is different.}
}
\description{
Annotate single cells using similarity.
}
\examples{
# Annotate cells using bulk RNA-seq data
data("pancreas1k")
data("ref_scMCA")
pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunKNNPredict(srt_query = pancreas1k, bulk_ref = ref_scMCA)
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)

# Removal of low credible cell types from the predicted results
pancreas1k <- RunKNNPredict(srt_query = pancreas1k, bulk_ref = ref_scMCA, filter_lowfreq = 30)
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)

# Annotate clusters using bulk RNA-seq data
pancreas1k <- RunKNNPredict(srt_query = pancreas1k, query_group = "SubCellType", bulk_ref = ref_scMCA)
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)

# Annotate using single cell RNA-seq data
data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(stringr::str_to_title(rownames(panc8_sub)))
panc8_sub <- RenameFeatures(panc8_sub, newnames = genenames)
panc8_sub <- check_srtMerge(panc8_sub, batch = "tech")[["srtMerge"]]

pancreas1k <- RunKNNPredict(srt_query = pancreas1k, srt_ref = panc8_sub, ref_group = "celltype")
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)
ExpDimPlot(pancreas1k, features = "knnpredict_simil")

pancreas1k <- RunKNNPredict(
  srt_query = pancreas1k, srt_ref = panc8_sub,
  ref_group = "celltype", ref_collapsing = FALSE
)
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)
ExpDimPlot(pancreas1k, features = "knnpredict_prob")

pancreas1k <- RunKNNPredict(
  srt_query = pancreas1k, srt_ref = panc8_sub,
  query_group = "SubCellType", ref_group = "celltype",
  query_collapsing = TRUE, ref_collapsing = TRUE
)
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)
ExpDimPlot(pancreas1k, features = "knnpredict_simil")

# Annotate with DE gene instead of HVF
pancreas1k <- RunKNNPredict(
  srt_query = pancreas1k, srt_ref = panc8_sub,
  ref_group = "celltype",
  features_type = "DE", feature_source = "ref"
)
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)
ExpDimPlot(pancreas1k, features = "knnpredict_simil")

pancreas1k <- RunKNNPredict(
  srt_query = pancreas1k, srt_ref = panc8_sub,
  query_group = "SubCellType", ref_group = "celltype",
  features_type = "DE", feature_source = "both"
)
ClassDimPlot(pancreas1k, group.by = "knnpredict_classification", label = TRUE)
ExpDimPlot(pancreas1k, features = "knnpredict_simil")

}
