% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunMDS}
\alias{RunMDS}
\alias{RunMDS.Seurat}
\alias{RunMDS.Assay}
\alias{RunMDS.default}
\title{Run multi-dimensional scaling(MDS)}
\usage{
RunMDS(object, ...)

\method{RunMDS}{Seurat}(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nmds = 50,
  dist.method = "euclidean",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "mds",
  reduction.key = "MDS_",
  verbose = TRUE,
  seed.use = 42,
  ...
)

\method{RunMDS}{Assay}(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nmds = 50,
  dist.method = "euclidean",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "MDS_",
  verbose = TRUE,
  seed.use = 42,
  ...
)

\method{RunMDS}{default}(
  object,
  assay = NULL,
  slot = "data",
  nmds = 50,
  dist.method = "euclidean",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "MDS_",
  seed.use = 42,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{}

\item{assay}{Name of Assay mds is being run on.}

\item{slot}{Name of slot mds is being run on.}

\item{features}{Features to compute mds on. If features=NULL, mds will be run
using the variable features for the Assay.}

\item{nmds}{Total Number of MDS to compute and store (50 by default).}

\item{rev.mds}{By default computes the mds on the cell x gene matrix. Setting
to true will compute it on gene x cell matrix.}

\item{ndims.print}{MDS to print genes for.}

\item{nfeatures.print}{Number of genes to print for each MDS.}

\item{reduction.name}{dimensional reduction name, 'mds' by default}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. "MDS_" by default.}

\item{verbose}{Print the top genes associated with high/low loadings for
the MDS.}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}
}
\value{
Returns Seurat object with the MDS calculation stored in the reductions slot.
}
\description{
Run a MDS dimensionality reduction.
}
\concept{dimensional_reduction}
