% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunMonocle2}
\alias{RunMonocle2}
\title{Run Monocle2 analysis}
\usage{
RunMonocle2(
  srt,
  annotation = NULL,
  assay = NULL,
  slot = "counts",
  expressionFamily = "negbinomial.size",
  features = NULL,
  feature_type = "HVF",
  disp_filter = "mean_expression >= 0.1 & dispersion_empirical >= 1 * dispersion_fit",
  max_components = 2,
  reduction_method = "DDRTree",
  norm_method = "log",
  residualModelFormulaStr = NULL,
  pseudo_expr = 1,
  root_state = NULL,
  seed = 11
)
}
\arguments{
\item{root_state}{}
}
\description{
Run Monocle2 analysis
}
\examples{
if (interactive()) {
  data("pancreas1k")
  pancreas1k <- RunMonocle2(srt = pancreas1k, annotation = "CellType")
  names(pancreas1k@tools$Monocle2)
  trajectory <- pancreas1k@tools$Monocle2$trajectory

  p1 <- ClassDimPlot(pancreas1k, group.by = "Monocle2_State", reduction = "DDRTree", label = TRUE) + trajectory
  p2 <- ClassDimPlot(pancreas1k, group.by = "Monocle2_State", reduction = "UMAP", label = TRUE)
  p3 <- ExpDimPlot(pancreas1k, features = "Monocle2_Pseudotime", reduction = "UMAP")
  print(p1 + p2 + p3)

  pancreas1k <- RunMonocle2(
    srt = pancreas1k, annotation = "CellType",
    feature_type = "Disp", disp_filter = "mean_expression >= 0.01 & dispersion_empirical >= 1 * dispersion_fit"
  )
  trajectory <- pancreas1k@tools$Monocle2$trajectory
  p1 <- ClassDimPlot(pancreas1k, group.by = "Monocle2_State", reduction = "DDRTree", label = TRUE) + trajectory
  p2 <- ClassDimPlot(pancreas1k, group.by = "Monocle2_State", reduction = "UMAP", label = TRUE)
  p3 <- ExpDimPlot(pancreas1k, features = "Monocle2_Pseudotime", reduction = "UMAP")
  print(p1 + p2 + p3)
}
}
