% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunMonocle3}
\alias{RunMonocle3}
\title{Run Monocle3 analysis}
\usage{
RunMonocle3(
  srt,
  annotation = NULL,
  assay = NULL,
  slot = "counts",
  reduction = DefaultReduction(srt),
  clusters = NULL,
  graph = NULL,
  partition_qval = 0.05,
  k = 50,
  cluster_method = "louvain",
  num_iter = 2,
  resolution = NULL,
  use_partition = NULL,
  close_loop = TRUE,
  learn_graph_control = NULL,
  root_pr_nodes = NULL,
  root_cells = NULL,
  seed = 11
)
}
\arguments{
\item{seed}{}
}
\description{
Run Monocle3 analysis
}
\examples{
if (interactive()) {
  data("pancreas1k")
  # Use Monocle clusters to infer the trajectories
  pancreas1k <- RunMonocle3(srt = pancreas1k, annotation = "CellType")
  names(pancreas1k@tools$Monocle3)
  trajectory <- pancreas1k@tools$Monocle3$trajectory
  milestones <- pancreas1k@tools$Monocle3$milestones

  p1 <- ClassDimPlot(pancreas1k, group.by = "Monocle3_partitions", reduction = "UMAP", label = TRUE) + trajectory + milestones
  p2 <- ClassDimPlot(pancreas1k, group.by = "Monocle3_clusters", reduction = "UMAP", label = TRUE) + trajectory
  p3 <- ExpDimPlot(pancreas1k, features = "Monocle3_Pseudotime", reduction = "UMAP") + trajectory
  print(p1 + p2 + p3)

  # Select the lineage using  monocle3::choose_graph_segments
  cds <- pancreas1k@tools$Monocle3$cds
  cds_sub <- monocle3::choose_graph_segments(cds, starting_pr_node = NULL, ending_pr_nodes = NULL)
  pancreas1k$Lineages_1 <- NA
  pancreas1k$Lineages_1[colnames(cds_sub)] <- pancreas1k$Monocle3_Pseudotime[colnames(cds_sub)]
  ClassDimPlot(pancreas1k, group.by = "SubCellType", lineages = "Lineages_1", lineages_span = 0.1)

  # Use Seurat clusters to infer the trajectories
  pancreas1k <- Standard_SCP(pancreas1k)
  ClassDimPlot(pancreas1k, group.by = c("Standardclusters", "CellType"), label = TRUE)
  pancreas1k <- RunMonocle3(srt = pancreas1k, annotation = "CellType", clusters = "Standardclusters")
  trajectory <- pancreas1k@tools$Monocle3$trajectory
  p1 <- ClassDimPlot(pancreas1k, group.by = "Monocle3_partitions", reduction = "StandardUMAP2D", label = TRUE) + trajectory
  p2 <- ClassDimPlot(pancreas1k, group.by = "Monocle3_clusters", reduction = "StandardUMAP2D", label = TRUE) + trajectory
  p3 <- ExpDimPlot(pancreas1k, features = "Monocle3_Pseudotime", reduction = "StandardUMAP2D") + trajectory
  print(p1 + p2 + p3)

  # Use custom graphs and cell clusters to infer the partitions and trajectories, respectively
  pancreas1k <- Standard_SCP(pancreas1k, cluster_resolution = 5)
  ClassDimPlot(pancreas1k, group.by = c("Standardclusters", "CellType"), label = TRUE)
  pancreas1k <- RunMonocle3(
    srt = pancreas1k, annotation = "CellType",
    clusters = "Standardclusters", graph = "Standardpca_SNN"
  )
  trajectory <- pancreas1k@tools$Monocle3$trajectory
  p1 <- ClassDimPlot(pancreas1k, group.by = "Monocle3_partitions", reduction = "StandardUMAP2D", label = TRUE) + trajectory
  p2 <- ClassDimPlot(pancreas1k, group.by = "Monocle3_clusters", reduction = "StandardUMAP2D", label = TRUE) + trajectory
  p3 <- ExpDimPlot(pancreas1k, features = "Monocle3_Pseudotime", reduction = "StandardUMAP2D") + trajectory
  print(p1 + p2 + p3)
}
}
