% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunPaCMAP}
\alias{RunPaCMAP}
\alias{RunPaCMAP.Seurat}
\alias{RunPaCMAP.default}
\title{Run PaCMAP}
\usage{
RunPaCMAP(object, ...)

\method{RunPaCMAP}{Seurat}(
  object,
  reduction = "pca",
  dims = NULL,
  features = NULL,
  assay = DefaultAssay(object = object),
  slot = "data",
  n_components = 2,
  n.neighbors = NULL,
  MN_ratio = 0.5,
  FP_ratio = 2,
  pair_neighbors = NULL,
  pair_MN = NULL,
  pair_FP = NULL,
  distance_method = "euclidean",
  lr = 1,
  num_iters = 450L,
  apply_pca = TRUE,
  init = "random",
  reduction.name = "pacmap",
  reduction.key = "PaCMAP_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)

\method{RunPaCMAP}{default}(
  object,
  assay = NULL,
  n_components = 2,
  n.neighbors = NULL,
  MN_ratio = 0.5,
  FP_ratio = 2,
  pair_neighbors = NULL,
  pair_MN = NULL,
  pair_FP = NULL,
  distance_method = "euclidean",
  lr = 1,
  num_iters = 450L,
  apply_pca = TRUE,
  init = "random",
  reduction.key = "PaCMAP_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}

\item{assay}{}

\item{n_components}{}
}
\description{
Run PaCMAP
}
\concept{dimensional_reduction}
