% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunPAGA}
\alias{RunPAGA}
\title{Run PAGA analysis}
\usage{
RunPAGA(
  srt = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  adata = NULL,
  h5ad = NULL,
  group_by = NULL,
  liner_reduction = NULL,
  nonliner_reduction = NULL,
  basis = NULL,
  n_pcs = 30,
  n_neighbors = 30,
  use_rna_velocity = FALSE,
  vkey = "stochastic",
  embedded_with_PAGA = FALSE,
  paga_layout = "fr",
  threshold = 0.1,
  point_size = 20,
  infer_pseudotime = FALSE,
  root_group = NULL,
  root_cell = NULL,
  n_dcs = 10,
  n_branchings = 0,
  min_group_size = 0.01,
  show_plot = TRUE,
  dpi = 300,
  save = FALSE,
  dirpath = "./",
  fileprefix = "",
  return_seurat = FALSE
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{adata}{An \code{anndata} object. Can be created through \code{\link{srt_to_adata}}}

\item{h5ad}{h5ad file path.}

\item{group_by}{group_by.}

\item{liner_reduction}{liner_reduction.}

\item{nonliner_reduction}{nonliner_reduction.}

\item{basis}{basis.}

\item{n_pcs}{n_pcs.}

\item{n_neighbors}{n_neighbors.}

\item{use_rna_velocity}{use_rna_velocity.}

\item{vkey}{Key for velocity. Default is "stochastic".}

\item{paga_layout}{Plotting layout that computes positions. See See \href{https://scanpy.readthedocs.io/en/stable/generated/scanpy.pl.paga.html}{layout} param in scanpy.pl.paga function.}

\item{threshold}{edge threshold.}

\item{point_size}{point_size.}

\item{dpi}{dpi.}

\item{save}{save.}

\item{dirpath}{dirpath.}

\item{fileprefix}{fileprefix.}

\item{return_seurat}{}
}
\value{
A \code{anndata} object.
}
\description{
Run PAGA analysis
}
\examples{
data("pancreas1k")
pancreas1k <- RunPAGA(srt = pancreas1k, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP", return_seurat = TRUE)

pancreas1k <- RunPAGA(
  srt = pancreas1k, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP",
  embedded_with_PAGA = TRUE, infer_pseudotime = TRUE, root_group = "Ductal", return_seurat = TRUE
)
head(pancreas1k[[]])
ExpDimPlot(pancreas1k, "dpt_pseudotime")

}
