% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunPCAMap}
\alias{RunPCAMap}
\title{Single-cell reference mapping with PCA method}
\usage{
RunPCAMap(
  srt_query,
  srt_ref,
  ref_pca = NULL,
  ref_dims = 1:30,
  ref_umap = NULL,
  force = FALSE,
  ref_group = NULL,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{vote_fun}{}
}
\description{
Single-cell reference mapping with PCA method
}
\examples{
if (interactive()) {
  if (!require("SeuratData", quietly = TRUE)) {
    devtools::install_github("zhanghao-njmu/seurat-data")
  }
  library(SeuratData)
  library(stringr)
  suppressWarnings(InstallData("panc8"))
  data("panc8")

  # Simply convert genes from human to mouse and preprocess the data
  genenm <- make.unique(str_to_title(rownames(panc8)))
  panc8 <- RenameFeatures(panc8, newnames = genenm)
  head(rownames(panc8))

  panc8 <- Integration_SCP(panc8, batch = "tech", integration_method = "Seurat")
  ClassDimPlot(panc8, "celltype")

  # Projection
  pancreas1k <- Standard_SCP(pancreas1k)
  pancreas1k <- RunPCAMap(srt_query = pancreas1k, srt_ref = panc8, ref_pca = "Seuratpca", ref_umap = "SeuratUMAP2D")
  ProjectionPlot(srt_query = pancreas1k, srt_ref = panc8, query_group = "SubCellType", ref_group = "celltype")
}
}
