% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunSCVELO}
\alias{RunSCVELO}
\title{Run scVelo analysis}
\usage{
RunSCVELO(
  srt = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  adata = NULL,
  h5ad = NULL,
  group_by = NULL,
  n_jobs = 1,
  liner_reduction = NULL,
  nonliner_reduction = NULL,
  basis = NULL,
  mode = "stochastic",
  fitting_by = "stochastic",
  magic_impute = FALSE,
  knn = 5,
  t = 2,
  min_shared_counts = 30,
  n_pcs = 30,
  n_neighbors = 30,
  approx = TRUE,
  stream_smooth = NULL,
  stream_density = 2,
  arrow_length = 5,
  arrow_size = 5,
  arrow_density = 0.5,
  denoise = FALSE,
  denoise_topn = 3,
  kinetics = FALSE,
  kinetics_topn = 100,
  calculate_velocity_genes = FALSE,
  s_genes = NULL,
  g2m_genes = NULL,
  show_plot = TRUE,
  dpi = 300,
  save = FALSE,
  dirpath = "./",
  fileprefix = "",
  return_seurat = FALSE
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{adata}{An \code{anndata} object. Can be created through \code{\link{srt_to_adata}}}

\item{h5ad}{h5ad file path.}

\item{group_by}{group_by.}

\item{n_jobs}{n_jobs.}

\item{liner_reduction}{liner_reduction.}

\item{nonliner_reduction}{nonliner_reduction.}

\item{basis}{basis.}

\item{mode}{mode.}

\item{fitting_by}{fitting_by.}

\item{min_shared_counts}{min_shared_counts.}

\item{n_pcs}{n_pcs.}

\item{n_neighbors}{n_neighbors.}

\item{approx}{approx.}

\item{stream_smooth}{stream_smooth.}

\item{stream_density}{stream_density.}

\item{arrow_length}{arrow_length.}

\item{arrow_size}{arrow_size.}

\item{arrow_density}{arrow_density.}

\item{calculate_velocity_genes}{calculate_velocity_genes.}

\item{s_genes}{s_genes.}

\item{g2m_genes}{g2m_genes.}

\item{dpi}{dpi.}

\item{save}{save.}

\item{dirpath}{dirpath.}

\item{fileprefix}{fileprefix.}

\item{return_seurat}{}
}
\value{
A \code{anndata} object.
}
\description{
scVelo is a scalable toolkit for RNA velocity analysis in single cells.
This function performs scVelo workflow in R by reticulate.
}
\examples{
data("pancreas1k")
pancreas1k <- RunSCVELO(srt = pancreas1k, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP", return_seurat = TRUE)
head(pancreas1k[[]])
names(pancreas1k@assays)
ExpDimPlot(pancreas1k, c("stochastic_length", "stochastic_confidence"))
ExpDimPlot(pancreas1k, "stochastic_pseudotime")

}
