% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunSeuratMap}
\alias{RunSeuratMap}
\title{Single-cell reference mapping with Seurat method}
\usage{
RunSeuratMap(
  srt_query,
  query_assay = NULL,
  srt_ref,
  ref_pca = NULL,
  ref_dims = 1:30,
  ref_umap = NULL,
  force = FALSE,
  ref_group = NULL,
  normalization.method = "LogNormalize",
  reduction_project_method = "pcaproject",
  k.anchor = 5,
  k.filter = 200,
  k.score = 30,
  k.weight = 100,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{vote_fun}{}
}
\description{
Single-cell reference mapping with Seurat method
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)

data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(stringr::str_to_title(rownames(panc8_sub)))
panc8_rename <- RenameFeatures(panc8_sub, newnames = genenames)
head(rownames(panc8_rename))
panc8_rename <- Integration_SCP(panc8_rename, batch = "tech", integration_method = "Seurat")
ClassDimPlot(panc8_rename, "celltype")

# Projection
pancreas_sub <- RunSeuratMap(srt_query = pancreas_sub, srt_ref = panc8_rename, ref_pca = "Seuratpca", ref_umap = "SeuratUMAP2D", k.weight = 80)
ProjectionPlot(srt_query = pancreas_sub, srt_ref = panc8_rename, query_group = "SubCellType", ref_group = "celltype")

}
