% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunTriMap}
\alias{RunTriMap}
\alias{RunTriMap.Seurat}
\alias{RunTriMap.default}
\title{Run TriMap}
\usage{
RunTriMap(object, ...)

\method{RunTriMap}{Seurat}(
  object,
  reduction = "pca",
  dims = NULL,
  features = NULL,
  distance = NULL,
  assay = DefaultAssay(object = object),
  slot = "data",
  n_components = 2,
  n_inliers = 12,
  n_outliers = 4,
  n_random = 3,
  distance_method = "euclidean",
  lr = 0.1,
  n_iters = 400,
  triplets = NULL,
  weights = NULL,
  use_dist_matrix = FALSE,
  knn_tuple = NULL,
  apply_pca = TRUE,
  opt_method = "dbd",
  reduction.name = "trimap",
  reduction.key = "TriMap_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)

\method{RunTriMap}{default}(
  object,
  assay = NULL,
  n_components = 2,
  n_inliers = 12,
  n_outliers = 4,
  n_random = 3,
  distance_method = "euclidean",
  lr = 0.1,
  n_iters = 400,
  triplets = NULL,
  weights = NULL,
  use_dist_matrix = FALSE,
  knn_tuple = NULL,
  apply_pca = TRUE,
  opt_method = "dbd",
  reduction.key = "TriMap_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}

\item{n_components}{}
}
\description{
Run TriMap
}
\concept{dimensional_reduction}
