% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunWOT}
\alias{RunWOT}
\title{Run WOT analysis}
\usage{
RunWOT(
  srt = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  adata = NULL,
  group_by = NULL,
  time_field = "Time",
  growth_iters = 3L,
  tmap_out = "tmaps/tmap_out",
  time_from = 1,
  time_to = NULL,
  get_coupling = FALSE,
  recalculate = FALSE,
  palette = "Paired",
  palcolor = NULL,
  show_plot = TRUE,
  dpi = 300,
  save = FALSE,
  dirpath = "./",
  fileprefix = "",
  return_seurat = !is.null(srt)
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{adata}{An \code{anndata} object. Can be created through \code{\link{srt_to_adata}}}

\item{group_by}{group_by.}

\item{dpi}{dpi.}

\item{save}{save.}

\item{dirpath}{dirpath.}

\item{fileprefix}{fileprefix.}

\item{return_seurat}{}
}
\value{
A \code{anndata} object.
}
\description{
Run WOT analysis
}
