% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data to demonstrate function usage in \code{scp}}
\format{
A \code{list} with 2 elements, which are:
\describe{
\item{s}{a \code{data.frame} with two coordinates of 1681 grid locations in \eqn{[0,1]^2}.}
\item{Y}{a vector with \eqn{Y(s)} observations corresponding to \code{s}, where \eqn{Y(s) = X(s)^3 + E(s)},
\eqn{X(s)} is a stationary Gaussian process process with a Matern covariance, and \eqn{E(s)} is a white noise process.}
}
}
\usage{
sample_data
}
\description{
A list containing locations, \code{s},
and corresponding observations, \code{Y}, to demostrating \code{\link{scp}},
\code{\link{plausibility}}, and \code{\link{plausibility_contour}} usage
}
\keyword{internal}
