% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp-defunct.R
\name{scp-defunct}
\alias{scp-defunct}
\alias{aggregateFeaturesOverAssays}
\title{Defunct Functions}
\description{
Defunct Functions
}
\details{
The \code{aggregateFeaturesOverAssays} function is defunct. Please use the
\code{QFeatures::aggregateFeatures} method from the \code{QFeatures} package instead.
}
