% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp}
\alias{scp}
\title{Spatial Conformal Prediction (SCP) At a Single Input Location}
\usage{
scp(
  s0,
  s,
  Y,
  global = TRUE,
  eta = Inf,
  m = NULL,
  pred_fun = krige_pred,
  theta = NULL,
  alpha = 0.05,
  dfun = "std_residual2",
  precision = 0.01
)
}
\arguments{
\item{s0}{prediction location}

\item{s}{an \eqn{n \times d}{n x d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{global}{logical; if TRUE , \code{scp} function returns the result of global spatial conformal prediction \code{\link{gscp}};
if FALSE, \code{scp} function returns the result of local spatial conformal prediction \code{\link{lscp}}.}

\item{eta}{numerical value of the kernel bandwidth for the weight schema in conformal prediction. Defauls to \eqn{Inf} meaning equal weight on surrounding \eqn{m} points.}

\item{m}{an postive integer representing the number of nearest locations used for prediction. Depends on eta.}

\item{alpha}{significance level. Defaults to 0.05.}

\item{dfun}{non-conformity measure}

\item{precision}{Defaults to 0.01.}
}
\value{

}
\description{
This function provides the conformal prediction interval for spatial location \code{s0} given obserations \code{s, Y}.
}
