% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scp_aggregateByProtein}
\alias{scp_aggregateByProtein}
\title{Aggregation of peptide expression data into protein expression data}
\usage{
scp_aggregateByProtein(obj)
}
\arguments{
\item{obj}{An \code{MSnSet} object containing the peptide expression data.}
}
\value{
A new \code{MSnSet} object containing the aggregated protein expression data.
Note that the feature data is adapted and will contain only protein
annotations.
}
\description{
The \code{scp_AggregateByProtein} function takes an \code{MSnSet} object and
merges the peptides (rows) of the expression data into proteins. This merging
is performed by taking the median value for each cell (column) for a given
protein group. The peptide to protein mapping should be present in the
feature data of the \code{\link{MSnSet}} object.
}
\examples{
data(specht2019)
scProt <- scp_aggregateByProtein(specht2019)

}
