% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scp_exprsToLong}
\alias{scp_exprsToLong}
\title{Transform channel expression data to long format}
\usage{
scp_exprsToLong(obj, f_sel = c("Raw.file", "sequence_charge"))
}
\arguments{
\item{obj}{An object of class \code{\link{MSnSet}} containing channel expression data.}

\item{f_sel}{A vector of columns names that must be kept as feature data.}
}
\value{
An object of class \code{\link{MSnSet}} containing all expression data in a
single column.
}
\description{
The function converts the TMT channel data to a single-column expression
data. The channel header is used to create the \code{channel} field in the
feature data.
}
\examples{
print("hello")

}
