% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scp_normalize}
\alias{scp_normalize}
\title{Normalization of single-cell proteomics data}
\usage{
scp_normalize(obj, what = "col", method = 1, fun = "-")
}
\arguments{
\item{obj}{An MSnSet object}

\item{what}{A character indicating whether columns (\code{"col"}), rows (\code{"row"}) or
both (\code{"both"}) should be normalized}

\item{method}{The type of normalization (see Details)}
}
\value{
An \code{MSnSet} object similar to the input object \code{obj}, but where the
expression values have been replaced by normalized expression values.
}
\description{
The function \code{scp_normalise} (and identically \code{scp_normalize})
performs normalization of cells or peptides/proteins or both of quantified
expression data (\code{MSnSet} objects). Different normalization methods are
implemented (see Details).
}
\details{
\code{method == 1}: each row (if \code{what == "row"}) or column (if
\code{what == "col"}) is subtracted by summary of the corresponding values.
The summary function is the mean for rows and the median for columns.
}
\examples{
sc <- scpdata("specht2019", type = "peptide")
scpNorm <- scp_normalize(sc, what = "both")

}
