% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{scVI_integrate}
\alias{scVI_integrate}
\title{scVI_integrate}
\usage{
scVI_integrate(
  srtMerge = NULL,
  batch = "orig.ident",
  append = TRUE,
  srtList = NULL,
  do_normalization = NULL,
  normalization_method = "logCPM",
  do_HVF_finding = TRUE,
  HVF_source = "separate",
  HVF_method = "vst",
  nHVF = 2000,
  HVF = NULL,
  nonliner_reduction = "umap",
  nonliner_reduction_dims = c(2, 3),
  nonliner_reduction_distance = "cosine",
  force_nonliner_reduction = TRUE,
  do_cluster_finding = TRUE,
  cluster_algorithm = "louvain",
  cluster_resolution = 0.6,
  cluster_reorder = TRUE,
  SCVI_params = list(),
  seed = 11
)
}
\arguments{
\item{seed}{}
}
\description{
scVI_integrate
}
