% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_eta.R
\name{select_eta}
\alias{select_eta}
\title{Select the optimal tunning parameter for \code{\link{scp}}}
\usage{
select_eta(s, Y, eta_cand = NULL, plot = TRUE)
}
\arguments{
\item{s}{an \eqn{n \times 2}{n x 2} \code{matrix} or a \code{data.frame} with two coordinates of \eqn{n} locations.}

\item{Y}{a vector with \eqn{n} values corresponding to \code{Y(s)}.}

\item{eta_cand}{a vector of candidate \code{eta} values. Defaults to \code{NULL}.}

\item{plot}{logical; if \code{TRUE}, \code{select_eta} plots the empirical interval score for the candidate \code{eta} values.
if \code{FALSE}, no plot will be generated. Defaults to \code{TRUE}.}
}
\value{
The optimum \code{eta} based on minimizing empirical interval score.
}
\description{
Select the optimal tunning parameter for \code{\link{scp}}
}
\examples{
#?sample_data
s  = sample_data$s
Y  = sample_data$Y

# CAUTION: the following command may take one hour to run.
opt_eta = select_eta(s = s, Y = Y)

use optimal eta to calculate prediction interval
s0  = c(0.5,0.5)
scp(s0 = s0, s = s, Y = Y, eta = opt_eta)
}
