% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{SeqModelSel}
\alias{SeqModelSel}
\title{Sequential Model Selection for Joint Rhythmicity}
\usage{
SeqModelSel(
  action = c(1, 2),
  pv = c(0.01, 0.02),
  alpha = 0.05,
  method = "Fisher_FS"
)
}
\arguments{
\item{action}{Vector indicating which group to test first}

\item{pv}{P-values for the tests}

\item{alpha}{Significance threshold}

\item{method}{Selection method: "Fisher_BS", "Fisher_FS", "Sidak_BS", "Sidak_FS",
"Nominal_BS", "Nominal_FS", "VDA", "AWFisher"}
}
\value{
Joint rhythmicity category
}
\description{
Sequential Model Selection for Joint Rhythmicity
}
