% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{srt_to_adata}
\alias{srt_to_adata}
\title{Convert a seurat object to an anndata object using reticulate}
\usage{
srt_to_adata(
  srt,
  features = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  convert_tools = FALSE,
  convert_misc = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{assay_X}{Assays to convert as X(main data matrix) in anndata object.}

\item{assay_layers}{Assays to convert as layers in anndata object.}

\item{verbose}{}
}
\value{
A \code{anndata} object.
}
\description{
Convert a seurat object to an anndata object using reticulate
}
\examples{
data("pancreas_sub")
adata <- srt_to_adata(pancreas_sub)
adata

### Or save as an h5ad file or a loom file
# adata$write_h5ad("pancreas_sub.h5ad")
# adata$write_loom("pancreas_sub.loom", write_obsm_varm = TRUE)

}
