% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{SrtAppend}
\alias{SrtAppend}
\title{Append a Seurat object to another}
\usage{
SrtAppend(
  srt_raw,
  srt_append,
  slots = slotNames(srt_append),
  pattern = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{srt_raw}{}

\item{verbose}{}
}
\description{
Append a Seurat object to another
}
