% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{SrtReorder}
\alias{SrtReorder}
\title{Reorder idents by the gene expression}
\usage{
SrtReorder(
  srt,
  features = NULL,
  reorder_by = NULL,
  slot = "data",
  assay = NULL,
  log = TRUE,
  distance_metric = "euclidean"
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{features}{Features used to reorder idents.}

\item{reorder_by}{Reorder groups instead of idents.}

\item{slot}{Specific slot to get data from.}

\item{assay}{Specific assay to get data from.}

\item{log}{Whether log1p transformation needs to be applied. Default is \code{TRUE}.}

\item{distance_metric}{Metric to compute distance. Default is "euclidean".}
}
\description{
Reorder idents by the gene expression
}
