% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{StatPlot}
\alias{StatPlot}
\title{StatPlot}
\usage{
StatPlot(
  meta.data,
  stat.by,
  group.by = NULL,
  split.by = NULL,
  flip = FALSE,
  NA_color = "grey",
  NA_stat = TRUE,
  keep_empty = FALSE,
  stat_single = FALSE,
  stat_level = NULL,
  plot_type = c("bar", "rose", "ring", "pie", "trend", "area", "dot", "sankey", "chord",
    "venn", "upset"),
  stat_type = c("percent", "count"),
  position = c("stack", "dodge"),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  label = FALSE,
  label.size = 3.5,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  theme_use = "theme_scp",
  aspect.ratio = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  force = FALSE,
  seed = 11
)
}
\arguments{
\item{meta.data}{}

\item{force}{}
}
\description{
StatPlot
}
\examples{
data("pancreas_sub")
head(pancreas_sub@meta.data)
StatPlot(pancreas_sub@meta.data, stat.by = "Phase", group.by = "CellType", plot_type = "bar", label = TRUE)

head(pancreas_sub[["RNA"]]@meta.features)
StatPlot(pancreas_sub[["RNA"]]@meta.features, stat.by = "highly_variable_genes", plot_type = "ring", label = TRUE)

pancreas_sub <- AnnotateFeatures(pancreas_sub, species = "Mus_musculus", IDtype = "symbol", db = "GeneType")
head(pancreas_sub[["RNA"]]@meta.features)
StatPlot(pancreas_sub[["RNA"]]@meta.features,
  stat.by = "highly_variable_genes", group.by = "GeneType",
  stat_type = "count", plot_type = "bar", position = "dodge", label = TRUE, NA_stat = FALSE
)

}
