% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{StopRule}
\alias{StopRule}
\title{Stop Rule for Sequential Testing}
\usage{
StopRule(action, pv, alpha, method)
}
\arguments{
\item{action}{Vector indicating which group to test first}

\item{pv}{P-values}

\item{alpha}{Significance threshold}

\item{method}{"FS" (forward stop) or "BS" (basic stop)}
}
\value{
List with type and one
}
\description{
Stop Rule for Sequential Testing
}
