% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{SummaryPlot}
\alias{SummaryPlot}
\title{SummaryPlot}
\usage{
SummaryPlot(
  srt,
  group.by = NULL,
  group.split.by = NULL,
  features = NULL,
  reduction = NULL,
  Class_palette = "Paired",
  Exp_palette = "Spectral",
  size = 2,
  margin = 0.1,
  units = "in"
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{group.by}{Name of one or more metadata columns to group (color) cells by (for example, orig.ident).}

\item{features}{Name of one or more metadata columns to group (color) cells by (for example, orig.ident).}

\item{reduction}{Which dimensionality reduction to use.}

\item{Class_palette}{Name of palette to use for ClassDimPlot.}

\item{Exp_palette}{Name of palette to use for ExpDimPlot.}

\item{size}{Size of each plot.}

\item{margin}{Margins around the plot when \code{file} is specified.}

\item{units}{The units in which \code{height}, \code{width} and \code{margin} are given. Can be \code{mm}, \code{cm}, \code{inches}, etc. See \code{\link[grid]{unit}}.}
}
\value{
A grob obejct with size attributes.
}
\description{
SummaryPlot
}
\examples{
data("pancreas1k")
p <- SummaryPlot(pancreas1k,
  group.by = c("CellType", "SubCellType"), group.split.by = "SubCellType",
  features = c("S_score", "G2M_score", "nFeature_RNA", "nCount_RNA")
)
p
\donttest{
# Save the plot with appropriate size
ggsave(
  filename = "summaryplot.png", plot = p,
  units = attr(p, "size")$units, width = attr(p, "size")$width, height = attr(p, "size")$height
)
}
}
