% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{theme_blank}
\alias{theme_blank}
\title{Blank theme}
\usage{
theme_blank(
  add_coord = TRUE,
  xlab = "Dim1",
  ylab = "Dim2",
  xlen_npc = 0.15,
  ylen_npc = 0.15,
  lab_cex = 1,
  ...
)
}
\arguments{
\item{add_coord}{}

\item{...}{}
}
\description{
Blank theme
}
