% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsankey.R
\name{theme_sankey}
\alias{theme_sankey}
\alias{theme_alluvial}
\alias{theme_sankey_bump}
\title{sankey_themes}
\usage{
theme_sankey(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_alluvial(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_sankey_bump(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Minimal themes for sankey, alluvial and sankey bump plots
}
