% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{theme_scp}
\alias{theme_scp}
\title{SCP theme}
\usage{
theme_scp(aspect.ratio = 1, text_size_scale = 1, ...)
}
\arguments{
\item{aspect.ratio}{Aspect ratio of the panel.}

\item{...}{Arguments passed to the \code{\link[ggplot2]{theme}}.}
}
\description{
The default theme for SCP plot function.
}
