% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tochunks}
\alias{tochunks}
\title{Split a vector into the chunks}
\usage{
tochunks(x, nchunks)
}
\arguments{
\item{x}{A vector.}

\item{nchunks}{Number of chunks.}
}
\description{
Split a vector into the chunks
}
\examples{
x <- 1:10
names(x) <- letters[1:10]
tochunks(x, nchunks = 3)
}
