% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_funs.R
\name{trans_dfun}
\alias{trans_dfun}
\title{Transform non-conformity measure}
\usage{
trans_dfun(dfun)
}
\arguments{
\item{dfun}{}
}
\value{

}
\description{
Transform non-conformity measure
}
\keyword{internal}
