% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transferColDataToAssay}
\alias{transferColDataToAssay}
\title{Transfer the `colData` to an Assay}
\usage{
transferColDataToAssay(obj, i)
}
\arguments{
\item{obj}{A `QFeatures` object}

\item{i}{A `numeric(1)` or `character(1)` indicating which assay to transfer 
the `colData` to.}
}
\value{
A `QFeatures` object
}
\description{
The function transfers the `colData` from a `QFeatures` object to one of the 
assays it contains. The transfered data is bound to the existing `colData` 
of the target assay.
}
\examples{
TODO
}
