% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR.R
\name{trend.spatial}
\alias{trend.spatial}
\title{Trend Spatial}
\usage{
trend.spatial(trend, geodata, add.to.trend)
}
\arguments{
\item{add.to.trend}{}
}
\value{

}
\description{
Trend Spatial
}
\keyword{internal}
