% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateAssay}
\alias{updateAssay}
\title{Update the quantitative data of an assay}
\usage{
updateAssay(obj, i, fun, ...)
}
\arguments{
\item{obj}{A `Features` object}

\item{i}{A `numeric(1)` or `character(1)` indicating which assay to transfer 
the `colData` to.}

\item{fun}{A custom function that takes as input the data matrix from assay 
`i` and return a matrix with same dimension and dimension names.}

\item{...}{Further arguments passed to `fun`}
}
\value{
A `Features` object
}
\description{
The function update the quantitative data of an assay in a `Features` object
given a custom transformation function.
}
\examples{
TODO
}
