% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{updateBioOptions}
\alias{updateBioOptions}
\title{Update CircadianBioOptions (like modifyList but preserves class)}
\usage{
updateBioOptions(opts, ...)
}
\arguments{
\item{opts}{Existing CircadianBioOptions}

\item{...}{Named arguments to override}
}
\value{
Updated CircadianBioOptions
}
\description{
Update CircadianBioOptions (like modifyList but preserves class)
}
