% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR.R
\name{variog}
\alias{variog}
\title{variog}
\usage{
variog(
  geodata,
  coords = geodata$coords,
  data = geodata$data,
  uvec = "default",
  breaks = "default",
  trend = "cte",
  lambda = 1,
  option = c("bin", "cloud", "smooth"),
  estimator.type = c("classical", "modulus"),
  nugget.tolerance,
  max.dist,
  pairs.min = 2,
  bin.cloud = FALSE,
  direction = "omnidirectional",
  tolerance = pi/8,
  unit.angle = c("radians", "degrees"),
  angles = FALSE,
  messages,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
variog
}
\keyword{internal}
