% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{VolcanoPlot}
\usage{
VolcanoPlot(
  srt,
  group_by = NULL,
  test.use = "wilcox",
  DE_threshold = "avg_log2FC > 0 & p_val_adj < 0.05",
  x_metric = "diff_pct",
  palette = "RdBu",
  palcolor = NULL,
  nlabel = 5,
  features_label = NULL,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  label.size = 4,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  align = "hv",
  axis = "lr"
)
}
\description{
VolcanoPlot
}
\examples{
data("pancreas1k")
pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType", only.pos = FALSE)
VolcanoPlot(pancreas1k, group_by = "CellType")
VolcanoPlot(pancreas1k, group_by = "CellType", x_metric = "avg_log2FC")

}
