% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{adata_to_srt}
\alias{adata_to_srt}
\title{Convert an anndata object to a seurat object using reticulate}
\usage{
adata_to_srt(adata)
}
\arguments{
\item{adata}{a connected python anndata object.}
}
\description{
Convert an anndata object to a seurat object using reticulate
}
\examples{
data("pancreas1k")
adata <- srt_to_adata(pancreas1k)
adata <- RunPAGA(adata = adata, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP")
srt <- adata_to_srt(adata)
srt

### Or convert a h5ad file to Seurat object
# library(reticulate)
# sc <- import("scanpy")
# adata <- sc$read_h5ad("pancreas.h5ad")
# srt <- adata_to_srt(adata)
# srt
}
