% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{adjcolors}
\alias{adjcolors}
\title{Convert a color with arbitrary transparency to a fixed color}
\usage{
adjcolors(colors, alpha)
}
\arguments{
\item{colors}{Color vectors.}

\item{alpha}{Alpha level in [0,1]}
}
\description{
Convert a color with arbitrary transparency to a fixed color
}
\examples{
colors <- c("red", "blue", "green")
adjcolors(colors, 0.5)

library(scales)
alpha(colors, 0.5)

show_palettes(list(colors, adjcolors(colors, 0.5), alpha(colors, 0.5)))

}
