% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_matrix}
\alias{as_matrix}
\title{Attempts to turn a dgCMatrix into a dense matrix}
\usage{
as_matrix(x)
}
\arguments{
\item{x}{A matrix.}
}
\description{
Attempts to turn a dgCMatrix into a dense matrix
}
\examples{
data("pancreas_sub")
system.time(mat1 <- as.matrix(slot(pancreas_sub[["RNA"]], "counts")))
system.time(mat2 <- as_matrix(slot(pancreas_sub[["RNA"]], "counts")))
identical(mat1, mat2)

}
