% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{calculate_CI_A.phase.Taylor}
\alias{calculate_CI_A.phase.Taylor}
\title{Calculate CI for Amplitude and Phase (Taylor method)}
\usage{
calculate_CI_A.phase.Taylor(XX.inv, A.t, phase.hat, A.hat, sigma.hat)
}
\description{
Calculate CI for Amplitude and Phase (Taylor method)
}
