% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{CellDimPlot3D}
\alias{CellDimPlot3D}
\title{3D-Dimensional reduction plot for cell classification visualization.}
\usage{
CellDimPlot3D(
  srt,
  group.by,
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  palette = "Paired",
  palcolor = NULL,
  bg_color = "grey80",
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  lineages = NULL,
  lineage_palette = "Dark2",
  span = 0.75,
  arrow_reverse = FALSE,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{group.by}{Name of one or more meta.data columns to group (color) cells by (for example, orig.ident).}

\item{reduction}{Which dimensionality reduction to use. If not specified, will use the reduction returned by \code{\link{DefaultReduction}}.}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{palette}{Name of a color palette name collected in SCP. Default is "Paired".}

\item{palcolor}{Custom colors used to create a color palette.}

\item{bg_color}{Color value for background(NA) points.}

\item{pt.size}{Point size.}

\item{cells.highlight}{A vector of cell names to highlight.}

\item{cols.highlight}{Color used to highlight the cells.}

\item{shape.highlight}{Shape of the cell to highlight. See \href{https://plotly.com/r/reference/scattergl/#scattergl-marker-symbol}{scattergl-marker-symbol}}

\item{sizes.highlight}{Size of highlighted cell points.}

\item{lineages}{Lineages/pseudotime to add to the plot. If specified, curves will be fitted using \code{\link[stats]{loess}} method.}

\item{width}{Width in pixels, defaults to automatic sizing.}

\item{height}{Height in pixels, defaults to automatic sizing.}

\item{force}{Whether to force drawing regardless of maximum levels in any cell group is greater than 100.}
}
\description{
Plotting cell points on a reduced 3D space and coloring according to the groups of the cells.
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
CellDimPlot3D(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")

pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")
CellDimPlot3D(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D", lineages = "Lineage1")
}
