% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{CellScoring}
\alias{CellScoring}
\title{Cell classification}
\usage{
CellScoring(
  srt,
  features,
  ncores = 1,
  method = "Seurat",
  classification = TRUE,
  name = "CC",
  slot = "data",
  assay = "RNA",
  new_assay = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Cell classification
}
\examples{
data("pancreas_sub")
ccgenes <- CC_GenePrefetch("Mus_musculus")
pancreas_sub <- CellScoring(srt = pancreas_sub, features = list(S = ccgenes$cc_S_genes, G2M = ccgenes$cc_G2M_genes), nbin = 10)
ClassDimPlot(pancreas_sub, "CC_classification")
ExpDimPlot(pancreas_sub, "CC_G2M")
}
