% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{CellScoring}
\alias{CellScoring}
\title{Cell classification}
\usage{
CellScoring(
  srt,
  features,
  ncores = 1,
  method = "Seurat",
  classification = TRUE,
  name = "CC",
  slot = "data",
  assay = "RNA",
  new_assay = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Cell classification
}
\examples{
data("pancreas1k")
ccgenes <- CC_GenePrefetch("Mus_musculus")
pancreas1k <- CellScoring(srt = pancreas1k, features = list(S = ccgenes$cc_S_genes, G2M = ccgenes$cc_G2M_genes), nbin = 20)
ClassDimPlot(pancreas1k, "CC_classification")
ExpDimPlot(pancreas1k, "CC_G2M")
}
