% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{CellStatPlot}
\alias{CellStatPlot}
\title{Statistical plot of cells}
\usage{
CellStatPlot(
  srt,
  stat.by,
  group.by = NULL,
  split.by = NULL,
  bg.by = NULL,
  cells = NULL,
  flip = FALSE,
  NA_color = "grey",
  NA_stat = TRUE,
  keep_empty = FALSE,
  stat_single = FALSE,
  stat_level = NULL,
  plot_type = c("bar", "rose", "ring", "pie", "trend", "area", "dot", "sankey", "chord",
    "venn", "upset"),
  stat_type = c("percent", "count"),
  position = c("stack", "dodge"),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  bg_palette = "Paired",
  bg_palcolor = NULL,
  bg_apha = 0.2,
  label = FALSE,
  label.size = 3.5,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  aspect.ratio = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  theme_use = "theme_scp",
  theme_args = list(),
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  force = FALSE,
  seed = 11
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{stat.by}{The name of a metadata column to be counted.}

\item{seed}{}
}
\description{
Statistical plot of cells
}
\examples{
library(dplyr)
data("pancreas_sub")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "SubCellType", label = TRUE)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "SubCellType", label = TRUE) \%>\% panel_fix(height = 2, width = 3)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "SubCellType", stat_type = "count", position = "dodge", label = TRUE)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "SubCellType", bg.by = "CellType", palette = "Set1", stat_type = "count", position = "dodge")

CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "bar")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "rose")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "ring")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "pie")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "dot")

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "bar")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "rose")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "ring")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "area")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "dot")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "trend")

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "bar", stat_single = TRUE)

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "bar")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "rose")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "ring")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "area")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "dot")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "trend")

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "bar", position = "dodge", label = TRUE)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "rose", position = "dodge", label = TRUE)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "ring", position = "dodge", label = TRUE)

CellStatPlot(pancreas_sub, stat.by = c("CellType", "Phase"), plot_type = "sankey")
CellStatPlot(pancreas_sub, stat.by = c("CellType", "Phase"), plot_type = "chord")

CellStatPlot(pancreas_sub,
  stat.by = c("CellType", "Phase"), plot_type = "venn",
  stat_level = list(CellType = c("Ductal", "Ngn3 low EP"), Phase = "S")
)
pancreas_sub$Progenitor <- pancreas_sub$CellType \%in\% c("Ngn3 low EP", "Ngn3 high EP")
pancreas_sub$G2M <- pancreas_sub$Phase == "G2M"
pancreas_sub$Sox9_Expressed <- pancreas_sub[["RNA"]]@counts["Sox9", ] > 0
pancreas_sub$Neurog3_Expressed <- pancreas_sub[["RNA"]]@counts["Neurog3", ] > 0
CellStatPlot(pancreas_sub, stat.by = c("Progenitor", "G2M", "Sox9_Expressed", "Neurog3_Expressed"), plot_type = "venn", stat_level = "TRUE")
CellStatPlot(pancreas_sub, stat.by = c("Progenitor", "G2M", "Sox9_Expressed", "Neurog3_Expressed"), plot_type = "upset", stat_level = "TRUE")
sum(pancreas_sub$Progenitor == "FALSE" &
  pancreas_sub$G2M == "FALSE" &
  pancreas_sub$Sox9_Expressed == "TRUE" &
  pancreas_sub$Neurog3_Expressed == "FALSE")
}
