% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{CellStatPlot}
\alias{CellStatPlot}
\title{Statistical plot of cell classification}
\usage{
CellStatPlot(
  srt,
  stat.by,
  group.by = NULL,
  split.by = NULL,
  cells = NULL,
  flip = FALSE,
  NA_color = "grey",
  NA_stat = TRUE,
  keep_empty = FALSE,
  stat_single = FALSE,
  stat_level = NULL,
  plot_type = c("bar", "rose", "ring", "pie", "trend", "area", "dot", "sankey", "chord",
    "venn", "upset"),
  stat_type = c("percent", "count"),
  position = c("stack", "dodge"),
  palette = "Paired",
  palcolor = NULL,
  alpha = 1,
  label = FALSE,
  label.size = 3.5,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  theme_use = "theme_scp",
  aspect.ratio = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  force = FALSE,
  seed = 11
)
}
\arguments{
\item{stat_level}{}
}
\description{
Statistical plot of cell classification
}
\examples{
data("pancreas_sub")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", label = TRUE)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", position = "dodge", label = TRUE)

CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "bar")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "rose")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "ring")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "pie")
CellStatPlot(pancreas_sub, stat.by = "Phase", plot_type = "dot")

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "bar")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "rose")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "ring")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "area")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "dot")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "trend")

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", plot_type = "bar", stat_single = TRUE)

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "bar")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "rose")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "ring")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "area")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "dot")
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "trend")

CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "bar", position = "dodge", label = TRUE)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "rose", position = "dodge", label = TRUE)
CellStatPlot(pancreas_sub, stat.by = "Phase", group.by = "CellType", stat_type = "count", plot_type = "ring", position = "dodge", label = TRUE)

CellStatPlot(pancreas_sub, stat.by = c("CellType", "Phase"), plot_type = "sankey")
CellStatPlot(pancreas_sub, stat.by = c("CellType", "Phase"), plot_type = "chord")

CellStatPlot(pancreas_sub,
  stat.by = c("CellType", "Phase"), plot_type = "venn",
  stat_level = list(CellType = c("Ductal", "Ngn3 low EP"), Phase = "S")
)

pancreas_sub$Ductal <- ifelse(pancreas_sub$CellType \%in\% c("Ductal", "Ngn3 low EP", "Ngn3 high EP"), TRUE, FALSE)
pancreas_sub$Endocrine <- ifelse(pancreas_sub$CellType \%in\% c("Pre-endocrine", "Endocrine"), TRUE, FALSE)
pancreas_sub$G1S <- ifelse(pancreas_sub$Phase \%in\% c("G1", "S"), TRUE, FALSE)
pancreas_sub$G2M <- ifelse(pancreas_sub$Phase \%in\% c("G2M"), TRUE, FALSE)
head(pancreas_sub@meta.data[, c("Ductal", "Endocrine", "G1S", "G2M")])
CellStatPlot(pancreas_sub, stat.by = c("Ductal", "Endocrine", "G1S", "G2M"), plot_type = "venn", stat_level = TRUE)
CellStatPlot(pancreas_sub, stat.by = c("Ductal", "Endocrine", "G1S", "G2M"), plot_type = "upset", stat_level = TRUE)

}
