% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{check_DataType}
\alias{check_DataType}
\title{Check and report the type of data}
\usage{
check_DataType(srt, data = NULL, slot = "data", assay = NULL)
}
\arguments{
\item{srt}{}

\item{assay}{}
}
\description{
Check and report the type of data
}
