% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_Python}
\alias{check_Python}
\title{Check and install python packages}
\usage{
check_Python(
  packages,
  envname = NULL,
  conda = "auto",
  force = FALSE,
  pip = TRUE,
  ...
)
}
\arguments{
\item{envname}{The name of, or path to, a Python environment.}

\item{force}{Whether to force package installation. Default is FALSE.}

\item{...}{}
}
\description{
Check and install python packages
}
