% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_Python}
\alias{check_Python}
\title{Check and install python packages}
\usage{
check_Python(pkgs, pkg_names = NULL, envname = "SCP", force = FALSE)
}
\arguments{
\item{force}{}
}
\description{
Check and install python packages
}
