% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_R}
\alias{check_R}
\title{Check and install R packages}
\usage{
check_R(
  pkgs,
  pkg_names = NULL,
  install_methods = c("BiocManager::install", "install.packages",
    "devtools::install_github"),
  lib = .libPaths()[1],
  force = FALSE
)
}
\arguments{
\item{pkgs}{Package to be installed. Package source can be CRAN, Bioconductor or Github, e.g. scmap, davidsjoberg/ggsankey.}

\item{pkg_names}{The name of the package that corresponds to the \code{pkgs} parameter, used to check if the package is already installed.
By default, the package name is extracted according to the \code{pkgs} parameter.}

\item{install_methods}{Functions for installing R packages. The default is to try to install packages from CRAN, Bioconductor and Github.}

\item{lib}{Character vector giving the library directories where to install the packages.}

\item{force}{Whether to force package installation. Default is FALSE.}
}
\description{
Check and install R packages
}
