% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_R}
\alias{check_R}
\title{Check and install R packages}
\usage{
check_R(
  pkgs,
  pkg_names = NULL,
  install_methods = c("BiocManager::install", "install.packages",
    "devtools::install_github"),
  lib = .libPaths()[1],
  force = FALSE
)
}
\arguments{
\item{pkgs}{}

\item{force}{}
}
\description{
Check and install R packages
}
