% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{check_srtMerge}
\alias{check_srtMerge}
\title{Check and preprocess a merged seurat object}
\usage{
check_srtMerge(
  srtMerge,
  batch = NULL,
  assay = NULL,
  do_normalization = NULL,
  normalization_method = "LogNormalize",
  do_HVF_finding = TRUE,
  HVF_source = "separate",
  HVF_method = "vst",
  nHVF = 2000,
  HVF_min_intersection = 1,
  HVF = NULL,
  vars_to_regress = NULL,
  seed = 11
)
}
\arguments{
\item{srtMerge}{A merged Seurat object that includes the batch information.}

\item{batch}{A character string specifying the batch variable name.}

\item{assay}{The name of the assay to be used for downstream analysis.}

\item{do_normalization}{A logical value indicating whether data normalization should be performed.}

\item{normalization_method}{The normalization method to be used. Possible values are "LogNormalize", "SCT", and "TFIDF". Default is "LogNormalize".}

\item{do_HVF_finding}{A logical value indicating whether highly variable feature (HVF) finding should be performed. Default is TRUE.}

\item{HVF_source}{The source of highly variable features. Possible values are "global" and "separate". Default is "separate".}

\item{HVF_method}{The method for selecting highly variable features. Default is "vst".}

\item{nHVF}{The number of highly variable features to select. Default is 2000.}

\item{HVF_min_intersection}{The feature needs to be present in batches for a minimum number of times in order to be considered as highly variable. The default value is 1.}

\item{HVF}{A vector of highly variable features. Default is NULL.}

\item{vars_to_regress}{A vector of variable names to include as additional regression variables. Default is NULL.}

\item{seed}{An integer specifying the random seed for reproducibility. Default is 11.}
}
\description{
This function checks and preprocesses a merged seurat object.
}
\seealso{
\code{\link{check_srtList}}
}
