% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeFDR}
\alias{computeFDR}
\title{Compute false discovery rates (FDRs) from posterior error probabilities (PEPs)}
\usage{
computeFDR(object, i, groupCol, pepCol)
}
\arguments{
\item{object}{A `QFeatures` object}

\item{i}{A `numeric()` or `character()` vector indicating from which assays 
the `rowData` should be taken.}

\item{groupCol}{A `character(1)` indicating the variable names in the 
`rowData` that contains the grouping variable. The FDR are usually 
computed for PSMs grouped by peptide ID.}

\item{pepCol}{A `character(1)` indicating the variable names in the 
`rowData` that contains the PEPs.}
}
\value{
A `QFeatures` object.
}
\description{
The functions takes the PEPs from the given assay's `rowData` and creates a 
new variable (`.FDR`) to it that contains the computed FDRs.
}
\examples{
scp1 <- computeFDR(scp1,
                   i = 1,
                   groupCol = "Sequence",
                   pepCol = "PEP")
## Check results
rowDataToDF(scp1, 1, c("PEP", ".FDR"))
}
