% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{computeSCR}
\alias{computeSCR}
\title{Compute the sample over carrier ratio (SCR)}
\usage{
computeSCR(obj, i, colDataCol, samplePattern, carrierPattern, verbose = TRUE)
}
\arguments{
\item{obj}{A \code{QFeatures} object.}

\item{i}{A \code{character()} or \code{integer()} indicating for which assay(s) the
SCR needs to be computed.}

\item{colDataCol}{A \code{character(1)} indicating the variable to take from
\code{colData(obj)} that gives the sample annotation.}

\item{samplePattern}{A \code{character(1)} pattern that matches the sample
encoding in \code{colDataCol}.}

\item{carrierPattern}{A \code{character(1)} pattern that matches the carrier
encoding in \code{colDataCol}. Only one match per assay is allowed, otherwise
only the first match is taken}
}
\value{
A \code{QFeatures} object for which the \code{rowData} of the given assay(s) is
augmented with the mean SCR (\code{meanSCR} variable).
}
\description{
The function computes the ratio of the intensities of sample channels over
the intentisty of the carrier channel for each feature. The ratios are
averaged within the assay.
}
\examples{
scp1 <- computeSCR(scp1, 
                   i = 1,
                   colDataCol = "SampleType",
                   carrierPattern = "Carrier",
                   samplePattern = "Blank|Macrophage|Monocyte")
## Check results
rowDataToDF(scp1, 1, "meanSCR")
}
