% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-app.R
\name{CreateMetaFile}
\alias{CreateMetaFile}
\title{CreateMetaFile}
\usage{
CreateMetaFile(
  srt,
  MetaFile,
  name = NULL,
  write_tools = FALSE,
  write_misc = FALSE,
  ignore_nlevel = 100,
  compression_level = 6,
  overwrite = FALSE
)
}
\arguments{
\item{srt}{}

\item{overwrite}{}
}
\description{
CreateMetaFile
}
