% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_DoubletDetection}
\alias{db_DoubletDetection}
\title{Run doublet-calling with DoubletDetection}
\usage{
db_DoubletDetection(srt, assay = "RNA", db_rate = ncol(srt)/1000 * 0.01, ...)
}
\arguments{
\item{...}{}
}
\description{
Run doublet-calling with DoubletDetection
}
\examples{
data("pancreas_sub")
pancreas_sub <- db_DoubletDetection(pancreas_sub)
CellDimPlot(pancreas_sub, reduction = "umap", group.by = "db.DoubletDetection_class")
FeatureDimPlot(pancreas_sub, reduction = "umap", features = "db.DoubletDetection_score")
}
