% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_scDblFinder}
\alias{db_scDblFinder}
\title{Run doublet-calling with scDblFinder}
\usage{
db_scDblFinder(srt, assay = "RNA", db_rate = ncol(srt)/1000 * 0.01, ...)
}
\arguments{
\item{...}{}
}
\description{
Run doublet-calling with scDblFinder
}
\examples{
data("pancreas_sub")
pancreas_sub <- db_scDblFinder(pancreas_sub)
CellDimPlot(pancreas_sub, reduction = "umap", group.by = "db.scDblFinder_class")
FeatureDimPlot(pancreas_sub, reduction = "umap", features = "db.scDblFinder_score")
}
