% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_scds}
\alias{db_scds}
\title{Run doublet-calling with scds}
\usage{
db_scds(
  srt,
  assay = "RNA",
  db_rate = ncol(srt)/1000 * 0.01,
  method = c("hybrid", "cxds", "bcds"),
  ...
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{assay}{The name of the assay to be used for doublet-calling. Default is "RNA".}

\item{db_rate}{The expected doublet rate. Default is calculated as ncol(srt) / 1000 * 0.01.}

\item{method}{The method to be used for doublet-calling. Options are "hybrid", "cxds", or "bcds".}

\item{...}{Additional arguments to be passed to scds::cxds_bcds_hybrid function.}
}
\description{
This function performs doublet-calling using the scds package on a Seurat object.
}
\examples{
data("pancreas_sub")
pancreas_sub <- db_scds(pancreas_sub, method = "hybrid")
CellDimPlot(pancreas_sub, reduction = "umap", group.by = "db.scds_hybrid_class")
FeatureDimPlot(pancreas_sub, reduction = "umap", features = "db.scds_hybrid_score")
}
