% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cellqc.R
\name{db_Scrublet}
\alias{db_Scrublet}
\title{Run doublet-calling with Scrublet}
\usage{
db_Scrublet(srt, assay = "RNA", db_rate = ncol(srt)/1000 * 0.01, ...)
}
\arguments{
\item{...}{}
}
\description{
Run doublet-calling with Scrublet
}
\examples{
data("pancreas_sub")
pancreas_sub <- db_Scrublet(pancreas_sub)
CellDimPlot(pancreas_sub, reduction = "umap", group.by = "db.Scrublet_class")
FeatureDimPlot(pancreas_sub, reduction = "umap", features = "db.Scrublet_score")
}
