% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{DCP_DiffPar}
\alias{DCP_DiffPar}
\title{Differential Parameter Test (DP and DA)}
\usage{
DCP_DiffPar(
  x,
  Par = c("A"),
  TOJR = NULL,
  alpha = 0.05,
  p.adjust.method = "BH",
  parallel.ncores = 1
)
}
\arguments{
\item{x}{Output from DCP_Rhythmicity(x1, x2)}

\item{Par}{One of "A", "phase", "M", "A&phase", or "A&phase&M"}

\item{TOJR}{TOJR output. If NULL, rhythm.joint object in x will be used}

\item{alpha}{P-value cutoff}

\item{p.adjust.method}{P-value adjustment method}

\item{parallel.ncores}{Number of cores for parallel processing}
}
\value{
Data frame with differential parameter test results
}
\description{
Tests for differences in amplitude, phase, and/or mesor between two groups
}
\examples{
\dontrun{
# rhythm.res is the output of DCP_Rhythmicity(x1, x2)
rhythm.diffPar = DCP_DiffPar(rhythm.res, Par = "A&phase")
}

}
\keyword{internal}
