% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{DCP_DiffR2}
\alias{DCP_DiffR2}
\title{Differential Rhythmicity Test (DR)}
\usage{
DCP_DiffR2(
  x,
  method = "LR",
  TOJR = NULL,
  alpha = 0.05,
  nSampling = 1000,
  Sampling.save = NULL,
  Sampling.file.label = "gII_vs_gI",
  p.adjust.method = "BH",
  parallel.ncores = 1
)
}
\arguments{
\item{x}{Output from DCP_Rhythmicity(x1, x2)}

\item{method}{One of "LR" (recommended) or "LR_sigma2"}

\item{TOJR}{TOJR output. If NULL, rhythm.joint object in x will be used}

\item{alpha}{P-value cutoff}

\item{nSampling}{Number of samplings for permutation/bootstrap}

\item{Sampling.save}{Directory to save sampling results}

\item{Sampling.file.label}{File label for saved results}

\item{p.adjust.method}{P-value adjustment method}

\item{parallel.ncores}{Number of cores for parallel processing}
}
\value{
Data frame with differential R^2 test results
}
\description{
Tests for differential rhythmicity between two groups using R^2 difference
}
\examples{
\dontrun{
# rhythm.res is the output of DCP_Rhythmicity(x1, x2)
rhythm.diffR2 = DCP_DiffR2(rhythm.res)
}

}
\keyword{internal}
