% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{DCP_DiffR2}
\alias{DCP_DiffR2}
\title{Differential Rhythmicity Test (DR)}
\usage{
DCP_DiffR2(
  x,
  method = "LR",
  TOJR = NULL,
  alpha = 0.05,
  nSampling = 1000,
  Sampling.save = NULL,
  Sampling.file.label = "gII_vs_gI",
  p.adjust.method = "BH",
  parallel.ncores = 1
)
}
\arguments{
\item{x}{Output from DCP_Rhythmicity(x1, x2)}

\item{method}{One of "LR" (recommended), "permutation", "bootstrap", "LR_sigma2"}

\item{TOJR}{TOJR output. If NULL, rhythm.joint object in x will be used}

\item{alpha}{P-value cutoff}

\item{nSampling}{Number of samplings for permutation/bootstrap}

\item{Sampling.save}{Directory to save sampling results}

\item{Sampling.file.label}{File label for saved results}

\item{p.adjust.method}{P-value adjustment method}

\item{parallel.ncores}{Number of cores for parallel processing}
}
\value{
Data frame with differential R² test results
}
\description{
Tests for differential rhythmicity between two groups using R² difference
}
\examples{
x = DCP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)
rhythm.res = DCP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
rhythm.diffR2 = DCP_DiffR2(rhythm.res)

}
