% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{DCP_Rhythmicity}
\alias{DCP_Rhythmicity}
\title{Rhythmicity Analysis with Cosinor Model}
\usage{
DCP_Rhythmicity(
  x1,
  x2 = NULL,
  method = "Sidak_FS",
  period = 24,
  amp.cutoff = 0,
  alpha = 0.05,
  alpha.FDR = 0.05,
  CI = FALSE,
  p.adjust.method = "BH",
  parallel.ncores = 1
)
}
\arguments{
\item{x1}{Group 1 data (list with data, time, gname components)}

\item{x2}{Group 2 data (optional, same format as x1)}

\item{method}{Algorithm for joint rhythmicity categorization. Should be one of
"Sidak_FS" (default, recommended), "Sidak_BS", "VDA", "AWFisher"}

\item{period}{Oscillation cycle length (default 24 for circadian)}

\item{amp.cutoff}{Only genes with amplitude > amp.cutoff are considered rhythmic}

\item{alpha}{Threshold for rhythmicity p-value}

\item{alpha.FDR}{Threshold for FDR-adjusted rhythmicity p-value}

\item{CI}{Calculate confidence intervals?}

\item{p.adjust.method}{P-value adjustment method}

\item{parallel.ncores}{Number of cores for parallel processing}
}
\value{
List with:
  \item{x1, x2}{Input data with added rhythmicity analysis}
  \item{gname_overlap}{Overlapping gene names}
  \item{rhythm.joint}{Data frame with joint rhythmicity categories}
}
\description{
This function either takes single-group data and performs only rhythmicity analysis,
or takes two-group data and also categorize the genes into types of joint rhythmicity (TOJR).
}
\details{
The methods "Sidak_FS" and "Sidak_BS" implement selective sequential model selection
with Sidak adjusted p-value. "FS" represents forward stop, and "BS" basic stop.
The method "Sidak_FS" has better type I error control compared to Venn diagram
analysis (VDA) and adaptively weighted Fisher's method (AWFisher).
}
\examples{
\dontrun{
# x1, x2 are lists with elements data (genes x samples), time, gname
rhythm.res = DCP_Rhythmicity(x1 = x1, x2 = x2)
}

}
\references{
Fithian, W., Taylor, J., Tibshirani, R., & Tibshirani, R. (2015). Selective
sequential model selection. arXiv preprint arXiv:1512.02565.
}
\keyword{internal}
