% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_contour.R
\name{.plausibility_contour}
\alias{.plausibility_contour}
\title{internal plausibility_contour calculation function}
\usage{
.plausibility_contour(Y_cand, s_aug, Y_aug, w_aug, d_aug, pred_fun, T_dfun)
}
\arguments{
\item{T_dfun}{}
}
\value{
a data frame of Y_cand and p_y
}
\description{
internal plausibility_contour calculation function
}
\keyword{internal}
