% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{.prime}
\alias{.prime}
\title{Transform the hyperparameters}
\usage{
.prime(s0, s, Y, global, eta, m, dfun)
}
\arguments{
\item{dfun}{}
}
\value{

}
\description{
Transform the hyperparameters
}
\keyword{internal}
