% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{drop_data}
\alias{drop_data}
\alias{drop_data.ggplot}
\alias{drop_data.patchwork}
\alias{drop_data.default}
\title{Drop all data in the plot (only one observation is kept)}
\usage{
drop_data(p)

\method{drop_data}{ggplot}(p)

\method{drop_data}{patchwork}(p)

\method{drop_data}{default}(p)
}
\arguments{
\item{p}{A \code{patchwork} object.}
}
\description{
Drop all data in the plot (only one observation is kept)
}
\examples{
library(ggplot2)
p <- ggplot(data = mtcars, aes(x = mpg, y = wt, colour = cyl)) +
  geom_point() +
  scale_x_continuous(limits = c(10, 30)) +
  scale_y_continuous(limits = c(1, 6)) +
  theme_scp()
object.size(p)

p_drop <- drop_data(p)
object.size(p_drop)

p / p_drop

}
