% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DynamicHeatmap}
\alias{DynamicHeatmap}
\title{Heatmap plot for dynamic features along lineages}
\usage{
DynamicHeatmap(
  srt,
  lineages,
  feature_from = lineages,
  exp_method = c("zscore", "raw", "log2fc", "log1p"),
  slot = "counts",
  assay = "RNA",
  use_fitted = FALSE,
  lib_normalize = TRUE,
  libsize = NULL,
  min_expcells = 20,
  r.sq = 0.2,
  dev.expl = 0.2,
  padjust = 0.05,
  cell_density = 1,
  order_by = c("peaktime", "valleytime"),
  decreasing = FALSE,
  feature_split = NULL,
  row_title_size = 12,
  n_split = NULL,
  heatmap_split_by = lineages,
  split_method = c("mfuzz", "kmeans", "kmeans-peaktime", "hclust", "hclust-peaktime"),
  fuzzification = NULL,
  show_fuzzification = FALSE,
  anno_terms = FALSE,
  anno_keys = FALSE,
  anno_features = FALSE,
  IDtype = "symbol",
  species = "Homo_sapiens",
  db_IDtype = "symbol",
  db_update = FALSE,
  db_version = "latest",
  Ensembl_version = 103,
  mirror = NULL,
  enrichment = "GO_BP",
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  universe = NULL,
  GO_simplify = FALSE,
  GO_simplify_padjustCutoff = 0.2,
  simplify_method = "Rel",
  simplify_similarityCutoff = 0.7,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  topTerm = 5,
  show_id = TRUE,
  topWord = 20,
  min_word_length = 3,
  exclude_words = c("cell", "cellular", "dna", "rna", "protein", "development", "system",
    "regulation", "positive", "negative", "response", "process"),
  anno_width = unit(c(4, 2, 2), "in"),
  anno_size = c(6, 10),
  nlabel = 20,
  features_label = NULL,
  label_size = 10,
  label_color = "black",
  pseudotime_label = NULL,
  pseudotime_label_color = "black",
  pseudotime_label_linetype = 2,
  pseudotime_label_linewidth = 2,
  heatmap_palette = "RdBu",
  pseudotime_palette = "cividis",
  feature_split_palette = "jama",
  cell_annotation = NULL,
  cell_palette = "Paired",
  cell_palcolor = NULL,
  feature_annotation = NULL,
  feature_palette = "Dark2",
  feature_palcolor = NULL,
  reverse_ht = NULL,
  use_raster = NULL,
  height = NULL,
  width = NULL,
  units = "inch",
  seed = 11
)
}
\arguments{
\item{seed}{}
}
\description{
Heatmap plot for dynamic features along lineages
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
pancreas_sub <- RunDynamicFeatures(pancreas_sub, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
ht_result1 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1"), order_by = "peaktime",
  cell_annotation = "SubCellType",
  n_split = 5, split_method = "kmeans-peaktime",
  height = 6, width = 7, use_raster = FALSE
)
ht_result1$plot

ht_result2 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_split = 5, reverse_ht = "Lineage1",
  height = 6, width = 7, use_raster = FALSE
)
ht_result2$plot

ht_result3 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_split = 5, reverse_ht = "Lineage1",
  species = "Mus_musculus", anno_terms = TRUE, anno_keys = TRUE, anno_features = TRUE,
  height = 6, width = 7, use_raster = FALSE
)
ht_result3$plot
}
