% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DynamicHeatmap}
\alias{DynamicHeatmap}
\title{Heatmap plot for dynamic genes along lineages}
\usage{
DynamicHeatmap(
  srt,
  lineages,
  feature_from = lineages,
  exp_method = c("zscore", "raw", "log2fc", "log1p"),
  slot = "counts",
  assay = "RNA",
  use_fitted = FALSE,
  lib_normalize = TRUE,
  lib_size = NULL,
  min_expcells = 20,
  r.sq = 0.2,
  dev.expl = 0.2,
  padjust = 0.05,
  cell_density = 1,
  order_by = "peak_time",
  decreasing = FALSE,
  feature_clusters = NULL,
  cluster_title_size = 12,
  n_cluster = NULL,
  cluster_by = lineages,
  cluster_method = "kmeans",
  fuzzification = NULL,
  show_fuzzification = FALSE,
  nlabel = 20,
  feature_label = NULL,
  label_size = 10,
  label_color = "black",
  pseudotime_label = NULL,
  pseudotime_label_color = "black",
  pseudotime_label_linetype = 2,
  pseudotime_label_linewidth = 2,
  ht_palette = "RdBu",
  pseudotime_palette = "cividis",
  cell_annotation = NULL,
  cell_palette = "Paired",
  cell_palcolor = NULL,
  feature_annotation = NULL,
  feature_palette = "Dark2",
  feature_palcolor = NULL,
  reverse_ht = NULL,
  use_raster = NULL,
  height = NULL,
  width = NULL,
  units = "inch",
  seed = 11
)
}
\description{
Heatmap plot for dynamic genes along lineages
}
\examples{
data("pancreas1k")
pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "UMAP")
pancreas1k <- RunDynamicFeatures(pancreas1k, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
ht_result <- DynamicHeatmap(
  srt = pancreas1k,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_cluster = 6, reverse_ht = "Lineage1",
  height = 5, width = 7, use_raster = FALSE
)
ht_result$plot
}
