% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DynamicHeatmap}
\alias{DynamicHeatmap}
\title{Heatmap plot for dynamic features along lineages}
\usage{
DynamicHeatmap(
  srt,
  lineages,
  features = NULL,
  feature_from = lineages,
  use_fitted = FALSE,
  border = TRUE,
  flip = FALSE,
  min_expcells = 20,
  r.sq = 0.2,
  dev.expl = 0.2,
  padjust = 0.05,
  cell_density = 1,
  order_by = c("peaktime", "valleytime"),
  slot = "counts",
  assay = NULL,
  exp_method = c("zscore", "raw", "fc", "log2fc", "log1p"),
  limits = NULL,
  lib_normalize = identical(slot, "counts"),
  libsize = NULL,
  family = NULL,
  cluster_features_by = NULL,
  cluster_rows = FALSE,
  cluster_row_slices = FALSE,
  cluster_columns = FALSE,
  cluster_column_slices = FALSE,
  show_row_names = FALSE,
  show_column_names = FALSE,
  row_names_side = ifelse(flip, "left", "right"),
  column_names_side = ifelse(flip, "bottom", "top"),
  row_names_rot = 0,
  column_names_rot = 90,
  row_title_side = "left",
  column_title_side = "top",
  row_title_rot = 0,
  column_title_rot = ifelse(flip, 90, 0),
  feature_split = NULL,
  feature_split_by = NULL,
  n_split = NULL,
  split_method = c("mfuzz", "kmeans", "kmeans-peaktime", "hclust", "hclust-peaktime"),
  decreasing = FALSE,
  fuzzification = NULL,
  show_fuzzification = FALSE,
  anno_terms = FALSE,
  anno_keys = FALSE,
  anno_features = FALSE,
  terms_width = unit(4, "in"),
  terms_fontsize = 8,
  keys_width = unit(2, "in"),
  keys_fontsize = c(6, 10),
  features_width = unit(2, "in"),
  features_fontsize = c(6, 10),
  IDtype = "symbol",
  species = "Homo_sapiens",
  db_update = FALSE,
  db_version = "latest",
  db_combine = FALSE,
  convert_species = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  db = "GO_BP",
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  GO_simplify = FALSE,
  GO_simplify_cutoff = "p.adjust < 0.05",
  simplify_method = "Wang",
  simplify_similarityCutoff = 0.7,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  topTerm = 5,
  show_termid = FALSE,
  topWord = 20,
  min_word_length = 3,
  exclude_words = c("cell", "cellular", "dna", "rna", "protein", "development",
    "organization", "system", "regulation", "positive", "negative", "response",
    "process"),
  nlabel = 20,
  features_label = NULL,
  label_size = 10,
  label_color = "black",
  pseudotime_label = NULL,
  pseudotime_label_color = "black",
  pseudotime_label_linetype = 2,
  pseudotime_label_linewidth = 3,
  heatmap_palette = "RdBu",
  heatmap_palcolor = NULL,
  pseudotime_palette = "cividis",
  pseudotime_palcolor = NULL,
  feature_split_palette = "jama",
  feature_split_palcolor = NULL,
  cell_annotation = NULL,
  cell_annotation_palette = "Paired",
  cell_annotation_palcolor = NULL,
  cell_annotation_params = list(),
  feature_annotation = NULL,
  feature_annotation_palette = "Dark2",
  feature_annotation_palcolor = NULL,
  feature_annotation_params = list(),
  separate_annotation = NULL,
  separate_annotation_palette = "Paired",
  separate_annotation_palcolor = NULL,
  separate_annotation_params = if (flip) list(width = grid::unit(2, "cm")) else
    list(height = grid::unit(2, "cm")),
  reverse_ht = NULL,
  use_raster = NULL,
  raster_device = "png",
  raster_by_magick = FALSE,
  height = NULL,
  width = NULL,
  units = "inch",
  seed = 11,
  ht_params = list()
)
}
\arguments{
\item{raster_by_magick}{}
}
\description{
Heatmap plot for dynamic features along lineages
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
pancreas_sub <- RunDynamicFeatures(pancreas_sub, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
ht1 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = "Lineage1",
  cell_annotation = "SubCellType",
  n_split = 5, split_method = "kmeans-peaktime"
)
ht1$plot
panel_fix(ht1$plot, raster = TRUE, dpi = 50)

ht2 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = "Lineage1",
  features = c("Sox9", "Neurod2", "Isl1", "Rbp4", "S_score", "G2M_score"),
  cell_annotation = "SubCellType"
)
ht2$plot
panel_fix(ht2$plot, height = 5, width = 5, raster = TRUE, dpi = 50)

ht3 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_split = 5, split_method = "kmeans",
  cluster_rows = TRUE
)
ht3$plot

pancreas_sub <- AnnotateFeatures(pancreas_sub, species = "Mus_musculus", db = c("TF", "SP"))
ht4 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  use_fitted = TRUE,
  n_split = 6, split_method = "mfuzz",
  reverse_ht = "Lineage1",
  heatmap_palette = "viridis",
  cell_annotation = c("SubCellType", "Phase", "G2M_score"),
  cell_annotation_palette = c("Paired", "jama", "Purples"),
  separate_annotation = list("SubCellType", c("Nnat", "Irx1")),
  separate_annotation_palette = c("Paired", "Set1"),
  separate_annotation_params = list(height = grid::unit(2, "cm")),
  feature_annotation = c("TF", "SP"),
  feature_annotation_palcolor = list(c("gold", "steelblue"), c("forestgreen")),
  pseudotime_label = 25,
  pseudotime_label_color = "red",
  width = 1, height = 2.5
)
ht4$plot

ht5 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  use_fitted = TRUE,
  n_split = 6, split_method = "mfuzz",
  reverse_ht = "Lineage1",
  heatmap_palette = "viridis",
  cell_annotation = c("SubCellType", "Phase", "G2M_score"),
  cell_annotation_palette = c("Paired", "jama", "Purples"),
  separate_annotation = list("SubCellType", c("Nnat", "Irx1")),
  separate_annotation_palette = c("Paired", "Set1"),
  separate_annotation_params = list(width = grid::unit(2, "cm")),
  feature_annotation = c("TF", "SP"),
  feature_annotation_palcolor = list(c("gold", "steelblue"), c("forestgreen")),
  pseudotime_label = 25,
  pseudotime_label_color = "red",
  flip = TRUE, column_title_rot = 45,
  width = 2.5, height = 1
)
ht5$plot

ht6 <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_split = 5, reverse_ht = "Lineage1",
  species = "Mus_musculus", db = "GO_BP",
  anno_terms = TRUE, anno_keys = TRUE, anno_features = TRUE
)
ht6$plot

}
