% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DynamicPlot}
\alias{DynamicPlot}
\title{DynamicPlot}
\usage{
DynamicPlot(
  srt,
  lineages,
  features,
  group.by = NULL,
  cells = NULL,
  slot = "counts",
  assay = NULL,
  family = NULL,
  exp_method = c("log1p", "raw", "zscore", "fc", "log2fc"),
  lib_normalize = identical(slot, "counts"),
  libsize = NULL,
  compare_lineages = TRUE,
  compare_features = FALSE,
  add_line = TRUE,
  add_interval = TRUE,
  line.size = 1,
  line_palette = "Dark2",
  line_palcolor = NULL,
  add_point = TRUE,
  pt.size = 1,
  point_palette = "Paired",
  point_palcolor = NULL,
  add_rug = TRUE,
  flip = FALSE,
  reverse = FALSE,
  x_order = c("value", "rank"),
  aspect.ratio = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  theme_use = "theme_scp",
  theme_args = list(),
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  seed = 11
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{lineages}{A character vector specifying the lineages to plot.}

\item{features}{A character vector specifying the features to plot.}

\item{group.by}{A character specifying a metadata column to group the cells by. Default is NULL.}

\item{cells}{A character vector specifying the cells to include in the plot. Default is NULL.}

\item{slot}{A character string specifying the slot to use for the analysis. Default is "counts".}

\item{assay}{A character string specifying the assay to use for the analysis. Default is NULL.}

\item{family}{A character specifying the model used to calculate the dynamic features if needed. By default, this parameter is set to NULL, and the appropriate family will be automatically determined.}

\item{exp_method}{A character specifying the method to transform the expression values. Default is "log1p" with options "log1p", "raw", "zscore", "fc", "log2fc".}

\item{lib_normalize}{A boolean specifying whether to normalize the expression values using library size. By default, if the \code{slot} is counts, this parameter is set to TRUE. Otherwise, it is set to FALSE.}

\item{libsize}{A numeric vector specifying the library size for each cell. Default is NULL.}

\item{compare_lineages}{A boolean specifying whether to compare the lineages in the plot. Default is TRUE.}

\item{compare_features}{A boolean specifying whether to compare the features in the plot. Default is FALSE.}

\item{add_line}{A boolean specifying whether to add lines to the plot. Default is TRUE.}

\item{add_interval}{A boolean specifying whether to add confidence intervals to the plot. Default is TRUE.}

\item{line.size}{A numeric specifying the size of the lines. Default is 1.}

\item{line_palette}{A character string specifying the name of the palette to use for the line colors. Default is "Dark2".}

\item{line_palcolor}{A vector specifying the colors to use for the line palette. Default is NULL.}

\item{add_point}{A boolean specifying whether to add points to the plot. Default is TRUE.}

\item{pt.size}{A numeric specifying the size of the points. Default is 1.}

\item{point_palette}{A character string specifying the name of the palette to use for the point colors. Default is "Paired".}

\item{point_palcolor}{A vector specifying the colors to use for the point palette. Default is NULL.}

\item{add_rug}{A boolean specifying whether to add rugs to the plot. Default is TRUE.}

\item{flip}{A boolean specifying whether to flip the x-axis. Default is FALSE.}

\item{reverse}{A boolean specifying whether to reverse the x-axis. Default is FALSE.}

\item{x_order}{A character specifying the order of the x-axis values. Default is c("value", "rank").}

\item{aspect.ratio}{A numeric specifying the aspect ratio of the plot. Default is NULL.}

\item{legend.position}{A character string specifying the position of the legend in the plot. Default is "right".}

\item{legend.direction}{A character string specifying the direction of the legend in the plot. Default is "vertical".}

\item{theme_use}{A character string specifying the name of the theme to use for the plot. Default is "theme_scp".}

\item{theme_args}{A list specifying the arguments to pass to the theme function. Default is list().}

\item{combine}{A boolean specifying whether to combine multiple plots into a single plot. Default is TRUE.}

\item{nrow}{A numeric specifying the number of rows in the combined plot. Default is NULL.}

\item{ncol}{A numeric specifying the number of columns in the combined plot. Default is NULL.}

\item{byrow}{A boolean specifying whether to fill plots by row in the combined plot. Default is TRUE.}

\item{seed}{A numeric specifying the random seed. Default is 11.}
}
\description{
Plot dynamic features across pseudotime.
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
CellDimPlot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP", lineages = paste0("Lineage", 1:3), lineages_span = 0.1)
DynamicPlot(
  srt = pancreas_sub,
  lineages = "Lineage1",
  features = c("Nnat", "Irx1", "G2M_score"),
  group.by = "SubCellType",
  compare_features = TRUE
)
DynamicPlot(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  features = c("Nnat", "Irx1", "G2M_score"),
  group.by = "SubCellType",
  compare_lineages = TRUE,
  compare_features = FALSE
)
DynamicPlot(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  features = c("Nnat", "Irx1", "G2M_score"),
  group.by = "SubCellType",
  compare_lineages = FALSE,
  compare_features = FALSE
)
}
\seealso{
\code{\link{RunDynamicFeatures}}
}
