% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{DynamicPlot}
\alias{DynamicPlot}
\title{Plot for dynamic features along lineages.}
\usage{
DynamicPlot(
  srt,
  features,
  lineages,
  group.by = NULL,
  cells = NULL,
  slot = "counts",
  assay = NULL,
  family = NULL,
  exp_method = c("log1p", "raw", "zscore", "fc", "log2fc"),
  lib_normalize = identical(slot, "counts"),
  libsize = NULL,
  compare_lineages = TRUE,
  compare_features = FALSE,
  add_line = TRUE,
  add_interval = TRUE,
  line.size = 1,
  line_palette = "Dark2",
  line_palcolor = NULL,
  add_point = TRUE,
  pt.size = 1,
  point_palette = "Paired",
  point_palcolor = NULL,
  add_rug = TRUE,
  flip = FALSE,
  reverse = FALSE,
  x_order = c("value", "rank"),
  aspect.ratio = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  theme_use = "theme_scp",
  theme_args = list(),
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  seed = 11
)
}
\arguments{
\item{x_order}{}
}
\description{
Plot for dynamic features along lineages.
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
DynamicPlot(
  srt = pancreas_sub,
  lineages = "Lineage1",
  features = c("Nnat", "Irx1", "G2M_score"),
  group.by = "SubCellType",
  compare_features = TRUE
)
DynamicPlot(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  features = c("Nnat", "Irx1", "G2M_score"),
  group.by = "SubCellType",
  compare_lineages = TRUE,
  compare_features = FALSE
)
DynamicPlot(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  features = c("Nnat", "Irx1", "G2M_score"),
  group.by = "SubCellType",
  compare_lineages = FALSE,
  compare_features = FALSE
)
}
