% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{EnrichmentPlot}
\alias{EnrichmentPlot}
\title{EnrichmentPlot}
\usage{
EnrichmentPlot(
  srt,
  db = "GO_BP",
  group_by = NULL,
  group_use = NULL,
  test.use = "wilcox",
  res = NULL,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  plot_type = c("bar", "lollipop", "wordcloud"),
  palette = "Spectral",
  topTerm = 6,
  topWord = 100,
  word_type = c("term", "feature"),
  word_size = c(2, 8),
  min_word_length = 3,
  exclude_words = c("cell", "cellular", "dna", "rna", "protein", "development",
    "organization", "system", "regulation", "positive", "negative", "response",
    "process"),
  aspect.ratio = 1,
  base_size = 12,
  character_width = 50,
  line_height = 0.7,
  legend.position = "right",
  legend.direction = "vertical",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  align = "hv",
  axis = "lr"
)
}
\arguments{
\item{group_use}{}

\item{pvalueCutoff}{pvalueCutoff}

\item{padjustCutoff}{padjustCutoff}

\item{plot_type}{Type of plot to be visualized.}

\item{palette}{palette}

\item{topTerm}{Number of terms to plot if \code{plot_type="bar"} or \code{plot_type="lollipop"}.}

\item{topWord}{Number of features to plot if \code{plot_type="wordcloud"}.}

\item{combine}{combine}

\item{nrow}{nrow}

\item{ncol}{ncol}

\item{byrow}{byrow}
}
\description{
EnrichmentPlot
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType")
pancreas_sub <- RunEnrichment(srt = pancreas_sub, group_by = "CellType", db = "GO_BP", species = "Mus_musculus")
EnrichmentPlot(pancreas_sub, group_by = "CellType", plot_type = "bar")
EnrichmentPlot(pancreas_sub, group_by = "CellType", plot_type = "bar", character_width = 30, base_size = 10)

EnrichmentPlot(pancreas_sub, group_by = "CellType", group_use = c("Ductal", "Endocrine"), plot_type = "lollipop", ncol = 1)
EnrichmentPlot(pancreas_sub, group_by = "CellType", group_use = c("Ductal", "Endocrine"), plot_type = "wordcloud")
EnrichmentPlot(pancreas_sub, group_by = "CellType", group_use = c("Ductal", "Endocrine"), plot_type = "wordcloud", word_type = "feature")

pancreas_sub <- RunEnrichment(srt = pancreas_sub, group_by = "CellType", db = c("MP", "DO"), convert_species = TRUE, species = "Mus_musculus")
EnrichmentPlot(pancreas_sub, db = c("MP", "DO"), group_by = "CellType", group_use = c("Ductal", "Endocrine"), plot_type = "lollipop")

}
