% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exist_Python_pkgs}
\alias{exist_Python_pkgs}
\title{Check if the python package exists in the environment}
\usage{
exist_Python_pkgs(packages, envname = "SCP")
}
\arguments{
\item{envname}{The name of, or path to, a Python environment.}
}
\description{
Check if the python package exists in the environment
}
