% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{ExpCorPlot}
\alias{ExpCorPlot}
\title{Features correlation plot}
\usage{
ExpCorPlot(
  srt,
  features,
  group.by = NULL,
  split.by = NULL,
  cells = NULL,
  slot = "data",
  assay = "RNA",
  cor_method = "pearson",
  adjust = 1,
  margin = 1,
  reverse = FALSE,
  add_equation = FALSE,
  add_r2 = TRUE,
  add_pvalue = TRUE,
  add_smooth = TRUE,
  palette = "Paired",
  palcolor = NULL,
  bg_color = "grey80",
  pt.size = NULL,
  pt.alpha = 1,
  cells.highlight = NULL,
  cols.highlight = "black",
  sizes.highlight = 1,
  alpha.highlight = 1,
  stroke.highlight = 0.5,
  calculate_coexp = FALSE,
  raster = NULL,
  raster.dpi = c(512, 512),
  theme_use = "theme_scp",
  title = NULL,
  subtitle = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  align = "hv",
  axis = "lr",
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\description{
Features correlation plot
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
ExpCorPlot(pancreas_sub, features = c("Ghrl", "Gcg", "Ins1", "Ins2"), group.by = "SubCellType")
}
