% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_scp.R
\name{fast_scp}
\alias{fast_scp}
\title{Fast Spatial Conformal Prediction Intervals}
\usage{
fast_scp(s0, s, Y, w, Q, alpha = 0.05, std = TRUE)
}
\arguments{
\item{s0}{prediction location}

\item{s}{an \eqn{n \times d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{w}{weights for the non-conformity measures.}

\item{Q}{an \eqn{n+1 \times n+1} inverse of estimated covariance matrix}

\item{alpha}{significance level. Defaults to 0.05.}
}
\value{
A vector of lower and upper bounds of the conformal prediction interval.
}
\description{
Internal function provides the conformal prediction interval
for spatial location \code{s0} given partial obserations \code{s1,...,sm, Y1,.., Ym}
when the square of the standard kriging residuals are used as the non-conformality measures,
and weights are provided.
}
