% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{FeatureDimPlot}
\alias{FeatureDimPlot}
\title{Dimensional reduction plot for expression.}
\usage{
FeatureDimPlot(
  srt,
  features,
  reduction = NULL,
  dims = c(1, 2),
  split.by = NULL,
  cells = NULL,
  slot = "data",
  assay = NULL,
  show_stat = TRUE,
  palette = ifelse(isTRUE(compare_features), "Set1", "Spectral"),
  palcolor = NULL,
  pt.size = NULL,
  pt.alpha = 1,
  bg_cutoff = 0,
  bg_color = "grey80",
  keep_scale = NULL,
  lower_quantile = 0,
  upper_quantile = 0.99,
  lower_cutoff = NULL,
  upper_cutoff = NULL,
  add_density = FALSE,
  density_color = "grey80",
  density_filled = FALSE,
  density_filled_palette = "Greys",
  density_filled_color = NULL,
  cells.highlight = NULL,
  cols.highlight = "black",
  sizes.highlight = 1,
  alpha.highlight = 1,
  stroke.highlight = 0.5,
  calculate_coexp = FALSE,
  compare_features = FALSE,
  color_blend_mode = c("blend", "mix", "screen", "multiply"),
  label = FALSE,
  label.size = 4,
  label.fg = "white",
  label.bg = "black",
  label.bg.r = 0.1,
  label_insitu = FALSE,
  label_repel = FALSE,
  label_repulsion = 20,
  label_point_size = 1,
  label_point_color = "black",
  label_segment_color = "black",
  lineages = NULL,
  lineages_trim = c(0.01, 0.99),
  lineages_span = 0.75,
  lineages_palette = "Dark2",
  lineages_palcolor = NULL,
  lineages_arrow = arrow(length = unit(0.1, "inches")),
  lineages_line_size = 1,
  lineages_line_bg = "white",
  lineages_line_bg_r = 0.5,
  lineages_whiskers = FALSE,
  lineages_whiskers_size = 0.5,
  lineages_whiskers_alpha = 0.5,
  graph = NULL,
  edge_size = c(0.05, 0.5),
  edge_alpha = 0.1,
  edge_color = "grey40",
  hex = FALSE,
  hex.linewidth = 0.5,
  hex.color = "grey90",
  hex.bins = 50,
  hex.binwidth = NULL,
  raster = NULL,
  raster.dpi = c(512, 512),
  theme_use = "theme_scp",
  aspect.ratio = 1,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  lab_cex = 1,
  xlen_npc = 0.15,
  ylen_npc = 0.15,
  legend.position = "right",
  legend.direction = "vertical",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  force = FALSE,
  seed = 11
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{features}{Name of one or more metadata columns to group (color) cells by (for example, orig.ident).}

\item{reduction}{Which dimensionality reduction to use.}

\item{split.by}{Name of a metadata column to split plot by.}

\item{palette}{Name of palette to use. Default is "Paired".}

\item{pt.size}{Point size for plotting.}

\item{pt.alpha}{Point transparency.}

\item{bg_color}{Color value for NA points.}

\item{keep_scale}{How to handle the color scale across multiple plots. Options are:
\itemize{
  \item{"feature" (default; by row/feature scaling):}{ The plots for each individual feature are scaled to the maximum expression of the feature across the conditions provided to 'split.by'.}
  \item{"all" (universal scaling):}{ The plots for all features and conditions are scaled to the maximum expression value for the feature with the highest overall expression.}
  \item{NULL (no scaling):}{ Each individual plot is scaled to the maximum expression value of the feature in the condition provided to 'split.by'. Be aware setting NULL will result in color scales that are not comparable between plots.}
}}

\item{upper_cutoff}{}

\item{cells.highlight}{A vector of names of cells to highlight.}

\item{cols.highlight}{A vector of colors to highlight the cells.}

\item{sizes.highlight}{Size of highlighted cells.}

\item{legend.position}{The position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{legend.direction}{Layout of items in legends ("horizontal" or "vertical")}

\item{combine}{Whether to arrange multiple plots into a grid.}

\item{nrow}{Number of rows in the plot grid.}

\item{ncol}{Number of columns in the plot grid.}

\item{byrow}{Logical value indicating if the plots should be arrange by row (default) or by column.}
}
\value{
A single ggplot object if combine = TRUE; otherwise, a list of ggplot objects
}
\description{
Plotting cell points on a reduced 2D plane and coloring according to the expression of the features.
}
\examples{
#' library(dplyr)
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
FeatureDimPlot(pancreas_sub, features = "G2M_score", reduction = "UMAP")
FeatureDimPlot(pancreas_sub, features = "G2M_score", reduction = "UMAP") \%>\% panel_fix(height = 2)
FeatureDimPlot(pancreas_sub, features = "G2M_score", reduction = "UMAP", label = TRUE)
FeatureDimPlot(pancreas_sub, features = "G2M_score", reduction = "UMAP", hex = TRUE)

pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
FeatureDimPlot(pancreas_sub, features = "Lineage3", reduction = "UMAP", lineages = "Lineage3")
FeatureDimPlot(pancreas_sub, features = "Lineage3", reduction = "UMAP", lineages = "Lineage3", lineages_span = 0.1)

FeatureDimPlot(pancreas_sub, c("Ins1", "Gcg", "Sst", "Ghrl"), reduction = "UMAP")
FeatureDimPlot(pancreas_sub, c("Ins1", "Gcg", "Sst", "Ghrl"), reduction = "UMAP", lower_quantile = 0, upper_quantile = 0.8)
FeatureDimPlot(pancreas_sub, c("Ins1", "Gcg", "Sst", "Ghrl"), reduction = "UMAP", lower_cutoff = 1, upper_cutoff = 4)
FeatureDimPlot(pancreas_sub, c("Ins1", "Gcg", "Sst", "Ghrl"), reduction = "UMAP", bg_cutoff = 1, lower_cutoff = 1, upper_cutoff = 4)
FeatureDimPlot(pancreas_sub, c("Ins1", "Gcg", "Sst", "Ghrl"), reduction = "UMAP", keep_scale = "all")
FeatureDimPlot(pancreas_sub, c("Sst", "Ghrl"), split.by = "Phase", reduction = "UMAP", keep_scale = "feature")

FeatureDimPlot(pancreas_sub,
  features = c("Ins1", "Gcg", "Sst", "Ghrl"), pt.size = 1,
  compare_features = TRUE, color_blend_mode = "blend",
  label = TRUE, label_insitu = TRUE
)
FeatureDimPlot(pancreas_sub,
  features = c("S_score", "G2M_score"), pt.size = 1, palcolor = c("red", "green"),
  compare_features = TRUE, color_blend_mode = "mix", title = "mix",
  label = TRUE, label_insitu = TRUE
)
FeatureDimPlot(pancreas_sub,
  features = c("S_score", "G2M_score"), pt.size = 1, palcolor = c("red", "green"),
  compare_features = TRUE, color_blend_mode = "blend", title = "blend",
  label = TRUE, label_insitu = TRUE
)
FeatureDimPlot(pancreas_sub,
  features = c("S_score", "G2M_score"), pt.size = 1, palcolor = c("red", "green"),
  compare_features = TRUE, color_blend_mode = "screen", title = "screen",
  label = TRUE, label_insitu = TRUE
)
FeatureDimPlot(pancreas_sub,
  features = c("S_score", "G2M_score"), pt.size = 1, palcolor = c("red", "green"),
  compare_features = TRUE, color_blend_mode = "multiply", title = "multiply",
  label = TRUE, label_insitu = TRUE
)

library(scales)
par(mfrow = c(2, 2))
show_col(c("red", "green", blendcolors(c("red", "green"), mode = "mix")), ncol = 4)
text(3.5, -0.5, "mix", cex = 1.2)
show_col(c("red", "green", blendcolors(c("red", "green"), mode = "blend")), ncol = 4)
text(3.5, -0.5, "blend", cex = 1.2)
show_col(c("red", "green", blendcolors(c("red", "green"), mode = "screen")), ncol = 4)
text(3.5, -0.5, "screen", cex = 1.2)
show_col(c("red", "green", blendcolors(c("red", "green"), mode = "multiply")), ncol = 4)
text(3.5, -0.5, "multiply", cex = 1.2)
}
