% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{FeatureDimPlot3D}
\alias{FeatureDimPlot3D}
\title{3D-Dimensional reduction plot for gene expression visualization.}
\usage{
FeatureDimPlot3D(
  srt,
  features = NULL,
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  split.by = NULL,
  slot = "data",
  assay = NULL,
  calculate_coexp = FALSE,
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{features}{Vector of features to plot. Features can be gene names in Assay or names of numeric columns in meta.data.}

\item{reduction}{Which dimensionality reduction to use. If not specified, will use the reduction returned by \code{\link{DefaultReduction}}.}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions.}

\item{split.by}{Name of a column in meta.data to split plot by.}

\item{slot}{Which slot to pull expression data from? Default is \code{data}.}

\item{assay}{Which assay to pull expression data from. If \code{NULL}, will use the assay returned by \code{\link{DefaultAssay}}.}

\item{calculate_coexp}{Whether to calculate the co-expression value (geometric mean) of the features.}

\item{pt.size}{Point size for plotting.}

\item{cells.highlight}{A vector of cell names to highlight.}

\item{cols.highlight}{Color used to highlight the cells.}

\item{shape.highlight}{Shape of the cell to highlight. See \href{https://plotly.com/r/reference/scattergl/#scattergl-marker-symbol}{scattergl-marker-symbol}}

\item{sizes.highlight}{Size of highlighted cells.}

\item{width}{Width in pixels, defaults to automatic sizing.}

\item{height}{Height in pixels, defaults to automatic sizing.}

\item{force}{Whether to force drawing regardless of the number of features greater than 100.}
}
\description{
Plotting cell points on a reduced 3D space and coloring according to the gene expression in the cells.
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
FeatureDimPlot3D(pancreas_sub, features = c("Ghrl", "Ins1", "Gcg", "Ins2"), reduction = "StandardpcaUMAP3D")
}
