% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{FeatureHeatmap}
\alias{FeatureHeatmap}
\title{FeatureHeatmap}
\usage{
FeatureHeatmap(
  srt,
  features = NULL,
  cells = NULL,
  group.by = NULL,
  split.by = NULL,
  max_cells = 100,
  cell_order = NULL,
  border = TRUE,
  flip = FALSE,
  slot = "counts",
  assay = NULL,
  exp_method = c("zscore", "raw", "fc", "log2fc", "log1p"),
  limits = NULL,
  lib_normalize = identical(slot, "counts"),
  libsize = NULL,
  feature_split = NULL,
  feature_split_by = NULL,
  n_split = NULL,
  split_method = c("kmeans", "hclust", "mfuzz"),
  decreasing = FALSE,
  fuzzification = NULL,
  show_fuzzification = FALSE,
  cluster_features_by = NULL,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  cluster_row_slices = FALSE,
  cluster_column_slices = FALSE,
  show_row_names = FALSE,
  show_column_names = FALSE,
  row_names_side = ifelse(flip, "left", "right"),
  column_names_side = ifelse(flip, "bottom", "top"),
  row_names_rot = 0,
  column_names_rot = 90,
  row_title_side = "left",
  column_title_side = "top",
  row_title_rot = 0,
  column_title_rot = ifelse(flip, 90, 0),
  anno_terms = FALSE,
  anno_keys = FALSE,
  anno_features = FALSE,
  terms_width = unit(4, "in"),
  terms_fontsize = 8,
  keys_width = unit(2, "in"),
  keys_fontsize = c(6, 10),
  features_width = unit(2, "in"),
  features_fontsize = c(6, 10),
  IDtype = "symbol",
  species = "Homo_sapiens",
  db_update = FALSE,
  db_version = "latest",
  db_combine = FALSE,
  convert_species = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  db = "GO_BP",
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  GO_simplify = FALSE,
  GO_simplify_cutoff = "p.adjust < 0.05",
  simplify_method = "Wang",
  simplify_similarityCutoff = 0.7,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  topTerm = 5,
  show_termid = FALSE,
  topWord = 20,
  min_word_length = 3,
  exclude_words = c("cell", "cellular", "dna", "rna", "protein", "development",
    "organization", "system", "regulation", "positive", "negative", "response",
    "process"),
  nlabel = 20,
  features_label = NULL,
  label_size = 10,
  label_color = "black",
  heatmap_palette = "RdBu",
  heatmap_palcolor = NULL,
  group_palette = "Paired",
  group_palcolor = NULL,
  cell_split_palette = "jama",
  cell_split_palcolor = NULL,
  feature_split_palette = "jama",
  feature_split_palcolor = NULL,
  cell_annotation = NULL,
  cell_annotation_palette = "Paired",
  cell_annotation_palcolor = NULL,
  cell_annotation_params = list(),
  feature_annotation = NULL,
  feature_annotation_palette = "Dark2",
  feature_annotation_palcolor = NULL,
  feature_annotation_params = list(),
  use_raster = NULL,
  raster_device = "png",
  raster_by_magick = FALSE,
  height = NULL,
  width = NULL,
  units = "inch",
  seed = 11,
  ht_params = list()
)
}
\arguments{
\item{db_combine}{}
}
\description{
FeatureHeatmap
}
\examples{
library(dplyr)
data("pancreas_sub")
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType")
de_filter <- filter(pancreas_sub@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
ht1 <- FeatureHeatmap(
  srt = pancreas_sub, features = de_filter$gene, group.by = "CellType",
  split.by = "Phase", cell_split_palette = "Dark2",
)
ht1$plot
panel_fix(ht1$plot, height = 4, width = 6, raster = TRUE, dpi = 50)

ht2 <- FeatureHeatmap(
  srt = pancreas_sub, features = de_filter$gene, group.by = c("CellType", "SubCellType"), n_split = 4,
  cluster_rows = TRUE, cluster_row_slices = TRUE, cluster_columns = TRUE, cluster_column_slices = TRUE,
  ht_params = list(row_gap = grid::unit(0, "mm"))
)
ht2$plot

ht3 <- FeatureHeatmap(
  srt = pancreas_sub, features = de_filter$gene, feature_split = de_filter$group1, group.by = "CellType",
  species = "Mus_musculus", db = "GO_BP", anno_terms = TRUE, anno_keys = TRUE, anno_features = TRUE
)
ht3$plot

pancreas_sub <- AnnotateFeatures(pancreas_sub, species = "Mus_musculus", db = c("TF", "SP"))
ht4 <- FeatureHeatmap(
  srt = pancreas_sub, features = de_filter$gene, n_split = 4, group.by = "CellType",
  heatmap_palette = "viridis",
  feature_annotation = c("TF", "SP"),
  feature_annotation_palcolor = list(c("gold", "steelblue"), c("forestgreen")),
  cell_annotation = c("Phase", "G2M_score"), cell_annotation_palette = c("Dark2", "Purples")
)
ht4$plot

ht5 <- FeatureHeatmap(
  srt = pancreas_sub, features = de_filter$gene, n_split = 4, group.by = "CellType",
  heatmap_palette = "viridis",
  feature_annotation = c("TF", "SP"),
  feature_annotation_palcolor = list(c("gold", "steelblue"), c("forestgreen")),
  cell_annotation = c("Phase", "G2M_score"), cell_annotation_palette = c("Dark2", "Purples"),
  flip = TRUE, column_title_rot = 45
)
ht5$plot

pancreas_sub <- RunPAGA(
  srt = pancreas_sub, assay_X = "RNA", group_by = "SubCellType",
  linear_reduction = "PCA", nonlinear_reduction = "UMAP", infer_pseudotime = TRUE, root_group = "Ductal"
)
ht6 <- FeatureHeatmap(
  srt = pancreas_sub, features = de_filter$gene, nlabel = 10,
  cell_order = names(sort(pancreas_sub$dpt_pseudotime)),
  cell_annotation = c("CellType", "dpt_pseudotime"),
  cell_annotation_palette = c("Paired", "cividis")
)
ht6$plot

}
