% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-app.R
\name{FetchH5}
\alias{FetchH5}
\title{Fetch data from the hdf5 file}
\usage{
FetchH5(
  DataFile,
  MetaFile,
  name = NULL,
  features = NULL,
  slot = NULL,
  assay = NULL,
  metanames = NULL,
  reduction = NULL
)
}
\arguments{
\item{reduction}{}
}
\description{
Fetch data from the hdf5 file
}
\examples{
\dontrun{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
PrepareSCExplorer(pancreas_sub, base_dir = "./SCExplorer")
srt <- FetchH5(DataFile = "./SCExplorer/Data.hdf5", MetaFile = "./SCExplorer/Meta.hdf5", features = c("Ins1", "Ghrl"), metanames = c("SubCellType", "Phase"), reduction = "UMAP")
CellDimPlot(srt, group.by = c("SubCellType", "Phase"), reduction = "UMAP")
FeatureDimPlot(srt, features = c("Ins1", "Ghrl"), reduction = "UMAP")
}
}
