% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-app.R
\name{FetchH5}
\alias{FetchH5}
\title{Fetch data from the hdf5 file}
\usage{
FetchH5(
  DataFile,
  MetaFile,
  name = NULL,
  features = NULL,
  slot = NULL,
  assay = NULL,
  metanames = NULL,
  reduction = NULL
)
}
\arguments{
\item{DataFile}{The path to the hdf5 file containing the data.}

\item{MetaFile}{The path to the hdf5 file containing the metadata.}

\item{name}{The name of the dataset in the hdf5 file. If not specified, the function will attempt to find the shared group name in both files.}

\item{features}{The names of the genes or features to fetch. If specified, only these features will be fetched.}

\item{slot}{The slot for the counts in the hdf5 file. If not specified, the first slot will be used.}

\item{assay}{The name of the assay to use. If not specified, the default assay in the hdf5 file will be used.}

\item{metanames}{The names of the metadata columns to fetch.}

\item{reduction}{The name of the reduction to fetch.}
}
\value{
A Seurat object with the fetched data.
}
\description{
This function fetches data from an hdf5 file. It can fetch gene expression data, metadata, and reduction data from the specified file and returns a Seurat object.
}
\examples{
\dontrun{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
PrepareSCExplorer(pancreas_sub, base_dir = "./SCExplorer")
srt <- FetchH5(
  DataFile = "./SCExplorer/Data.hdf5",
  MetaFile = "./SCExplorer/Meta.hdf5",
  features = c("Ins1", "Ghrl"),
  metanames = c("SubCellType", "Phase"),
  reduction = "UMAP"
)
CellDimPlot(srt, group.by = c("SubCellType", "Phase"), reduction = "UMAP")
FeatureDimPlot(srt, features = c("Ins1", "Ghrl"), reduction = "UMAP")
}
}
\seealso{
\code{\link{CreateDataFile}} \code{\link{CreateMetaFile}} \code{\link{PrepareSCExplorer}} \code{\link{RunSCExplorer}}
}
