% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{fitSinCurve}
\alias{fitSinCurve}
\title{Differential Circadian Analysis for PowerSim}
\usage{
fitSinCurve(
  tt,
  yy,
  period = 24,
  parStart = list(amp = 3, phase = 0, offset = 0)
)
}
\arguments{
\item{tt}{Time vector}

\item{yy}{Expression vector}

\item{period}{Period of the sine curve (default 24)}

\item{parStart}{Initial value for optimization (default list(amp=3,phase=0,offset=0))}
}
\value{
A list with components:
  \item{amp}{Amplitude}
  \item{phase}{Phase (0 to period)}
  \item{offset}{Basal level}
  \item{peak}{Peak time}
  \item{A}{A coefficient (sin component)}
  \item{B}{B coefficient (cos component)}
  \item{tss}{Total sum of squares}
  \item{rss}{Residual sum of squares}
  \item{R2}{R-squared}
}
\description{
This file provides differential circadian rhythm analysis methods
Based on the DiffCircadian package implementation
}
\details{
Main Functions:
  DCP_Rhythmicity() - Single/two-group rhythmicity analysis with TOJR
  DCP_DiffR2()      - Differential rhythmicity test (DR)
  DCP_DiffPar()     - Differential parameter tests (DP, DA)
  DCP_Analyze()     - Complete differential analysis workflow

Author: Based on DiffCircadian package
Date: 2025-02-12
Fit Sinusoidal Curve

Fit a sine curve where tt is time, and yy is expression value.
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24)
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/24 * (tt + Phase)) + Offset + rnorm(n,0,1)
fitSinCurve(tt, yy)

}
