% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theta.R
\name{get_theta}
\alias{get_theta}
\title{Estimate theta via variogram fitting}
\usage{
get_theta(s, Y, dists = NULL)
}
\arguments{
\item{s}{spatial location}

\item{Y}{data}

\item{dists}{breakpoints for bins}

\item{plot_fitted}{will plot a emperical variogram if specified as TRUE}
}
\value{
Matern covariance parameters
}
\description{
Estimate theta via variogram fitting
}
\keyword{internal}
