% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{GroupHeatmap}
\alias{GroupHeatmap}
\title{GroupHeatmap}
\usage{
GroupHeatmap(
  srt,
  features = NULL,
  group.by = NULL,
  split.by = NULL,
  grouping.var = NULL,
  numerator = NULL,
  cells = NULL,
  aggregate_fun = base::mean,
  exp_cutoff = 0,
  border = TRUE,
  flip = FALSE,
  slot = "counts",
  assay = NULL,
  exp_method = c("zscore", "raw", "fc", "log2fc", "log1p"),
  limits = NULL,
  lib_normalize = identical(slot, "counts"),
  libsize = NULL,
  feature_split = NULL,
  feature_split_by = NULL,
  n_split = NULL,
  split_method = c("kmeans", "hclust", "mfuzz"),
  decreasing = FALSE,
  fuzzification = NULL,
  show_fuzzification = FALSE,
  cluster_features_by = NULL,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  cluster_row_slices = FALSE,
  cluster_column_slices = FALSE,
  show_row_names = FALSE,
  show_column_names = FALSE,
  row_names_side = ifelse(flip, "left", "right"),
  column_names_side = ifelse(flip, "bottom", "top"),
  row_names_rot = 0,
  column_names_rot = 90,
  row_title_side = "left",
  column_title_side = "top",
  row_title_rot = 0,
  column_title_rot = ifelse(flip, 90, 0),
  anno_terms = FALSE,
  anno_keys = FALSE,
  anno_features = FALSE,
  terms_width = unit(4, "in"),
  terms_fontsize = 8,
  keys_width = unit(2, "in"),
  keys_fontsize = c(6, 10),
  features_width = unit(2, "in"),
  features_fontsize = c(6, 10),
  IDtype = "symbol",
  species = "Homo_sapiens",
  db_update = FALSE,
  db_version = "latest",
  db_combine = FALSE,
  convert_species = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  db = "GO_BP",
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  GO_simplify = FALSE,
  GO_simplify_cutoff = "p.adjust < 0.05",
  simplify_method = "Wang",
  simplify_similarityCutoff = 0.7,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  topTerm = 5,
  show_termid = FALSE,
  topWord = 20,
  min_word_length = 3,
  exclude_words = c("cell", "cellular", "dna", "rna", "protein", "development",
    "organization", "system", "regulation", "positive", "negative", "response",
    "process"),
  nlabel = 0,
  features_label = NULL,
  label_size = 10,
  label_color = "black",
  add_bg = FALSE,
  bg_alpha = 0.5,
  add_dot = FALSE,
  dot_size = unit(8, "mm"),
  add_reticle = FALSE,
  reticle_color = "grey",
  add_violin = FALSE,
  fill.by = "feature",
  fill_palette = "Dark2",
  fill_palcolor = NULL,
  heatmap_palette = "RdBu",
  heatmap_palcolor = NULL,
  group_palette = "Paired",
  group_palcolor = NULL,
  cell_split_palette = "jama",
  cell_split_palcolor = NULL,
  feature_split_palette = "jama",
  feature_split_palcolor = NULL,
  cell_annotation = NULL,
  cell_annotation_palette = "Paired",
  cell_annotation_palcolor = NULL,
  cell_annotation_params = if (flip) list(width = grid::unit(1, "cm")) else list(height =
    grid::unit(1, "cm")),
  feature_annotation = NULL,
  feature_annotation_palette = "Dark2",
  feature_annotation_palcolor = NULL,
  feature_annotation_params = list(),
  use_raster = NULL,
  raster_device = "png",
  raster_by_magick = FALSE,
  height = NULL,
  width = NULL,
  units = "inch",
  seed = 11,
  ht_params = list()
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{features}{A vector of gene names to plot.}

\item{group.by}{Columns used to calculate cell expression. One heatmap per column name.}

\item{assay}{Assay used to calculate the expression.}

\item{exp_method}{Method used to calculate cell expression.}

\item{feature_split}{A vector of group names for features.}

\item{db_combine}{}

\item{heatmap_palette}{Heatmap expression palette.}

\item{cell_annotation_palette}{Column palette.}

\item{feature_annotation_palette}{Feature groups palette.}
}
\description{
GroupHeatmap
}
\examples{
library(dplyr)
data("pancreas_sub")
ht1 <- GroupHeatmap(pancreas_sub,
  features = c(
    "Sox9", "Anxa2", "Bicc1", # Ductal
    "Neurog3", "Hes6", # EPs
    "Fev", "Neurod1", # Pre-endocrine
    "Rbp4", "Pyy", # Endocrine
    "Ins1", "Gcg", "Sst", "Ghrl" # Beta, Alpha, Delta, Epsilon
  ),
  group.by = c("CellType", "SubCellType"),
  show_row_names = TRUE
)
ht1$plot
panel_fix(ht1$plot, height = 3, width = 6, raster = TRUE, dpi = 50)

pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType")
de_filter <- filter(pancreas_sub@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)

ht2 <- GroupHeatmap(
  srt = pancreas_sub, features = de_filter$gene, group.by = "CellType",
  split.by = "Phase", cell_split_palette = "Dark2",
  nlabel = 10, show_row_names = FALSE
)
ht2$plot

ht3 <- GroupHeatmap(
  srt = pancreas_sub, features = de_filter$gene, feature_split = de_filter$group1, group.by = "CellType",
  nlabel = 20, show_row_names = FALSE,
  species = "Mus_musculus", db = "GO_BP", anno_terms = TRUE, anno_keys = TRUE, anno_features = TRUE
)
ht3$plot

pancreas_sub <- AnnotateFeatures(pancreas_sub, species = "Mus_musculus", db = c("TF", "SP"))
de_top <- de_filter \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)
ht4 <- GroupHeatmap(pancreas_sub,
  features = de_top$gene, feature_split = de_top$group1, group.by = "CellType",
  heatmap_palette = "YlOrRd",
  cell_annotation = c("Phase", "G2M_score", "Neurod2"), cell_annotation_palette = c("Dark2", "Paired", "Paired"),
  cell_annotation_params = list(height = grid::unit(0.5, "in")),
  feature_annotation = c("TF", "SP"),
  feature_annotation_palcolor = list(c("gold", "steelblue"), c("forestgreen")),
  add_dot = TRUE, add_bg = TRUE, show_row_names = TRUE
)
ht4$plot

ht5 <- GroupHeatmap(pancreas_sub,
  features = de_top$gene, feature_split = de_top$group1, group.by = "CellType",
  heatmap_palette = "YlOrRd",
  cell_annotation = c("Phase", "G2M_score", "Neurod2"), cell_annotation_palette = c("Dark2", "Paired", "Paired"),
  cell_annotation_params = list(width = grid::unit(0.5, "in")),
  feature_annotation = c("TF", "SP"),
  feature_annotation_palcolor = list(c("gold", "steelblue"), c("forestgreen")),
  add_dot = TRUE, add_bg = TRUE,
  flip = TRUE, column_title_rot = 45, show_column_names = TRUE
)
ht5$plot

ht6 <- GroupHeatmap(pancreas_sub,
  features = de_top$gene, feature_split = de_top$group1, group.by = "CellType",
  add_violin = TRUE, cluster_rows = TRUE,
  show_row_names = TRUE
)
ht6$plot

ht7 <- GroupHeatmap(pancreas_sub,
  features = de_top$gene, feature_split = de_top$group1, group.by = "CellType",
  add_violin = TRUE, fill.by = "expression", fill_palette = "Blues", cluster_rows = TRUE,
  show_row_names = TRUE
)
ht7$plot

ht8 <- GroupHeatmap(pancreas_sub,
  features = de_top$gene, group.by = "CellType", split.by = "Phase", n_split = 4,
  cluster_rows = TRUE, cluster_columns = TRUE, cluster_row_slices = TRUE, cluster_column_slices = TRUE,
  add_dot = TRUE, add_reticle = TRUE, heatmap_palette = "viridis",
  show_row_names = TRUE, ht_params = list(row_gap = grid::unit(0, "mm"), row_names_gp = grid::gpar(fontsize = 10))
)
ht8$plot

}
