% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{GSEAPlot}
\alias{GSEAPlot}
\title{GSEA Plot}
\usage{
GSEAPlot(
  srt,
  group_by = NULL,
  test.use = "wilcox",
  group_use = NULL,
  enrichment = "GO_BP",
  res = NULL,
  geneSetID = NULL,
  title = NULL,
  subtitle = NULL,
  base_size = 12,
  rel_heights = c(1.5, 0.5, 1),
  subplots = 1:3,
  ES_geom = c("line", "dot"),
  color = "#6BB82D",
  n_coregene = 10,
  sample_coregene = FALSE,
  label_gene = NULL,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  label.size = 4
)
}
\arguments{
\item{term_use}{}
}
\description{
GSEA Plot
}
\examples{
if (interactive()) {
  data("pancreas1k")
  pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType", only.pos = FALSE, fc.threshold = 1)
  pancreas1k <- RunGSEA(pancreas1k, group_by = "CellType", enrichment = "GO_BP", species = "Mus_musculus")
  GSEAPlot(pancreas1k, group_by = "CellType", group_use = "Ductal", enrichment = "GO_BP", geneSetID = "GO:0006412")
  GSEAPlot(pancreas1k, group_by = "CellType", group_use = "Endocrine", enrichment = "GO_BP", geneSetID = "GO:0007186")
}
}
