% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{GSEAPlot}
\alias{GSEAPlot}
\title{GSEA Plot}
\usage{
GSEAPlot(
  x,
  geneSetID,
  title = NULL,
  subtitle = NULL,
  base_size = 12,
  rel_heights = c(1.5, 0.5, 1),
  subplots = 1:3,
  ES_geom = c("line", "dot"),
  color = "#6BB82D",
  n_coregene = 10,
  sample_coregene = FALSE,
  label_gene = NULL,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  label.size = 4
)
}
\description{
GSEA Plot
}
\examples{
if (interactive()) {
  data("pancreas1k")
  library(dplyr)
  pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType", only.pos = FALSE, fc.threshold = 1)
  de_filter <- filter(pancreas1k@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05)
  res <- RunGSEA(geneID = de_filter$gene, geneScore = de_filter$avg_log2FC, geneID_groups = de_filter$group1, species = "Mus_musculus")
  GSEAPlot(x = res$results[[1]], geneSetID = res$results[[1]]@result$ID[1])
}
}
