% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{GSEAPlot}
\alias{GSEAPlot}
\title{GSEA Plot}
\usage{
GSEAPlot(
  srt,
  db = "GO_BP",
  group_by = NULL,
  group_use = NULL,
  test.use = "wilcox",
  res = NULL,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  topTerm = 6,
  geneSetID = NULL,
  ES_geom = c("line", "dot"),
  subplots = 1:3,
  rel_heights = c(1.5, 0.5, 1),
  rel_width = 3,
  size = 1.5,
  alpha = 1,
  color = "#6BB82D",
  palette = "Set1",
  n_coregene = 10,
  sample_coregene = FALSE,
  features_label = NULL,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  label.size = 4,
  base_size = 12,
  title = NULL,
  subtitle = NULL,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE
)
}
\arguments{
\item{term_use}{}
}
\description{
GSEA Plot
}
\examples{
library(dplyr)
data("pancreas_sub")
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType", only.pos = FALSE, fc.threshold = 1)
pancreas_sub <- RunGSEA(pancreas_sub, group_by = "CellType", db = "GO_BP", species = "Mus_musculus")
GSEAPlot(pancreas_sub, group_by = "CellType", group_use = c("Ngn3 low EP", "Endocrine"))
GSEAPlot(pancreas_sub, group_by = "CellType", group_use = "Ductal", geneSetID = "GO:0006412")
GSEAPlot(pancreas_sub, group_by = "CellType", group_use = "Endocrine", geneSetID = c("GO:0046903", "GO:0015031", "GO:0007600")) \%>\%
  panel_fix_single(width = 10)
}
