% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{GSEAPlot}
\alias{GSEAPlot}
\title{GSEA Plot}
\usage{
GSEAPlot(
  srt,
  db = "GO_BP",
  group_by = NULL,
  group_use = NULL,
  test.use = "wilcox",
  res = NULL,
  pvalueCutoff = NULL,
  padjustCutoff = 0.05,
  plot_type = c("line", "comparison"),
  palette = "Spectral",
  palcolor = NULL,
  topTerm = 6,
  geneSetID = NULL,
  only_pos = FALSE,
  only_sig = FALSE,
  subplots = 1:3,
  rel_heights = c(1.5, 0.5, 1),
  rel_width = 3,
  line_size = 1.5,
  line_alpha = 1,
  line_color = "#6BB82D",
  n_coregene = 10,
  sample_coregene = FALSE,
  features_label = NULL,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  label.size = 4,
  aspect.ratio = NULL,
  base_size = 12,
  character_width = 50,
  lineheight = 0.7,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  seed = 11
)
}
\arguments{
\item{group_use}{}
}
\description{
GSEA Plot
}
\examples{
library(dplyr)
data("pancreas_sub")
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType", only.pos = FALSE, fc.threshold = 1)
pancreas_sub <- RunGSEA(pancreas_sub, group_by = "CellType", db = "GO_BP", species = "Mus_musculus")
GSEAPlot(pancreas_sub, db = "GO_BP", group_by = "CellType", group_use = "Ductal")
GSEAPlot(pancreas_sub, db = "GO_BP", group_by = "CellType", group_use = "Ductal", geneSetID = "GO:0006412")
GSEAPlot(pancreas_sub, db = "GO_BP", group_by = "CellType", group_use = "Endocrine", geneSetID = c("GO:0046903", "GO:0015031", "GO:0007600")) \%>\%
  panel_fix_single(width = 5) # Because the plot is made by combining, we want to adjust the overall height and width
GSEAPlot(pancreas_sub, db = "GO_BP", group_by = "CellType", plot_type = "comparison")
GSEAPlot(pancreas_sub, db = "GO_BP", group_by = "CellType", plot_type = "comparison", only_sig = TRUE)
GSEAPlot(pancreas_sub, db = "GO_BP", group_by = "CellType", plot_type = "comparison", pvalueCutoff = 0.05, padjustCutoff = NULL, only_pos = TRUE, only_sig = TRUE)
}
