% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_pred.R
\name{krige_pred}
\alias{krige_pred}
\title{Kriging prediction function}
\usage{
krige_pred(s0, s, Y, return_sd = FALSE)
}
\arguments{
\item{s0}{prediction location, a numeric vector with \code{length = 2}.}

\item{s}{an \eqn{n \times 2}{n x 2} \code{matrix} or \code{data.frame} with two coordinates of \eqn{n} locations.}

\item{Y}{a vector with \eqn{n} values corresponding to \code{Y(s)}.}

\item{return_sd}{logical. if \code{TRUE}, \code{Krige_pred} function returns the
standard error of \code{Y(s0)} along with the point prediction;
if \code{FALSE}, \code{Krige_pred} function only returns the point prediction. Defaults to \code{FALSE}.}
}
\value{
The output is a value of point prediction for \code{Y(s0)} if \code{return_sd} is FALSE
or a \code{list} with the following elements if \code{return_sd} is TRUE.
   \item{yhat }{point prediction for \code{Y(s0)}}
   \item{sd }{standard error for \code{Y(s0)}}
}
\description{
This function provides an example for \code{pred_fun} in \code{\link{scp}},
\code{\link{plausibility}}, and \code{\link{plausibility_contour}}, which provides a point prediction
for location \code{s0} (and corresponding standard error), given obserations \code{s} and \code{Y}.
}
\examples{
N = 21; n = N^2
S = seq(0,1,length=N)
s = expand.grid(S,S)
d = as.matrix(dist(s))

theta        = c(0,3,0.1,0.7)
names(theta) = c("Nugget","PartialSill","Range","Smoothness")
C = mat_cov(d,theta)
X = t(chol(C))\%*\%rnorm(n)
Y = X^3 + rnorm(n)

s0 = c(0.5, 0.5)
krige_pred(s0,s,Y)
krige_pred(s0,s,Y,interval=TRUE)
}
