% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{LineagePlot}
\alias{LineagePlot}
\title{LineagePlot}
\usage{
LineagePlot(
  srt,
  lineages,
  reduction = NULL,
  dims = c(1, 2),
  cells = NULL,
  trim = c(0.01, 0.99),
  span = 0.75,
  palette = "Dark2",
  palcolor = NULL,
  lineages_arrow = arrow(length = unit(0.1, "inches")),
  line_size = 1,
  line_bg = "white",
  line_bg_r = 0.5,
  whiskers = FALSE,
  whiskers_size = 0.5,
  whiskers_alpha = 0.5,
  aspect.ratio = 1,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend.position = "right",
  legend.direction = "vertical",
  theme_use = "theme_scp",
  theme_args = list(),
  return_layer = FALSE,
  seed = 11
)
}
\arguments{
\item{cells}{}
}
\description{
LineagePlot
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP", show_plot = FALSE)
LineagePlot(pancreas_sub, lineages = paste0("Lineage", 1:3))
LineagePlot(pancreas_sub, lineages = paste0("Lineage", 1:3), whiskers = TRUE)
}
