% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{ListDB}
\alias{ListDB}
\title{ListDB}
\usage{
ListDB(
  species = c("Homo_sapiens", "Mus_musculus"),
  db = c("GO", "GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway", "Reactome",
    "ProteinComplex", "DGI", "MP", "DO", "HPO", "PFAM", "Chromosome", "GeneType",
    "Enzyme", "TF", "CSPA", "Surfaceome", "SPRomeDB", "VerSeDa", "TFLink", "hTFtarget",
    "TRRUST", "JASPAR", "ENCODE", "CellTalk", "CellChat")
)
}
\arguments{
\item{db}{}
}
\description{
ListDB
}
\examples{
ListDB(species = "Homo_sapiens")

}
