% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{ListDB}
\alias{ListDB}
\title{ListDB}
\usage{
ListDB(species = "Homo_sapiens", db = NULL)
}
\arguments{
\item{species}{The species for which to retrieve database information. Default is "Homo_sapiens".}

\item{db}{The pattern to match against the database names. Default is NULL, which matches all databases.}
}
\description{
Retrieves information about databases based on a given species and database name.
}
\examples{
ListDB(species = "Homo_sapiens")
ListDB(species = "Mus_musculus", db = "GO_BP")

}
\seealso{
\code{\link{PrepareDB}}
}
