% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_cov.R
\name{mat_cov}
\alias{mat_cov}
\title{Matern covariance function}
\usage{
mat_cov(d, theta)
}
\arguments{
\item{d}{distances}

\item{theta}{covariance parameters}
}
\value{
a covariance matrix
}
\description{
Matern covariance function
}
