% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergeRows}
\alias{MergeRows}
\title{Merge Rows of a Matrix by Groupings}
\usage{
MergeRows(x, groupings)
}
\arguments{
\item{x}{A matrix.}

\item{groupings}{A vector specifying the groupings for merging rows.}
}
\description{
MergeRows is a function that sum the values of rows based on specified groupings.
}
\examples{
library(Matrix)
ncells <- 3000
nfeatures <- 1000
expressed <- 500
n <- ncells * expressed
dimnames <- list(paste0("feature", seq_len(nfeatures)), paste0("cell", seq_len(ncells)))

sparse <- sparseMatrix(
  i = sample(seq_len(nfeatures), size = n, replace = TRUE),
  j = sample(seq_len(ncells), size = n, replace = TRUE),
  x = sample(1:10, size = n, replace = TRUE),
  dimnames = dimnames
)
dup_index <- sample(seq_len(nfeatures), size = 50)
rownames(sparse)[dup_index] <- paste0("duplicated", 1:10)
table(rownames(sparse)[dup_index])

dense <- matrix(
  data = sample(1:10, size = nfeatures * ncells, replace = TRUE),
  nrow = nfeatures, ncol = ncells,
  dimnames = dimnames
)
dup_index <- sample(seq_len(nfeatures), size = 50)
rownames(dense)[dup_index] <- paste0("duplicated", 1:10)
table(rownames(dense)[dup_index])

system.time(MergeRows(dense, rownames(dense)))
system.time(MergeRows(sparse, rownames(sparse)))
system.time(aggregate(dense, by = list(rownames(dense)), FUN = sum))
}
