% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{palette_scp}
\alias{palette_scp}
\title{SCP palette}
\usage{
palette_scp(
  x,
  n = 100,
  palette = "Paired",
  palcolor = NULL,
  type = "auto",
  matched = FALSE,
  reverse = FALSE,
  NA_keep = FALSE,
  NA_color = "grey80"
)
}
\arguments{
\item{x}{A vector of character/factor or numeric values for color mapping.}

\item{palette}{Palette name. All palette names can be queried with \code{names(SCP:::palette_list)}.}

\item{palcolor}{}

\item{type}{Type of \code{x}.Can be one of "auto","discrete" or "continuous". The default is "auto", which automatically detects if \code{x} is a numeric value.
If it is numeric, then \code{x} is interpolated to the palette color of length 100 in ascending order.
Otherwise, the color is assigned to \code{x} according to its type number.}

\item{matched}{Whether to return a matched color vector of length \code{x}}

\item{reverse}{Whether to invert the colors.}

\item{NA_keep}{Whether to keep the color assignment to NA in \code{x}.}

\item{NA_color}{NA colors if NA_keep is \code{TRUE}.}
}
\description{
The default palette for SCP plot function.
}
\examples{
x <- c(1:3, NA, 3:5)
(pal1 <- palette_scp(x, palette = "Spectral"))
(pal2 <- palette_scp(x, palette = "Spectral", matched = TRUE))
(pal3 <- palette_scp(x, palette = "Spectral", matched = TRUE, NA_keep = TRUE))
(pal4 <- palette_scp(x, palette = "Paired", type = "discrete"))
show_palettes(list(pal1, pal2, pal3, pal4))

all_palettes <- show_palettes(return_palettes = TRUE)
names(all_palettes)

}
\seealso{

}
