% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{palette_scp}
\alias{palette_scp}
\title{Color palettes collected in SCP.}
\usage{
palette_scp(
  x,
  n = 100,
  palette = "Paired",
  palcolor = NULL,
  type = "auto",
  matched = FALSE,
  reverse = FALSE,
  NA_keep = FALSE,
  NA_color = "grey80"
)
}
\arguments{
\item{x}{A vector of character/factor or numeric values. If missing, numeric values 1:n will be used as x.}

\item{n}{The number of colors to return for numeric values.}

\item{palette}{Palette name. All available palette names can be queried with \code{show_palettes()}.}

\item{palcolor}{Custom colors used to create a color palette.}

\item{type}{Type of \code{x}. Can be one of "auto", "discrete" or "continuous". The default is "auto", which automatically detects if \code{x} is a numeric value.}

\item{matched}{If \code{TRUE}, will return a color vector of the same length as \code{x}.}

\item{reverse}{Whether to invert the colors.}

\item{NA_keep}{Whether to keep the color assignment to NA in \code{x}.}

\item{NA_color}{Color assigned to NA if NA_keep is \code{TRUE}.}
}
\description{
Color palettes collected in SCP.
}
\examples{
x <- c(1:3, NA, 3:5)
(pal1 <- palette_scp(x, palette = "Spectral"))
(pal2 <- palette_scp(x, palcolor = c("red", "white", "blue")))
(pal3 <- palette_scp(x, palette = "Spectral", n = 10))
(pal4 <- palette_scp(x, palette = "Spectral", n = 10, reverse = TRUE))
(pal5 <- palette_scp(x, palette = "Spectral", matched = TRUE))
(pal6 <- palette_scp(x, palette = "Spectral", matched = TRUE, NA_keep = TRUE))
(pal7 <- palette_scp(x, palette = "Paired", type = "discrete"))
show_palettes(list(pal1, pal2, pal3, pal4, pal5, pal6, pal7))

all_palettes <- show_palettes(return_palettes = TRUE)
names(all_palettes)

}
\seealso{
\code{\link{show_palettes}} \code{\link{palette_list}}
}
