% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{panel_fix_single}
\alias{panel_fix_single}
\title{Set the panel width/height for the single plot.}
\usage{
panel_fix_single(
  x,
  panel_index = NULL,
  respect = NULL,
  width = NULL,
  height = NULL,
  margin = 0,
  units = "in",
  raster = FALSE,
  dpi = 300,
  BPPARAM = BiocParallel::bpparam(),
  return_grob = FALSE,
  bg_color = "white",
  save = NULL,
  device = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{}

\item{verbose}{}
}
\description{
Set the panel width/height for the single plot.
}
