% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_contour.R
\name{plausibility_contour}
\alias{plausibility_contour}
\title{generate plausibility contour}
\usage{
plausibility_contour(
  s0,
  s,
  Y,
  global = TRUE,
  eta = Inf,
  m = NULL,
  pred_fun = krige_pred,
  thetaHat = NULL,
  dfun = c("residual2", "std_residual2"),
  precision = NULL
)
}
\arguments{
\item{s0}{prediction location, a numeric vector with \code{length = 2}
or a \code{matrix} with 1 row and 2 cols,
or a data.frame with 1 row and 2 cordinates.}

\item{s}{an \eqn{n \times 2}{n x 2} \code{matrix} or a \code{data.frame} with two coordinates of \eqn{n} locations.}

\item{Y}{a vector with \eqn{n} values corresponding to \code{Y(s)}.}

\item{global}{logical; if \code{TRUE} , \code{scp} function returns the result of global spatial conformal prediction (GSCP);
if \code{FALSE}, \code{scp} function returns the result of local spatial conformal prediction (LSCP) and users need to
specify \code{eta < Inf} or \code{m} \eqn{\le} \code{n}. Defaults to \code{TRUE}.}

\item{eta}{kernel bandwidth for weight schema, a positve scalar with smaller value meaning more localized procedure.
Defauls to \code{Inf}, which puts equal weight on surrounding \code{m} points.}

\item{m}{an postive integer representing the number of nearest locations to use for prediction.
Default to \code{NULL}. If \code{global = TRUE}, \code{m = n};
if \code{global = FALSE} and \code{m} is not specified, \code{m} would be determined by \code{eta}.}

\item{pred_fun}{spatial prediction function with inputs being \code{s0, s, Y} and ouputs being predicted \code{Y(s0)}
(and its standard error). Defaults to \code{\link{krige_pred}}.}

\item{thetaHat}{a vector of Matern parameters, representing nugget, partial sill, range, and smoothness as in Mao. et al. (2020).
Defaults to \code{NULL}. It will be ignored if \code{pred_fun} is not \code{krige_pred}.}

\item{dfun}{non-conformity measure with four options.
In which, \code{"residual2"} (default) represents squared residual and
\code{"std_residual2"} represents standardized squared residual.}

\item{precision}{a positive scalar represents how dense \code{Y(s)} candidates (\code{Y_cand}) are. Defaults to \code{NULL}.}
}
\value{
The output is a ``plausibility_contour'' object.
}
\description{
This function provides the plausibility contour for \code{Y(s0)},
given observations \code{s} and \code{Y},
using spatial conformal prediction algorithms.
}
\examples{
## generate plausibility contour for Y(s0), where s0 = c(0.5,0.5), using sample data

#?sample_data
s0 = c(0.5,0.5)
s  = sample_data$s
Y  = sample_data$Y

pc = plausibility_contour(s0=s0,s=s,Y=Y)
plot(pc)

idx = which(s[, 1] == s0[1] & s[, 2] == s0[2])
abline(v = Y[idx], col = "red", lty = 2)
legend("topright", col=1:2, lty=1:2, c("plausibility", "true value"))

}
\references{
to be entered
}
\seealso{
\code{\link{plausibility}}
}
\author{
Huiying Mao, \email{hmao@samsi.info}, Brian Reich \email{bjreich@ncsu.edu}
}
