% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{PrepareDB}
\alias{PrepareDB}
\title{Prepare the database for enrichment analysis}
\usage{
PrepareDB(
  species = c("Homo_sapiens", "Mus_musculus"),
  db = c("GO", "GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway", "Reactome",
    "ProteinComplex", "DGI", "MP", "DO", "PFAM", "Chromosome", "GeneType", "Enzyme",
    "TF", "SP", "CellTalk", "CellChat"),
  db_IDtypes = c("symbol", "entrez_id", "ensembl_id"),
  db_version = "latest",
  db_update = FALSE,
  convert_species = TRUE,
  Ensembl_version = 103,
  mirror = NULL,
  custom_TERM2GENE = NULL,
  custom_TERM2NAME = NULL,
  custom_species = NULL,
  custom_IDtype = NULL,
  custom_version = NULL
)
}
\arguments{
\item{species}{species}

\item{db}{Enrichment database name.}

\item{db_update}{Whether update the database to the latest version.}

\item{custom_version}{}
}
\value{
A list containing the database.
}
\description{
Prepare the database for enrichment analysis
}
\examples{
db_list <- PrepareDB(species = "Homo_sapiens", db = "GO_BP", db_update = TRUE)
ListDB(species = "Homo_sapiens", db = "GO_BP")
head(db_list[["Homo_sapiens"]][["GO_BP"]][["TERM2GENE"]])

db_list <- PrepareDB(species = "Homo_sapiens", db = "MP", convert_species = TRUE, db_update = TRUE)
ListDB(species = "Homo_sapiens", db = "MP")
head(db_list[["Homo_sapiens"]][["MP"]][["TERM2GENE"]])

db_list <- PrepareDB(species = "Macaca_fascicularis", db = "GO_BP", convert_species = TRUE, db_update = TRUE)
ListDB(species = "Macaca_fascicularis", db = "GO_BP")
head(db_list[["Macaca_fascicularis"]][["GO_BP"]][["TERM2GENE"]])

ccgenes <- CC_GenePrefetch("Homo_sapiens")
custom_TERM2GENE <- rbind(
  data.frame(term = "S_genes", gene = ccgenes[["cc_S_genes"]]),
  data.frame(term = "G2M_genes", gene = ccgenes[["cc_G2M_genes"]])
)
str(custom_TERM2GENE)
db_list <- PrepareDB(
  species = c("Homo_sapiens", "Mus_musculus"), db = "CellCycle", convert_species = TRUE,
  custom_TERM2GENE = custom_TERM2GENE, custom_species = "Homo_sapiens", custom_IDtype = "symbol", custom_version = "Seurat_v4"
)
ListDB(db = "CellCycle")

db_list <- PrepareDB(species = "Mus_musculus", db = "CellCycle")
head(db_list[["Mus_musculus"]][["CellCycle"]][["TERM2GENE"]])

}
