% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PrepareEnv}
\alias{PrepareEnv}
\title{Prepare SCP python environment}
\usage{
PrepareEnv(
  conda_binary = NULL,
  python_version = "3.8",
  miniconda_repo = "https://repo.anaconda.com/miniconda",
  force = FALSE,
  ...
)
}
\arguments{
\item{python_version}{The version of python to install. Default is \code{3.8}}

\item{...}{}
}
\description{
Prepare SCP python environment
}
