% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-app.R
\name{PrepareSCExplorer}
\alias{PrepareSCExplorer}
\title{Prepare Seurat object data for the SCExplorer into hdf5 file}
\usage{
PrepareSCExplorer(
  object,
  base_dir = "SCExplorer",
  DataFile = "Data.hdf5",
  MetaFile = "Meta.hdf5",
  assays = "RNA",
  slots = c("counts", "data"),
  ignore_nlevel = 50,
  write_tools = FALSE,
  write_misc = FALSE,
  compression_level = 6,
  overwrite = FALSE
)
}
\description{
Prepare Seurat object data for the SCExplorer into hdf5 file
}
\examples{
data("pancreas1k")
pancreas1k <- Standard_SCP(pancreas1k)
PrepareSCExplorer(pancreas1k, base_dir = "./SCExplorer")
}
