% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.estimate_r_from_data}
\alias{print.estimate_r_from_data}
\title{Print method for estimate_r_from_data}
\usage{
\method{print}{estimate_r_from_data}(x, ...)
}
\arguments{
\item{x}{Result from estimate_r_from_data}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for estimate_r_from_data
}
