% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSCP.R
\name{readSCPfromDIANN}
\alias{readSCPfromDIANN}
\title{Read single-cell DIA-NN output as a QFeatures objects}
\usage{
readSCPfromDIANN(
  colAnnotation,
  assayData,
  extractedData = NULL,
  ecol = "MS1.Area",
  multiplexing = "none",
  ...
)
}
\arguments{
\item{colAnnotation}{A \code{data.frame} or any object that can be
coerced to a \code{data.frame}. \code{colAnnotation} is expected to
contains all the sample annotations. We require the table to
contain a column called \code{File.Name} that links to the
\code{File.Name} in the DIA-NN report table. If \code{multiplexing = "mTRAQ"}, we require a second column called \code{Label} that links
the label to the sample (the labels identified by DIA-NN can
be retrieved from \verb{Modified Sequence} column in the report
table).}

\item{assayData}{A \code{data.frame} or any object that can be coerced
to a data.frame that contains the data from the \code{Report.tsv}
file generated by DIA-NN.}

\item{extractedData}{A data.frame or any object that can be coerced
to a data.frame that contains the data from the \verb{*_ms1_extracted.tsv}
file generated by DIA-NN. This argument is optional and is
currently only applicable for mTRAQ multiplexed experiments
where DIA-NN was run using the \code{plexdia} module.}

\item{ecol}{A \code{character(1)} indicating which column in
\code{assayData} contains the quantitative information. Default is
\code{"Ms1.Area"}.}

\item{multiplexing}{A \code{character(1)} indicating the type of
multiplexing used in the experiment. Provide \code{"none"} if the
experiment is label-free (default). Alternative options are:
\code{"mTRAQ"}.}

\item{...}{Further arguments passed to \code{\link[QFeatures:readQFeatures]{readQFeatures()}}.}
}
\value{
An instance of class \code{QFeatures}. The expression data of
each acquisition run is stored in a separate set as a
\code{SingleCellExperiment} object.
}
\description{
This function takes the output tables from DIA-NN and converts
them into a QFeatures object using the scp framework.
}
