% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RecoverCounts}
\alias{RecoverCounts}
\title{Attempt to recover raw counts from the normalized matrix of the Seurat object.}
\usage{
RecoverCounts(
  srt,
  assay = "RNA",
  min_count = c(1, 2, 3),
  tolerance = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{min_count}{Minimum UMI count of genes.}

\item{tolerance}{When restoring the original count, the nCount of each cell is theoretically calculated as an integer.
However, due to decimal point preservation during normalization, the calculated nCount is usually a floating point number close to the integer.
The tolerance is its difference from the integer.}
}
\description{
Attempt to recover raw counts from the normalized matrix of the Seurat object.
}
\examples{
data("pancreas1k")
pancreas1k <- Seurat::NormalizeData(pancreas1k)
raw_counts <- pancreas1k@assays$RNA@counts
pancreas1k@assays$RNA@counts <- pancreas1k@assays$RNA@data
pancreas1k <- RecoverCounts(pancreas1k)
identical(raw_counts, pancreas1k@assays$RNA@counts)

}
