% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{\%>\%}
\alias{unit}
\alias{arrow}
\alias{gpar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{grid}{\code{\link[grid]{arrow}}, \code{\link[grid]{gpar}}, \code{\link[grid]{unit}}}

  \item{rlang}{\code{\link[rlang:op-null-default]{\%||\%}}}
}}

