% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RenameClusters}
\alias{RenameClusters}
\title{Rename clusters for the Seurat object}
\usage{
RenameClusters(
  srt,
  group.by,
  nameslist = list(),
  name = "newclusters",
  keep_levels = FALSE
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{group.by}{The old group used to rename cells.}

\item{nameslist}{A named list of new cluster value.}

\item{name}{The name of the new cluster stored in the Seurat object.}

\item{keep_levels}{If the old group is a factor, keep the order of the levels.}
}
\description{
Rename clusters for the Seurat object
}
\examples{
data("pancreas_sub")
levels(pancreas_sub@meta.data[["SubCellType"]])

# Rename all clusters
pancreas_sub <- RenameClusters(pancreas_sub, group.by = "SubCellType", nameslist = letters[1:8])
CellDimPlot(pancreas_sub, "newclusters")

# Rename specified clusters
pancreas_sub <- RenameClusters(pancreas_sub,
  group.by = "SubCellType",
  nameslist = list("a" = "Alpha", "b" = "Beta")
)
CellDimPlot(pancreas_sub, "newclusters")

# Merge and rename clusters
pancreas_sub <- RenameClusters(pancreas_sub,
  group.by = "SubCellType",
  nameslist = list("EndocrineClusters" = c("Alpha", "Beta", "Epsilon", "Delta")),
  name = "Merged", keep_levels = TRUE
)
CellDimPlot(pancreas_sub, "Merged")

}
