% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RenameFeatures}
\alias{RenameFeatures}
\title{Rename features for the Seurat object}
\usage{
RenameFeatures(srt, newnames = NULL, oldnames = NULL, assays = NULL)
}
\arguments{
\item{srt}{}

\item{assays}{}
}
\description{
Rename features for the Seurat object
}
