% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RenameFeatures}
\alias{RenameFeatures}
\title{Rename features for the Seurat object}
\usage{
RenameFeatures(srt, newnames = NULL, assays = NULL)
}
\arguments{
\item{srt}{}

\item{assays}{}
}
\description{
Rename features for the Seurat object
}
\examples{
data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(stringr::str_to_title(rownames(panc8_sub)))
panc8_rename <- RenameFeatures(panc8_sub, newnames = genenames)
head(rownames(panc8_rename))

}
