% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowDataToDF}
\alias{rowDataToDF}
\title{Extract the \code{rowData} of a \code{QFeatures} object to a \code{DataFrame}}
\usage{
rowDataToDF(obj, i, vars)
}
\arguments{
\item{obj}{A \code{QFeatures} object}

\item{i}{A \code{numeric()} or \code{character()} vector indicating from which assays
the \code{rowData} should be taken.}

\item{vars}{A \code{character()} vector indicating which variables from the
\code{rowData} should be extracted.}
}
\value{
A \code{DataFrame} object with the \code{rowData} row-binded over the required
assays.
}
\description{
The methods takes the \code{rowData} of one or more given assay in a \code{QFeatures}
object and combines the data in a single \code{DataFrame}.
}
\details{
Along with the reuired \code{rowData} an additional \code{.assay} variable is created
and holds the assay name from which the metadata was taken.
}
\examples{
## Extract the peptide length and sequence from the first 3 assays
rowDataToDF(scp1, i = 1:3, c("Length", "Sequence"))
}
