% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowDataToDF}
\alias{rowDataToDF}
\title{Extract the `rowData` of a `QFeatures` object to a `DataFrame`}
\usage{
rowDataToDF(obj, i, vars)
}
\arguments{
\item{obj}{A `QFeatures` object}

\item{i}{A `numeric()` or `character()` vector indicating from which assays 
the `rowData` should be taken.}

\item{vars}{A `character()` vector indicating which variables from the 
`rowData` should be extracted.}
}
\value{
A `DataFrame` object with the `rowData` row-binded over the required
    assays.
}
\description{
The methods takes the `rowData` of one or more given assay in a `QFeatures`
object and combines the data in a single `DataFrame`.
}
\details{
Along with the reuired `rowData` an additional `.assay` variable is created 
and holds the assay name from which the metadata was taken.
}
\examples{
TODO
}
