% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunCSSMap}
\alias{RunCSSMap}
\title{Single-cell reference mapping with CSS method}
\usage{
RunCSSMap(
  srt_query,
  srt_ref,
  ref_css = NULL,
  ref_umap = NULL,
  force = FALSE,
  ref_group = NULL,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{vote_fun}{}
}
\description{
Single-cell reference mapping with CSS method
}
\examples{
data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(stringr::str_to_title(rownames(panc8_sub)))
panc8_sub <- RenameFeatures(panc8_sub, newnames = genenames)
head(rownames(panc8_sub))

panc8_sub <- Integration_SCP(panc8_sub, batch = "tech", integration_method = "CSS")
ClassDimPlot(panc8_sub, "celltype")

# Projection
data("pancreas1k")
pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunCSSMap(srt_query = pancreas1k, srt_ref = panc8_sub, ref_css = "CSS", ref_umap = "CSSUMAP2D")
ProjectionPlot(srt_query = pancreas1k, srt_ref = panc8_sub, query_group = "SubCellType", ref_group = "celltype")

}
