% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDEtest}
\alias{RunDEtest}
\title{Differential gene test}
\usage{
RunDEtest(
  srt,
  group_by = NULL,
  group1 = NULL,
  group2 = NULL,
  cells1 = NULL,
  cells2 = NULL,
  FindAllMarkers = TRUE,
  FindPairedMarkers = FALSE,
  FindConservedMarkers = FALSE,
  grouping.var = NULL,
  meta.method = metap::maximump,
  test.use = "wilcox",
  only.pos = TRUE,
  fc.threshold = 1.5,
  base = 2,
  pseudocount.use = 1,
  mean.fxn = NULL,
  min.pct = 0.1,
  min.diff.pct = -Inf,
  max.cells.per.ident = Inf,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  norm.method = "LogNormalize",
  p.adjust.method = "bonferroni",
  slot = "data",
  assay = "RNA",
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE,
  force = FALSE,
  seed = 11,
  ...
)
}
\arguments{
\item{srt}{A \code{Seurat} object}

\item{group_by}{group_by}

\item{FindAllMarkers}{Whether to find markers across all cells.}

\item{FindPairedMarkers}{Whether to find markers between pairs of the cell groups.}

\item{fc.threshold}{fc.threshold}

\item{min.pct}{min.pct}

\item{p.adjust.method}{}

\item{force}{force}
}
\description{
Differential gene test
}
\examples{
library(dplyr)
data(pancreas_sub)
pancreas_sub <- Standard_SCP(pancreas_sub)
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType")

# Heatmap
de_filter <- filter(pancreas_sub@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
ht_result <- ExpHeatmap(pancreas_sub, features = de_filter$gene, feature_split = de_filter$group1, group.by = "CellType")
ht_result$plot

# Dot plot
de_top <- de_filter \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)
GroupHeatmap(pancreas_sub, features = de_top$gene, feature_split = de_top$group1, group.by = "CellType")

data("panc8_sub")
panc8_sub <- Integration_SCP(panc8_sub, batch = "tech", integration_method = "Seurat")
ClassDimPlot(panc8_sub, group.by = c("tech", "celltype"))
panc8_sub <- RunDEtest(panc8_sub, group_by = "celltype", FindConservedMarkers = TRUE, grouping.var = "tech")

# Heatmap

# Dot plot
conserved_filter <- filter(panc8_sub@tools$DEtest_celltype$ConservedMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
conserved_top <- conserved_filter \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)

ExpDimPlot(panc8_sub,
  features = conserved_top$gene[1], split.by = "tech",
  cells.highlight = colnames(panc8_sub)[panc8_sub$celltype == conserved_top$group1[1]]
)

}
