% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDEtest}
\alias{RunDEtest}
\title{Differential gene test}
\usage{
RunDEtest(
  srt,
  group_by = NULL,
  group1 = NULL,
  group2 = NULL,
  cells1 = NULL,
  cells2 = NULL,
  features = NULL,
  FindAllMarkers = TRUE,
  FindPairedMarkers = FALSE,
  FindConservedMarkers = FALSE,
  grouping.var = NULL,
  meta.method = metap::maximump,
  test.use = "wilcox",
  only.pos = TRUE,
  fc.threshold = 1.5,
  base = 2,
  pseudocount.use = 1,
  mean.fxn = NULL,
  min.pct = 0.1,
  min.diff.pct = -Inf,
  max.cells.per.ident = Inf,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  norm.method = "LogNormalize",
  p.adjust.method = "bonferroni",
  slot = "data",
  assay = "RNA",
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE,
  force = FALSE,
  seed = 11,
  ...
)
}
\arguments{
\item{srt}{A \code{Seurat} object}

\item{group_by}{group_by}

\item{FindAllMarkers}{Whether to find markers across all cells.}

\item{FindPairedMarkers}{Whether to find markers between pairs of the cell groups.}

\item{fc.threshold}{fc.threshold}

\item{min.pct}{min.pct}

\item{force}{force}

\item{seed}{}
}
\description{
Differential gene test
}
\examples{
library(dplyr)
data("pancreas_sub")
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType")

de_filter <- filter(pancreas_sub@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
ht <- ExpHeatmap(pancreas_sub, features = de_filter$gene, feature_split = de_filter$group1, group.by = "CellType")
ht$plot

de_top <- de_filter \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)
ht <- GroupHeatmap(pancreas_sub, features = de_top$gene, feature_split = de_top$group1, group.by = "CellType")
ht$plot

}
