% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDEtest}
\alias{RunDEtest}
\title{Differential gene test}
\usage{
RunDEtest(
  srt,
  group_by = NULL,
  group1 = NULL,
  group2 = NULL,
  cells1 = NULL,
  cells2 = NULL,
  features = NULL,
  FindAllMarkers = TRUE,
  FindPairedMarkers = FALSE,
  FindConservedMarkers = FALSE,
  grouping.var = NULL,
  meta.method = metap::maximump,
  test.use = "wilcox",
  only.pos = TRUE,
  fc.threshold = 1.5,
  base = 2,
  pseudocount.use = 1,
  mean.fxn = NULL,
  min.pct = 0.1,
  min.diff.pct = -Inf,
  max.cells.per.ident = Inf,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  norm.method = "LogNormalize",
  p.adjust.method = "bonferroni",
  slot = "data",
  assay = "RNA",
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE,
  force = FALSE,
  seed = 11,
  ...
)
}
\arguments{
\item{srt}{A \code{Seurat} object}

\item{group_by}{group_by}

\item{FindAllMarkers}{Whether to find markers across all cells.}

\item{FindPairedMarkers}{Whether to find markers between pairs of the cell groups.}

\item{fc.threshold}{fc.threshold}

\item{min.pct}{min.pct}

\item{force}{force}

\item{seed}{}
}
\description{
Differential gene test
}
\examples{
library(dplyr)
data("pancreas_sub")

pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType")
AllMarkers <- filter(pancreas_sub@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
table(AllMarkers$group1)
ht1 <- ExpHeatmap(pancreas_sub, features = AllMarkers$gene, feature_split = AllMarkers$group1, group.by = "CellType")
ht1$plot

TopMarkers <- AllMarkers \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)
ht2 <- GroupHeatmap(pancreas_sub, features = TopMarkers$gene, feature_split = TopMarkers$group1, group.by = "CellType")
ht2$plot

pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType", FindAllMarkers = FALSE, FindPairedMarkers = TRUE)
PairedMarkers <- filter(pancreas_sub@tools$DEtest_CellType$PairedMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
table(PairedMarkers$group1)
ht3 <- ExpHeatmap(pancreas_sub, features = PairedMarkers$gene, feature_split = PairedMarkers$group1, group.by = "CellType")
ht3$plot

ifnb_sub <- RunDEtest(ifnb_sub, group_by = "seurat_annotations", grouping.var = "stim", FindAllMarkers = FALSE, FindConservedMarkers = TRUE)
ConservedMarkers <- filter(ifnb_sub@tools$DEtest_seurat_annotations$ConservedMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
table(ConservedMarkers$group1)
ht4 <- ExpHeatmap(ifnb_sub,
  features = ConservedMarkers$gene, feature_split = ConservedMarkers$group1,
  group.by = "seurat_annotations", split.by = "stim"
)
ht4$plot

}
