% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDEtest}
\alias{RunDEtest}
\title{Differential gene test}
\usage{
RunDEtest(
  srt,
  group_by = NULL,
  group1 = NULL,
  group2 = NULL,
  cells1 = NULL,
  cells2 = NULL,
  FindAllMarkers = TRUE,
  FindPairedMarkers = FALSE,
  test.use = "wilcox",
  only.pos = TRUE,
  fc.threshold = 1.5,
  base = 2,
  min.pct = 0.1,
  min.diff.pct = -Inf,
  max.cells.per.ident = Inf,
  latent.vars = NULL,
  slot = "data",
  assay = "RNA",
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE,
  force = FALSE,
  ...
)
}
\arguments{
\item{srt}{A \code{Seurat} object}

\item{group_by}{group_by}

\item{FindAllMarkers}{Whether to find markers across all cells.}

\item{FindPairedMarkers}{Whether to find markers between pairs of the cell groups.}

\item{fc.threshold}{fc.threshold}

\item{min.pct}{min.pct}

\item{force}{force}

\item{...}{}

\item{cell_group1}{cell_group1}

\item{cell_group2}{cell_group2}
}
\description{
Differential gene test
}
\examples{
library(dplyr)
data(pancreas1k)
pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType")

# Heatmap
de_filter <- filter(pancreas1k@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
ExpHeatmap(pancreas1k, features = de_filter$gene, feature_split = de_filter$group1, cell_split_by = "CellType")

# Dot plot
de_top <- de_filter \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)
ExpDotPlot(pancreas1k, features = de_top$gene, feature_split = de_top$group1, cell_split_by = "CellType")
}
