% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDEtest}
\alias{RunDEtest}
\title{Differential gene test}
\usage{
RunDEtest(
  srt,
  group_by = NULL,
  group1 = NULL,
  group2 = NULL,
  cells1 = NULL,
  cells2 = NULL,
  features = NULL,
  markers_type = c("all", "paired", "conserved", "disturbed"),
  grouping.var = NULL,
  meta.method = c("maximump", "minimump", "wilkinsonp", "meanp", "sump", "votep"),
  test.use = "wilcox",
  only.pos = TRUE,
  fc.threshold = 1.5,
  base = 2,
  pseudocount.use = 1,
  mean.fxn = NULL,
  min.pct = 0.1,
  min.diff.pct = -Inf,
  max.cells.per.ident = Inf,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  norm.method = "LogNormalize",
  p.adjust.method = "bonferroni",
  slot = "data",
  assay = NULL,
  BPPARAM = BiocParallel::bpparam(),
  seed = 11,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{srt}{A \code{Seurat} object}

\item{group_by}{group_by}

\item{fc.threshold}{fc.threshold}

\item{min.pct}{min.pct}

\item{verbose}{}

\item{force}{force}
}
\description{
Differential gene test
}
\examples{
library(dplyr)
data("ifnb_sub")

ifnb_sub <- RunDEtest(ifnb_sub, group_by = "seurat_annotations")
AllMarkers <- filter(ifnb_sub@tools$DEtest_seurat_annotations$AllMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
table(AllMarkers$group1)
ht1 <- GroupHeatmap(ifnb_sub, features = AllMarkers$gene, feature_split = AllMarkers$group1, group.by = "seurat_annotations")
ht1$plot

TopMarkers <- AllMarkers \%>\%
  group_by(gene) \%>\%
  top_n(1, avg_log2FC) \%>\%
  group_by(group1) \%>\%
  top_n(3, avg_log2FC)
ht2 <- GroupHeatmap(ifnb_sub, features = TopMarkers$gene, feature_split = TopMarkers$group1, group.by = "seurat_annotations", show_row_names = TRUE)
ht2$plot

ifnb_sub <- RunDEtest(ifnb_sub, group_by = "seurat_annotations", markers_type = "paired")
PairedMarkers <- filter(ifnb_sub@tools$DEtest_seurat_annotations$PairedMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
table(PairedMarkers$group1)
ht3 <- GroupHeatmap(ifnb_sub, features = PairedMarkers$gene, feature_split = PairedMarkers$group1, group.by = "seurat_annotations")
ht3$plot

ifnb_sub <- RunDEtest(ifnb_sub, group_by = "seurat_annotations", grouping.var = "stim", markers_type = "conserved")
ConservedMarkers <- filter(ifnb_sub@tools$DEtest_seurat_annotations$ConservedMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1)
table(ConservedMarkers$group1)
ht4 <- GroupHeatmap(ifnb_sub,
  features = ConservedMarkers$gene, feature_split = ConservedMarkers$group1,
  group.by = "seurat_annotations", split.by = "stim"
)
ht4$plot

ifnb_sub <- RunDEtest(ifnb_sub, group_by = "seurat_annotations", grouping.var = "stim", markers_type = "disturbed")
DisturbedMarkers_STIM <- filter(ifnb_sub@tools$DEtest_seurat_annotations$DisturbedMarkers_wilcox, p_val_adj < 0.05 & avg_log2FC > 1 & var1 == "STIM")
table(DisturbedMarkers_STIM$group1)
ht5 <- GroupHeatmap(ifnb_sub,
  features = DisturbedMarkers_STIM$gene, feature_split = DisturbedMarkers_STIM$group1,
  group.by = "seurat_annotations", split.by = "stim"
)
ht5$plot

gene_specific <- names(which(table(DisturbedMarkers_STIM$gene) <= 2))
DisturbedMarkers_STIM_specific <- DisturbedMarkers_STIM[DisturbedMarkers_STIM$gene \%in\% gene_specific, ]
table(DisturbedMarkers_STIM_specific$group1)
ht6 <- GroupHeatmap(ifnb_sub,
  features = DisturbedMarkers_STIM_specific$gene, feature_split = DisturbedMarkers_STIM_specific$group1,
  group.by = "seurat_annotations", split.by = "stim"
)
ht6$plot

ht7 <- GroupHeatmap(ifnb_sub,
  features = DisturbedMarkers_STIM_specific$gene, feature_split = DisturbedMarkers_STIM_specific$group1,
  group.by = "seurat_annotations", grouping.var = "stim", numerator = "STIM", limits = c(-2, 2)
)
ht7$plot

ht8 <- GroupHeatmap(ifnb_sub,
  features = DisturbedMarkers_STIM_specific$gene, feature_split = DisturbedMarkers_STIM_specific$group1,
  group.by = "seurat_annotations", grouping.var = "stim", numerator = "STIM",
  slot = "data", aggregate_fun = function(x) mean(expm1(x)) + 1
)
ht8$plot

}
