% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{RunDimReduction}
\alias{RunDimReduction}
\title{RunDimReduction}
\usage{
RunDimReduction(
  srt,
  prefix = "",
  features = NULL,
  assay = NULL,
  slot = "data",
  linear_reduction = NULL,
  linear_reduction_dims = 100,
  linear_reduction_params = list(),
  force_linear_reduction = FALSE,
  nonlinear_reduction = NULL,
  nonlinear_reduction_dims = 2,
  reduction_use = NULL,
  reduction_dims = NULL,
  neighbor_use = NULL,
  graph_use = NULL,
  distance_use = NULL,
  nonlinear_reduction_params = list(),
  force_nonlinear_reduction = TRUE,
  verbose = TRUE,
  seed = 11
)
}
\arguments{
\item{linear_reduction}{"pca", "ica", "nmf", "mds", "glmpca"}

\item{nonlinear_reduction}{"umap", "umap-naive", "tsne", "dm", "phate", "pacmap", "trimap", "largevis"}

\item{force_nonlinear_reduction}{}
}
\description{
RunDimReduction
}
