% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunDM}
\alias{RunDM}
\alias{RunDM.Seurat}
\alias{RunDM.matrix}
\alias{RunDM.Neighbor}
\title{Run DiffusionMap(DM)}
\usage{
RunDM(object, ...)

\method{RunDM}{Seurat}(
  object,
  reduction = "pca",
  dims = 1:30,
  features = NULL,
  assay = NULL,
  slot = "data",
  graph = NULL,
  neighbor = NULL,
  ndcs = 2,
  sigma = "local",
  k = 30,
  dist.method = "euclidean",
  reduction.name = "dm",
  reduction.key = "DM_",
  seed.use = 42,
  verbose = TRUE,
  ...
)

\method{RunDM}{matrix}(
  object,
  ndcs = 2,
  sigma = "local",
  k = 30,
  dist.method = "euclidean",
  assay = NULL,
  slot = "data",
  reduction.key = "DM_",
  seed.use = 42,
  verbose = TRUE,
  ...
)

\method{RunDM}{Neighbor}(
  object,
  ndcs = 2,
  sigma = "local",
  k = 30,
  slot = "data",
  assay = NULL,
  reduction.key = "DM_",
  verbose = TRUE,
  seed.use = 42,
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}

\item{assay}{Name of Assay dm is being run on.}

\item{slot}{Name of slot dm is being run on.}

\item{graph}{}

\item{ndcs}{Total Number of DM to compute and store (2 by default).}

\item{reduction.name}{dimensional reduction name, 'dm' by default}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. "DM_" by default.}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}

\item{verbose}{Show a progressbar and other progress information}
}
\value{
Returns Seurat object with the DM calculation stored in the reductions slot.
}
\description{
Run a DM dimensionality reduction.
}
\concept{dimensional_reduction}
