% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunDM}
\alias{RunDM}
\alias{RunDM.Seurat}
\alias{RunDM.default}
\title{Run DM (diffusion map)}
\usage{
RunDM(object, ...)

\method{RunDM}{Seurat}(
  object,
  reduction = "pca",
  dims = 1:30,
  features = NULL,
  assay = NULL,
  slot = "data",
  ndcs = 2,
  sigma = "local",
  k = 30,
  dist.method = "euclidean",
  reduction.name = "dm",
  reduction.key = "DM_",
  verbose = TRUE,
  seed.use = 11,
  ...
)

\method{RunDM}{default}(
  object,
  assay = NULL,
  slot = "data",
  ndcs = 2,
  sigma = "local",
  k = 30,
  dist.method = "euclidean",
  reduction.key = "DM_",
  verbose = TRUE,
  seed.use = 11,
  ...
)
}
\arguments{
\item{object}{An object. This can be a Seurat object or a matrix-like object.}

\item{...}{Additional arguments to be passed to the \link[destiny]{DiffusionMap} function.}

\item{reduction}{A character string specifying the reduction to be used.}

\item{dims}{An integer vector specifying the dimensions to be used. Default is 1:30.}

\item{features}{A character vector specifying the features to be used. Default is NULL.}

\item{assay}{A character string specifying the assay to be used. Default is NULL.}

\item{slot}{A character string specifying the slot name to be used. Default is "data".}

\item{ndcs}{An integer specifying the number of diffusion components (dimensions) to be computed. Default is 2.}

\item{sigma}{A character string specifying the diffusion scale parameter of the Gaussian kernel. Currently supported values are "local" (default) and "global".}

\item{k}{An integer specifying the number of nearest neighbors to be used for the construction of the graph. Default is 30.}

\item{dist.method}{A character string specifying the distance metric to be used for the construction of the knn graph. Currently supported values are "euclidean" and "cosine". Default is "euclidean".}

\item{reduction.name}{A character string specifying the name of the reduction to be stored in the Seurat object. Default is "dm".}

\item{reduction.key}{A character string specifying the prefix for the column names of the basis vectors. Default is "DM_".}

\item{verbose}{A logical value indicating whether to print verbose output. Default is TRUE.}

\item{seed.use}{An integer specifying the random seed to be used. Default is 11.}
}
\description{
Run DM (diffusion map)
}
\examples{
pancreas_sub <- Seurat::FindVariableFeatures(pancreas_sub)
pancreas_sub <- RunDM(object = pancreas_sub, features = Seurat::VariableFeatures(pancreas_sub))
CellDimPlot(pancreas_sub, group.by = "CellType", reduction = "dm")

}
