% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunDynamicFeatures}
\alias{RunDynamicFeatures}
\title{RunDynamicFeatures}
\usage{
RunDynamicFeatures(
  srt,
  lineages,
  features = NULL,
  suffix = lineages,
  n_candidates = 1000,
  minfreq = 5,
  family = NULL,
  slot = "counts",
  assay = NULL,
  libsize = NULL,
  BPPARAM = BiocParallel::bpparam(),
  seed = 11
)
}
\arguments{
\item{seed}{}
}
\description{
RunDynamicFeatures
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP")
pancreas_sub <- RunDynamicFeatures(pancreas_sub, lineages = c("Lineage1", "Lineage2"), n_candidates = 200)
names(pancreas_sub@tools$DynamicFeatures_Lineage1)
head(pancreas_sub@tools$DynamicFeatures_Lineage1$DynamicFeatures)
ht <- DynamicHeatmap(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  cell_annotation = "SubCellType",
  n_split = 6, reverse_ht = "Lineage1"
)
ht$plot

DynamicPlot(
  srt = pancreas_sub,
  lineages = c("Lineage1", "Lineage2"),
  features = c("Nnat", "Irx1"),
  group.by = "SubCellType",
  compare_lineages = TRUE,
  compare_features = FALSE
)
}
