% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunEnrichment}
\alias{RunEnrichment}
\title{Perform the enrichment analysis(over-representation) on the genes}
\usage{
RunEnrichment(
  geneID = NULL,
  geneID_groups = NULL,
  IDtype = "symbol",
  result_IDtype = "symbol",
  species = "Homo_sapiens",
  enrichment = "GO_BP",
  db_IDtype = "symbol",
  db_update = FALSE,
  db_version = "latest",
  Ensembl_version = 103,
  mirror = NULL,
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  universe = NULL,
  GO_simplify = TRUE,
  GO_simplify_padjustCutoff = 0.2,
  simplify_method = "Rel",
  simplify_similarityCutoff = 0.7,
  BPPARAM = BiocParallel::bpparam(),
  progressbar = TRUE
)
}
\arguments{
\item{geneID}{geneID}

\item{geneID_groups}{geneID_groups}

\item{IDtype}{IDtype}

\item{result_IDtype}{result_IDtype}

\item{species}{species}

\item{enrichment}{enrichment}

\item{db_update}{db_update}

\item{TERM2GENE}{TERM2GENE}

\item{TERM2NAME}{TERM2NAME}

\item{minGSSize}{minGSSize}

\item{maxGSSize}{maxGSSize}

\item{universe}{universe}

\item{GO_simplify}{GO_simplify}

\item{GO_simplify_padjustCutoff}{GO_simplify_padjustCutoff}

\item{simplify_method}{simplify_method}

\item{simplify_similarityCutoff}{simplify_similarityCutoff}
}
\description{
Perform the enrichment analysis(over-representation) on the genes
}
\examples{
if (interactive()) {
  data("pancreas1k")
  library(dplyr)
  pancreas1k <- RunDEtest(pancreas1k, group_by = "CellType")
  de_filter <- filter(pancreas1k@tools$DEtest_CellType$AllMarkers_wilcox, p_val_adj < 0.05)
  res <- RunEnrichment(geneID = de_filter$gene, geneID_groups = de_filter$group1, enrichment = "GO_BP", species = "Mus_musculus")
  EnrichmentPlot(res$enrichment, plot_type = "bar")
  EnrichmentPlot(res$enrichment, plot_type = "bar", character_width = 30, text_size_scale = 0.8)

  EnrichmentPlot(filter(res$enrichment, Groups \%in\% c("Ductal", "Endocrine")), plot_type = "lollipop", ncol = 1)
  EnrichmentPlot(filter(res$enrichment, Groups \%in\% c("Ductal", "Endocrine")), plot_type = "wordcloud")
}
}
