% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunGLMPCA}
\alias{RunGLMPCA}
\alias{RunGLMPCA.Seurat}
\alias{RunGLMPCA.Assay}
\alias{RunGLMPCA.default}
\title{Run GLMPCA}
\usage{
RunGLMPCA(object, ...)

\method{RunGLMPCA}{Seurat}(
  object,
  L = 5,
  fam = c("poi", "nb", "nb2", "binom", "mult", "bern"),
  assay = NULL,
  slot = "counts",
  features = NULL,
  reduction.name = "glmpca",
  reduction.key = "GLMPC_",
  verbose = TRUE,
  ...
)

\method{RunGLMPCA}{Assay}(
  object,
  L = 5,
  fam = c("poi", "nb", "nb2", "binom", "mult", "bern"),
  assay = NULL,
  slot = "counts",
  features = NULL,
  reduction.name = "glmpca",
  reduction.key = "GLMPC_",
  verbose = TRUE,
  ...
)

\method{RunGLMPCA}{default}(
  object,
  L = 5,
  fam = c("poi", "nb", "nb2", "binom", "mult", "bern"),
  assay = NULL,
  slot = "counts",
  features = NULL,
  reduction.key = "GLMPC_",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{}

\item{L}{The number of dimensions to return (defaults to 5)}

\item{assay}{Assay to use, defaults to the default assay}

\item{features}{A list of features to use when performing GLM-PCA. If null, defaults to variable features.}

\item{reduction.name}{Name to store resulting DimReduc object as. Defaults to glmpca}

\item{reduction.key}{Key for resulting DimReduc. Defaults to GLMPC_}
}
\value{
A Seurat object containing the output of GLMPCA stored as a DimReduc object.
}
\description{
Run GLMPCA
}
\concept{dimensional_reduction}
