% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunHarmony2}
\alias{RunHarmony2}
\alias{RunHarmony2.Seurat}
\title{Harmony single cell integration}
\usage{
RunHarmony2(object, ...)

\method{RunHarmony2}{Seurat}(
  object,
  group.by.vars,
  reduction = "pca",
  dims.use = NULL,
  theta = NULL,
  lambda = NULL,
  sigma = 0.1,
  nclust = NULL,
  tau = 0,
  block.size = 0.05,
  max.iter.harmony = 10,
  max.iter.cluster = 20,
  epsilon.cluster = 1e-05,
  epsilon.harmony = 1e-04,
  plot_convergence = FALSE,
  verbose = TRUE,
  reference_values = NULL,
  reduction.save = "Harmony",
  assay.use = "RNA",
  project.dim = TRUE,
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Run Harmony algorithm with Seurat pipelines.
}
