% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunKNNMap}
\alias{RunKNNMap}
\title{Single-cell reference mapping with KNN method}
\usage{
RunKNNMap(
  srt_query,
  srt_ref,
  ref_umap = NULL,
  force = FALSE,
  ref_group = NULL,
  features = NULL,
  nfeatures = 2000,
  query_reduction = NULL,
  ref_reduction = NULL,
  query_dims = ref_dims,
  ref_dims = 1:30,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{vote_fun}{}
}
\description{
Single-cell reference mapping with KNN method
}
\examples{
if (!require("SeuratData", quietly = TRUE)) {
  devtools::install_github("zhanghao-njmu/seurat-data")
}
library(SeuratData)
library(stringr)
suppressWarnings(InstallData("panc8"))
data("panc8")
cell_sub <- unlist(lapply(split(colnames(panc8), panc8$tech), function(x) sample(x, size = 500)))
panc8 <- subset(panc8, cells = cell_sub)

# Simply convert genes from human to mouse and preprocess the data
genenm <- make.unique(str_to_title(rownames(panc8)))
panc8 <- RenameFeatures(panc8, newnames = genenm)
head(rownames(panc8))

panc8 <- Integration_SCP(panc8, batch = "tech", integration_method = "Seurat")
ClassDimPlot(panc8, "celltype")

# Projection
data("pancreas1k")
pancreas1k <- Standard_SCP(pancreas1k)
head(intersect(rownames(panc8), rownames(pancreas1k)))
pancreas1k <- RunKNNMap(srt_query = pancreas1k, srt_ref = panc8, ref_umap = "SeuratUMAP2D")
ProjectionPlot(srt_query = pancreas1k, srt_ref = panc8, query_group = "SubCellType", ref_group = "celltype")
}
