% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunMDS}
\alias{RunMDS}
\alias{RunMDS.Seurat}
\alias{RunMDS.Assay}
\alias{RunMDS.default}
\title{Run MDS (multi-dimensional scaling)}
\usage{
RunMDS(object, ...)

\method{RunMDS}{Seurat}(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nmds = 50,
  dist.method = "euclidean",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  reduction.name = "mds",
  reduction.key = "MDS_",
  verbose = TRUE,
  seed.use = 11,
  ...
)

\method{RunMDS}{Assay}(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nmds = 50,
  dist.method = "euclidean",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  reduction.key = "MDS_",
  verbose = TRUE,
  seed.use = 11,
  ...
)

\method{RunMDS}{default}(
  object,
  assay = NULL,
  slot = "data",
  nmds = 50,
  dist.method = "euclidean",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  reduction.key = "MDS_",
  verbose = TRUE,
  seed.use = 11,
  ...
)
}
\arguments{
\item{object}{An object. This can be a Seurat object, an assay object, or a matrix-like object.}

\item{...}{Additional arguments to be passed to the stats::\link[stats]{cmdscale}, MASS::\link[MASS]{isoMDS} or MASS::\link[MASS]{sammon} function.}

\item{assay}{A character string specifying the assay to be used for the analysis. Default is NULL.}

\item{slot}{A character string specifying the slot name to be used for the analysis. Default is "data".}

\item{features}{A character vector specifying the features to be used for the analysis. Default is NULL, which uses all variable features.}

\item{nmds}{An integer specifying the number of dimensions to be computed. Default is 50.}

\item{dist.method}{A character string specifying the distance metric to be used. Currently supported values are "euclidean", "chisquared","kullback", "jeffreys", "jensen", "manhattan", "maximum", "canberra", "minkowski", and "hamming". Default is "euclidean".}

\item{mds.method}{A character string specifying the MDS algorithm to be used. Currently supported values are "cmdscale", "isoMDS", and "sammon". Default is "cmdscale".}

\item{rev.mds}{A logical value indicating whether to perform reverse MDS (i.e., transpose the input matrix) before running the analysis. Default is FALSE.}

\item{reduction.name}{A character string specifying the name of the reduction to be stored in the Seurat object. Default is "mds".}

\item{reduction.key}{A character string specifying the prefix for the column names of the basis vectors. Default is "MDS_".}

\item{verbose}{A logical value indicating whether to print verbose output. Default is TRUE.}

\item{seed.use}{An integer specifying the random seed to be used. Default is 11.}
}
\description{
Run MDS (multi-dimensional scaling)
}
\examples{
pancreas_sub <- RunMDS(object = pancreas_sub)
CellDimPlot(pancreas_sub, group.by = "CellType", reduction = "mds")

}
