% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunMonocle3}
\alias{RunMonocle3}
\title{Run Monocle3 analysis}
\usage{
RunMonocle3(
  srt,
  annotation = NULL,
  assay = NULL,
  slot = "counts",
  reduction = DefaultReduction(srt),
  clusters = NULL,
  graph = NULL,
  partition_qval = 0.05,
  k = 50,
  cluster_method = "louvain",
  num_iter = 2,
  resolution = NULL,
  use_partition = NULL,
  close_loop = TRUE,
  learn_graph_control = NULL,
  root_pr_nodes = NULL,
  root_cells = NULL,
  seed = 11
)
}
\arguments{
\item{clusters}{}
}
\description{
Run Monocle3 analysis
}
\examples{
if (interactive()) {
  data("pancreas_sub")
  # Use Monocle clusters to infer the trajectories
  pancreas_sub <- RunMonocle3(srt = pancreas_sub, annotation = "CellType")
  names(pancreas_sub@tools$Monocle3)
  trajectory <- pancreas_sub@tools$Monocle3$trajectory
  milestones <- pancreas_sub@tools$Monocle3$milestones

  p1 <- CellDimPlot(pancreas_sub, group.by = "Monocle3_partitions", reduction = "UMAP", label = TRUE, theme = "theme_blank") + trajectory + milestones
  p2 <- CellDimPlot(pancreas_sub, group.by = "Monocle3_clusters", reduction = "UMAP", label = TRUE, theme = "theme_blank") + trajectory
  p3 <- FeatureDimPlot(pancreas_sub, features = "Monocle3_Pseudotime", reduction = "UMAP", theme = "theme_blank") + trajectory
  print(p1 + p2 + p3)

  # Select the lineage using monocle3::choose_graph_segments
  cds <- pancreas_sub@tools$Monocle3$cds
  cds_sub <- monocle3::choose_graph_segments(cds, starting_pr_node = NULL, ending_pr_nodes = NULL)
  pancreas_sub$Lineages_1 <- NA
  pancreas_sub$Lineages_1[colnames(cds_sub)] <- pancreas_sub$Monocle3_Pseudotime[colnames(cds_sub)]
  CellDimPlot(pancreas_sub, group.by = "SubCellType", lineages = "Lineages_1", lineages_span = 0.1, theme = "theme_blank")

  # Use Seurat clusters to infer the trajectories
  pancreas_sub <- Standard_SCP(pancreas_sub)
  CellDimPlot(pancreas_sub, group.by = c("Standardclusters", "CellType"), label = TRUE, theme = "theme_blank")
  pancreas_sub <- RunMonocle3(srt = pancreas_sub, annotation = "CellType", clusters = "Standardclusters")
  trajectory <- pancreas_sub@tools$Monocle3$trajectory
  p1 <- CellDimPlot(pancreas_sub, group.by = "Monocle3_partitions", reduction = "StandardUMAP2D", label = TRUE, theme = "theme_blank") + trajectory
  p2 <- CellDimPlot(pancreas_sub, group.by = "Monocle3_clusters", reduction = "StandardUMAP2D", label = TRUE, theme = "theme_blank") + trajectory
  p3 <- FeatureDimPlot(pancreas_sub, features = "Monocle3_Pseudotime", reduction = "StandardUMAP2D", theme = "theme_blank") + trajectory
  print(p1 + p2 + p3)

  # Use custom graphs and cell clusters to infer the partitions and trajectories, respectively
  pancreas_sub <- Standard_SCP(pancreas_sub, cluster_resolution = 5)
  CellDimPlot(pancreas_sub, group.by = c("Standardclusters", "CellType"), label = TRUE)
  pancreas_sub <- RunMonocle3(
    srt = pancreas_sub, annotation = "CellType",
    clusters = "Standardclusters", graph = "Standardpca_SNN"
  )
  trajectory <- pancreas_sub@tools$Monocle3$trajectory
  p1 <- CellDimPlot(pancreas_sub, group.by = "Monocle3_partitions", reduction = "StandardUMAP2D", label = TRUE, theme = "theme_blank") + trajectory
  p2 <- CellDimPlot(pancreas_sub, group.by = "Monocle3_clusters", reduction = "StandardUMAP2D", label = TRUE, theme = "theme_blank") + trajectory
  p3 <- FeatureDimPlot(pancreas_sub, features = "Monocle3_Pseudotime", reduction = "StandardUMAP2D", theme = "theme_blank") + trajectory
  print(p1 + p2 + p3)
}
}
