% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunNMF}
\alias{RunNMF}
\alias{RunNMF.Seurat}
\alias{RunNMF.Assay}
\alias{RunNMF.default}
\title{Run NMF (non-negative matrix factorization)}
\usage{
RunNMF(object, ...)

\method{RunNMF}{Seurat}(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nbes = 50,
  nmf.method = "RcppML",
  tol = 1e-05,
  maxit = 100,
  rev.nmf = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "nmf",
  reduction.key = "BE_",
  verbose = TRUE,
  seed.use = 11,
  ...
)

\method{RunNMF}{Assay}(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nbes = 50,
  nmf.method = "RcppML",
  tol = 1e-05,
  maxit = 100,
  rev.nmf = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "BE_",
  verbose = TRUE,
  seed.use = 11,
  ...
)

\method{RunNMF}{default}(
  object,
  assay = NULL,
  slot = "data",
  nbes = 50,
  nmf.method = "RcppML",
  tol = 1e-05,
  maxit = 100,
  rev.nmf = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "BE_",
  verbose = TRUE,
  seed.use = 11,
  ...
)
}
\arguments{
\item{object}{An object. This can be a Seurat object, an Assay object, or a matrix-like object.}

\item{...}{Additional arguments passed to RcppML::\link[RcppML]{nmf} or NMF::\link[NMF]{nmf} function.}

\item{assay}{A character string specifying the assay to be used for the analysis. Default is NULL.}

\item{slot}{A character string specifying the slot name to be used for the analysis. Default is "data".}

\item{features}{A character vector specifying the features to be used for the analysis. Default is NULL, which uses all variable features.}

\item{nbes}{An integer specifying the number of basis vectors (components) to be computed. Default is 50.}

\item{nmf.method}{A character string specifying the NMF algorithm to be used. Currently supported values are "RcppML" and "NMF". Default is "RcppML".}

\item{tol}{A numeric value specifying the tolerance for convergence (only applicable when nmf.method is "RcppML"). Default is 1e-5.}

\item{maxit}{An integer specifying the maximum number of iterations for convergence (only applicable when nmf.method is "RcppML"). Default is 100.}

\item{rev.nmf}{A logical value indicating whether to perform reverse NMF (i.e., transpose the input matrix) before running the analysis. Default is FALSE.}

\item{ndims.print}{An integer vector specifying the dimensions (number of basis vectors) to print in the output. Default is 1:5.}

\item{nfeatures.print}{An integer specifying the number of features to print in the output. Default is 30.}

\item{reduction.name}{A character string specifying the name of the reduction to be stored in the Seurat object. Default is "nmf".}

\item{reduction.key}{A character string specifying the prefix for the column names of the basis vectors. Default is "BE_".}

\item{verbose}{A logical value indicating whether to print verbose output. Default is TRUE.}

\item{seed.use}{An integer specifying the random seed to be used. Default is 11.}
}
\description{
Run NMF (non-negative matrix factorization)
}
\examples{
pancreas_sub <- RunNMF(object = pancreas_sub)
CellDimPlot(pancreas_sub, group.by = "CellType", reduction = "nmf")

}
