% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunPCAMap}
\alias{RunPCAMap}
\title{Single-cell reference mapping with PCA method}
\usage{
RunPCAMap(
  srt_query,
  srt_ref,
  ref_pca = NULL,
  ref_dims = 1:30,
  ref_umap = NULL,
  force = FALSE,
  ref_group = NULL,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{vote_fun}{}
}
\description{
Single-cell reference mapping with PCA method
}
\examples{
data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(stringr::str_to_title(rownames(panc8_sub)))
panc8_sub <- RenameFeatures(panc8_sub, newnames = genenames)
head(rownames(panc8_sub))

panc8_sub <- Integration_SCP(panc8_sub, batch = "tech", integration_method = "Seurat")
ClassDimPlot(panc8_sub, "celltype")

# Projection
data("pancreas1k")
pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunPCAMap(srt_query = pancreas1k, srt_ref = panc8_sub, ref_pca = "Seuratpca", ref_umap = "SeuratUMAP2D")
ProjectionPlot(srt_query = pancreas1k, srt_ref = panc8_sub, query_group = "SubCellType", ref_group = "celltype")

}
