% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunPHATE}
\alias{RunPHATE}
\alias{RunPHATE.Seurat}
\alias{RunPHATE.default}
\title{Run PHATE}
\usage{
RunPHATE(object, ...)

\method{RunPHATE}{Seurat}(
  object,
  reduction = "pca",
  dims = NULL,
  features = NULL,
  assay = DefaultAssay(object = object),
  slot = "data",
  n_components = 2,
  knn = 5,
  decay = 40,
  n_landmark = 2000,
  t = "auto",
  gamma = 1,
  n_pca = 100,
  knn_dist = "euclidean",
  knn_max = NULL,
  mds = "metric",
  mds_dist = "euclidean",
  mds_solver = "sgd",
  t_max = 100,
  n_jobs = 1,
  reduction.name = "phate",
  reduction.key = "PHATE_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)

\method{RunPHATE}{default}(
  object,
  assay = NULL,
  n_components = 2,
  knn = 5,
  decay = 40,
  n_landmark = 2000,
  t = "auto",
  gamma = 1,
  n_pca = 100,
  knn_dist = "euclidean",
  knn_max = NULL,
  mds = "metric",
  mds_dist = "euclidean",
  mds_solver = "sgd",
  t_max = 100,
  n_jobs = 1,
  reduction.key = "PHATE_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}

\item{assay}{}
}
\description{
Run PHATE
}
\concept{dimensional_reduction}
