% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-app.R
\name{RunSCExplorer}
\alias{RunSCExplorer}
\title{RunSCExplorer}
\usage{
RunSCExplorer(
  base_dir = "SCExplorer",
  DataFile = "Data.hdf5",
  MetaFile = "Meta.hdf5",
  title = "SCExplorer",
  initial_dataset = NULL,
  initial_reduction = NULL,
  initial_metaname = NULL,
  initial_feature = NULL,
  initial_assay = NULL,
  initial_slot = NULL,
  initial_label = "No",
  initial_palette1 = "Paired",
  initial_palette2 = "Spectral",
  initial_theme1 = "theme_scp",
  initial_theme2 = "theme_scp",
  initial_coExp = "No",
  initial_size = 4,
  initial_ncol = 3,
  initial_arrange = "Row",
  initial_plot_dpi = 100,
  create_script = TRUE,
  style_script = require("styler", quietly = TRUE),
  overwrite = FALSE,
  return_app = TRUE
)
}
\arguments{
\item{return_app}{}
}
\description{
RunSCExplorer
}
\examples{
data("pancreas1k")
pancreas1k <- Standard_SCP(pancreas1k)
PrepareSCExplorer(pancreas1k, base_dir = "./SCExplorer")

# Create the app.R script
app <- RunSCExplorer(base_dir = "./SCExplorer", return_app = TRUE)

# Run shiny app
if (interactive()) {
  shiny::runApp(app)
}
}
