% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cell_annotation.R
\name{RunScmap}
\alias{RunScmap}
\title{Annotate single cells using scmap.}
\usage{
RunScmap(
  srt_query,
  srt_ref,
  ref_group = NULL,
  method = "scmapCluster",
  n_features = 500,
  threshold = 0.5,
  k = 10,
  query_assay = "RNA",
  ref_assay = "RNA",
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\description{
Annotate single cells using scmap.
}
\examples{
if (!require("SeuratData", quietly = TRUE)) {
  devtools::install_github("zhanghao-njmu/seurat-data")
}
library(SeuratData)
library(stringr)
suppressWarnings(InstallData("panc8"))
data("panc8")
cell_sub <- unlist(lapply(split(colnames(panc8), panc8$tech), function(x) sample(x, size = 500)))
panc8 <- subset(panc8, cells = cell_sub)

# Simply convert genes from human to mouse and preprocess the data
genenm <- make.unique(str_to_title(rownames(panc8)))
panc8 <- RenameFeatures(panc8, newnames = genenm)
panc8 <- check_srtMerge(panc8, batch = "tech")[["srtMerge"]]

# Annotation
data("pancreas1k")
pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunScmap(
  srt_query = pancreas1k, srt_ref = panc8,
  ref_group = "celltype", method = "scmapCluster"
)
ClassDimPlot(pancreas1k, group.by = "scmap_annotation")

pancreas1k <- RunScmap(
  srt_query = pancreas1k, srt_ref = panc8,
  ref_group = "celltype", method = "scmapCell"
)
ClassDimPlot(pancreas1k, group.by = "scmap_annotation")

}
