% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunSCVELO}
\alias{RunSCVELO}
\title{Run scVelo workflow}
\usage{
RunSCVELO(
  srt = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  adata = NULL,
  group_by = NULL,
  linear_reduction = NULL,
  nonlinear_reduction = NULL,
  basis = NULL,
  mode = "stochastic",
  fitting_by = "stochastic",
  magic_impute = FALSE,
  knn = 5,
  t = 2,
  min_shared_counts = 30,
  n_pcs = 30,
  n_neighbors = 30,
  stream_smooth = NULL,
  stream_density = 2,
  arrow_length = 5,
  arrow_size = 5,
  arrow_density = 0.5,
  denoise = FALSE,
  denoise_topn = 3,
  kinetics = FALSE,
  kinetics_topn = 100,
  calculate_velocity_genes = FALSE,
  top_n = 6,
  n_jobs = 1,
  palette = "Paired",
  palcolor = NULL,
  show_plot = TRUE,
  save = FALSE,
  dpi = 300,
  dirpath = "./",
  fileprefix = "",
  return_seurat = !is.null(srt)
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{assay_X}{Assay to convert as the main data matrix (X) in the anndata object.}

\item{slot_X}{Slot name for assay_X in the Seurat object.}

\item{assay_layers}{Assays to convert as layers in the anndata object.}

\item{slot_layers}{Slot names for the assay_layers in the Seurat object.}

\item{adata}{An anndata object.}

\item{group_by}{Variable to use for grouping cells in the Seurat object.}

\item{linear_reduction}{Linear reduction method to use, e.g., "PCA".}

\item{nonlinear_reduction}{Non-linear reduction method to use, e.g., "UMAP".}

\item{basis}{The basis to use for reduction, e.g., "UMAP".}

\item{mode}{Velocity estimation model to use, e.g., "stochastic".}

\item{fitting_by}{Method used to fit gene velocities, e.g., "stochastic".}

\item{magic_impute}{Flag indicating whether to perform magic imputation.}

\item{knn}{The number of nearest neighbors for magic.MAGIC.}

\item{t}{power to which the diffusion operator is powered for magic.MAGIC.}

\item{min_shared_counts}{Minimum number of counts (both unspliced and spliced) required for a gene.}

\item{n_pcs}{Number of principal components (PCs) used for velocity estimation.}

\item{n_neighbors}{Number of nearest neighbors used for velocity estimation.}

\item{stream_smooth}{Multiplication factor for scale in Gaussian kernel around grid point.}

\item{stream_density}{Controls the closeness of streamlines. When density = 2 (default), the domain is divided into a 60x60 grid, whereas density linearly scales this grid. Each cell in the grid can have, at most, one traversing streamline.}

\item{arrow_length}{Length of arrows.}

\item{arrow_size}{Size of arrows.}

\item{arrow_density}{Amount of velocities to show.}

\item{denoise}{Boolean flag indicating whether to denoise.}

\item{denoise_topn}{Number of genes with highest likelihood selected to infer velocity directions.}

\item{kinetics}{Boolean flag indicating whether to estimate RNA kinetics.}

\item{kinetics_topn}{Number of genes with highest likelihood selected to infer velocity directions.}

\item{calculate_velocity_genes}{Boolean flag indicating whether to calculate velocity genes.}

\item{top_n}{The number of top features to plot.}

\item{n_jobs}{The number of parallel jobs to run.}

\item{palette}{The palette to use for coloring cells.}

\item{palcolor}{A vector of colors to use as the palette.}

\item{show_plot}{Whether to show the PAGA plot.}

\item{save}{Whether to save the PAGA plots.}

\item{dpi}{The DPI (dots per inch) for saving the PAGA plot.}

\item{dirpath}{The directory to save the PAGA plots.}

\item{fileprefix}{The file prefix to use for the PAGA plots.}

\item{return_seurat}{Whether to return a Seurat object instead of an anndata object. Default is TRUE.}
}
\description{
scVelo is a scalable toolkit for RNA velocity analysis in single cells. This function runs scVelo workflow on a Seurat object.
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunSCVELO(srt = pancreas_sub, assay_X = "RNA", group_by = "SubCellType", linear_reduction = "PCA", nonlinear_reduction = "UMAP")
head(pancreas_sub[[]])
names(pancreas_sub@assays)

FeatureDimPlot(pancreas_sub, c("stochastic_length", "stochastic_confidence"))
FeatureDimPlot(pancreas_sub, "stochastic_pseudotime")
VelocityPlot(pancreas_sub, reduction = "UMAP", plot_type = "stream")
CellDimPlot(pancreas_sub, group.by = "SubCellType", reduction = "UMAP", pt.size = NA, velocity = "stochastic")

pancreas_sub <- Standard_SCP(pancreas_sub, normalization_method = "SCT", nonlinear_reduction = "tsne")
pancreas_sub <- RunSCVELO(srt = pancreas_sub, assay_X = "SCT", group_by = "SubCellType", linear_reduction = "Standardpca", nonlinear_reduction = "StandardTSNE2D")

}
\seealso{
\code{\link{srt_to_adata}} \code{\link{VelocityPlot}} \code{\link{CellDimPlot}} \code{\link{RunPAGA}}
}
