% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cell_annotation.R
\name{RunSingleR}
\alias{RunSingleR}
\title{Annotate single cells using scmap.}
\usage{
RunSingleR(
  srt_query,
  srt_ref,
  query_group = NULL,
  ref_group = NULL,
  genes = "de",
  de.method = "wilcox",
  sd.thresh = 1,
  de.n = NULL,
  aggr.ref = FALSE,
  aggr.args = list(),
  quantile = 0.8,
  fine.tune = TRUE,
  tune.thresh = 0.05,
  prune = TRUE,
  BPPARAM = BiocParallel::bpparam(),
  query_assay = "RNA",
  ref_assay = "RNA",
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\description{
Annotate single cells using scmap.
}
\examples{
if (!require("SeuratData", quietly = TRUE)) {
  devtools::install_github("zhanghao-njmu/seurat-data")
}
library(SeuratData)
library(stringr)
suppressWarnings(InstallData("panc8"))
data("panc8")
cell_sub <- unlist(lapply(split(colnames(panc8), panc8$tech), function(x) sample(x, size = 500)))
panc8 <- subset(panc8, cells = cell_sub)

# Simply convert genes from human to mouse and preprocess the data
genenm <- make.unique(str_to_title(rownames(panc8)))
panc8 <- RenameFeatures(panc8, newnames = genenm)
panc8 <- check_srtMerge(panc8, batch = "tech")[["srtMerge"]]

# Annotation
data("pancreas1k")
pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunSingleR(
  srt_query = pancreas1k, srt_ref = panc8,
  query_group = "Standardclusters", ref_group = "celltype",
)
ClassDimPlot(pancreas1k, group.by = "singler_annotation")

pancreas1k <- RunSingleR(
  srt_query = pancreas1k, srt_ref = panc8,
  query_group = NULL, ref_group = "celltype"
)
ClassDimPlot(pancreas1k, group.by = "singler_annotation")

}
