% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{RunSlingshot}
\alias{RunSlingshot}
\title{RunSlingshot}
\usage{
RunSlingshot(
  srt,
  group.by,
  reduction = NULL,
  start = NULL,
  end = NULL,
  prefix = NULL,
  reverse = FALSE,
  align_start = FALSE,
  show_plot = TRUE,
  lineage_palette = "Dark2",
  seed = 11,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
RunSlingshot
}
\examples{
data("pancreas1k")
pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "UMAP")

# 3D lineage
pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")
}
