% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunSymphonyMap}
\alias{RunSymphonyMap}
\title{Single-cell reference mapping with Symphony method}
\usage{
RunSymphonyMap(
  srt_query,
  srt_ref,
  ref_pca = NULL,
  ref_harmony = NULL,
  ref_umap = NULL,
  force = FALSE,
  ref_group = NULL,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\description{
Single-cell reference mapping with Symphony method
}
\examples{
if (!require("SeuratData", quietly = TRUE)) {
  devtools::install_github("satijalab/seurat-data")
}
library(SeuratData)
library(stringr)
suppressWarnings(InstallData("panc8"))
data("panc8")

# Simply convert genes from human to mouse and preprocess the data
genenm <- make.unique(str_to_title(rownames(panc8)))
panc8 <- RenameFeatures(panc8, newnames = genenm)
head(rownames(panc8))

panc8 <- Integration_SCP(panc8, batch = "tech", integration_method = "Harmony")
ClassDimPlot(panc8, "celltype")

# Projection
pancreas1k <- Standard_SCP(pancreas1k)
pancreas1k <- RunSymphonyMap(srt_query = pancreas1k, srt_ref = panc8, ref_pca = "Harmonypca", ref_harmony = "Harmony", ref_umap = "HarmonyUMAP2D")
ProjectionPlot(srt_query = pancreas1k, srt_ref = panc8, query_group = "SubCellType", ref_group = "celltype")
}
