% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunSymphonyMap}
\alias{RunSymphonyMap}
\title{Single-cell reference mapping with Symphony method}
\usage{
RunSymphonyMap(
  srt_query,
  srt_ref,
  query_assay = NULL,
  ref_assay = srt_ref[[ref_pca]]@assay.used,
  ref_pca = NULL,
  ref_harmony = NULL,
  ref_umap = NULL,
  ref_group = NULL,
  projection_method = "model",
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{vote_fun}{}
}
\description{
Single-cell reference mapping with Symphony method
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)

data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(stringr::str_to_title(rownames(panc8_sub)))
panc8_sub <- RenameFeatures(panc8_sub, newnames = genenames)
head(rownames(panc8_sub))
panc8_sub <- Integration_SCP(panc8_sub, batch = "tech", integration_method = "Harmony")
ClassDimPlot(panc8_sub, "celltype")

# Projection
pancreas_sub <- RunSymphonyMap(srt_query = pancreas_sub, srt_ref = panc8_sub, ref_pca = "Harmonypca", ref_harmony = "Harmony", ref_umap = "HarmonyUMAP2D")
ProjectionPlot(srt_query = pancreas_sub, srt_ref = panc8_sub, query_group = "SubCellType", ref_group = "celltype")

}
