% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-projection.R
\name{RunSymphonyMap}
\alias{RunSymphonyMap}
\title{Single-cell reference mapping with Symphony method}
\usage{
RunSymphonyMap(
  srt_query,
  srt_ref,
  query_assay = NULL,
  ref_assay = srt_ref[[ref_pca]]@assay.used,
  ref_pca = NULL,
  ref_harmony = NULL,
  ref_umap = NULL,
  ref_group = NULL,
  projection_method = c("model", "knn"),
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)
}
\arguments{
\item{srt_query}{An object of class Seurat storing the query cells.}

\item{srt_ref}{An object of class Seurat storing the reference cells.}

\item{query_assay}{A character string specifying the assay name for the query cells. If not provided, the default assay for the query object will be used.}

\item{ref_assay}{A character string specifying the assay name for the reference cells. If not provided, the default assay for the reference object will be used.}

\item{ref_pca}{A character string specifying the name of the PCA reduction in the reference object to use for calculating the distance metric.}

\item{ref_harmony}{A character string specifying the name of the Harmony reduction in the reference object to use for calculating the distance metric.}

\item{ref_umap}{A character string specifying the name of the UMAP reduction in the reference object. If not provided, the first UMAP reduction found in the reference object will be used.}

\item{ref_group}{A character string specifying a metadata column name in the reference object to use for grouping.}

\item{projection_method}{A character string specifying the projection method to use. Options are "model" and "knn". If "model" is selected, the function will try to use a pre-trained UMAP model in the reference object for projection. If "knn" is selected, the function will directly find the nearest neighbors using the distance metric.}

\item{nn_method}{A character string specifying the nearest neighbor search method to use. Options are "raw", "annoy", and "rann". If "raw" is selected, the function will use the brute-force method to find the nearest neighbors. If "annoy" is selected, the function will use the Annoy library for approximate nearest neighbor search. If "rann" is selected, the function will use the RANN library for approximate nearest neighbor search. If not provided, the function will choose the search method based on the size of the query and reference datasets.}

\item{k}{An integer specifying the number of nearest neighbors to find for each cell in the query object.}

\item{distance_metric}{A character string specifying the distance metric to use for calculating the pairwise distances between cells. Options include: "pearson", "spearman", "cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamman", "simple matching", and "faith". Additional distance metrics can also be used, such as "euclidean", "manhattan", "hamming", etc.}

\item{vote_fun}{A character string specifying the function to be used for aggregating the nearest neighbors in the reference object. Options are "mean", "median", "sum", "min", "max", "sd", "var", etc. If not provided, the default is "mean".}
}
\description{
Single-cell reference mapping with Symphony method
}
\examples{
data("panc8_sub")
srt_ref <- panc8_sub[, panc8_sub$tech != "fluidigmc1"]
srt_query <- panc8_sub[, panc8_sub$tech == "fluidigmc1"]
srt_ref <- Integration_SCP(srt_ref, batch = "tech", integration_method = "Harmony")
CellDimPlot(srt_ref, group.by = c("celltype", "tech"))

# Projection
srt_query <- RunSymphonyMap(srt_query = srt_query, srt_ref = srt_ref, ref_pca = "Harmonypca", ref_harmony = "Harmony", ref_umap = "HarmonyUMAP2D")
ProjectionPlot(srt_query = srt_query, srt_ref = srt_ref, query_group = "celltype", ref_group = "celltype")

}
