% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat-function.R
\name{RunUMAP2}
\alias{RunUMAP2}
\alias{RunUMAP2.Seurat}
\alias{RunUMAP2.default}
\title{Run UMAP}
\usage{
RunUMAP2(object, ...)

\method{RunUMAP2}{Seurat}(
  object,
  reduction = "pca",
  dims = NULL,
  features = NULL,
  neighbor = NULL,
  graph = NULL,
  assay = DefaultAssay(object = object),
  slot = "data",
  umap.method = "uwot",
  reduction.model = NULL,
  return.model = FALSE,
  n.neighbors = 30L,
  n.components = 2L,
  metric = "cosine",
  n.epochs = 200L,
  spread = 1,
  min.dist = 0.3,
  set.op.mix.ratio = 1,
  local.connectivity = 1L,
  negative.sample.rate = 5L,
  a = NULL,
  b = NULL,
  learning.rate = 1,
  repulsion.strength = 1,
  seed.use = 42L,
  verbose = TRUE,
  reduction.name = "umap",
  reduction.key = "UMAP_",
  ...
)

\method{RunUMAP2}{default}(
  object,
  assay = NULL,
  umap.method = "uwot",
  reduction.model = NULL,
  return.model = FALSE,
  n.neighbors = 30L,
  n.components = 2L,
  metric = "cosine",
  n.epochs = 200L,
  spread = 1,
  min.dist = 0.3,
  set.op.mix.ratio = 1,
  local.connectivity = 1L,
  negative.sample.rate = 5L,
  a = NULL,
  b = NULL,
  learning.rate = 1,
  repulsion.strength = 1,
  seed.use = 42L,
  verbose = TRUE,
  reduction.key = "UMAP_",
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Run UMAP
}
\concept{dimensional_reduction}
