% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleAnnotation}
\alias{sampleAnnotation}
\title{Single cell sample annotation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7408 rows and 6 columns.
}
\usage{
sampleAnnotation
}
\description{
A data frame with 7408 observations on the following 6 variables.
\itemize{
\item Set: a character vector
\item Channel: a character vector
\item SampleAnnotation: a character vector
\item lcbatch: a character vector
\item sortday: a character vector
\item digest: a character vector
}
}
\seealso{
\code{\link[=readSCP]{readSCP()}} to see how this file is used.
}
\keyword{datasets}
