% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp}
\alias{scp}
\title{Spatial Conformal Prediction (SCP) At a Single Input Location}
\usage{
scp(
  s0,
  s,
  Y,
  global = FALSE,
  eta = Inf,
  m = NULL,
  pred_fun = krige_pred,
  alpha = 0.05,
  dfun = "std_residual2",
  precision = 0.01
)
}
\arguments{
\item{s0}{prediction location}

\item{s}{an \eqn{n \times d}{n x d} matrix or data-frame with \eqn{d} coordinates of the \eqn{n} data locations.}

\item{Y}{a vector with \eqn{n} data values.}

\item{global}{logical; if TRUE , \code{scp} function returns the result of global spatial conformal prediction \code{\link{gscp}};
if FALSE, \code{scp} function returns the result of local spatial conformal prediction \code{\link{lscp}}.}

\item{eta}{numerical value of the kernel bandwidth for the weight schema in conformal prediction. Defauls to \eqn{Inf} meaning equal weight on surrounding \eqn{m} points.}

\item{m}{an postive integer representing the number of nearest locations used for prediction. Depends on eta.}

\item{alpha}{significance level. Defaults to 0.05.}

\item{dfun}{non-conformity measure}

\item{precision}{Defaults to 0.01.}
}
\value{

}
\description{
This function provides the conformal prediction interval for spatial location \code{s0} given obserations \code{s, Y}.
}
\examples{
N = 41; n = N^2
S = seq(0,1,length=N)
s = expand.grid(S,S)
d = as.matrix(dist(s))

theta        = c(0,3,0.1,0.7)
names(theta) = c("Nugget","PartialSill","Range","Smoothness")
C = mat_cov(d,theta)
X = t(chol(C))\%*\%rnorm(n)
Y = X^3 + rnorm(n)

s0 = c(0.5, 0.5)
idx = which(s[,1]==s0[1] & s[,2]==s0[2])
pred_fun = function(s0,s,Y,alpha=0.05) return(mean(Y))
PI = scp(s0,s[-idx,],Y[-idx],pred_fun=pred_fun, dfun="abs_residual",precision=0.1)
cat(paste("True value: ", Y[idx], "\n"))
cat(paste("Prediction Interval: [ ", PI[1], ",", PI[2], "]"))
}
