% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scp1}
\alias{scp1}
\title{Single Cell QFeatures data}
\format{
An object of class \code{QFeatures} of length 5.
}
\usage{
scp1
}
\description{
A small \link{QFeatures} object with SCoPE2 data. The object is
composed of 5 assays, including 3 PSM-level assays, 1 peptide
assay and 1 protein assay.
}
\examples{
data(scp1)
scp1
}
\keyword{datasets}
