% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{scVI_integrate}
\alias{scVI_integrate}
\title{scVI_integrate}
\usage{
scVI_integrate(
  srtMerge = NULL,
  batch = "orig.ident",
  append = TRUE,
  srtList = NULL,
  do_normalization = NULL,
  normalization_method = "logCPM",
  do_HVF_finding = TRUE,
  HVF_source = "separate",
  HVF_method = "vst",
  nHVF = 2000,
  HVF_intersect = FALSE,
  HVF_min_intersection = 1,
  HVF = NULL,
  nonlinear_reduction = "umap",
  nonlinear_reduction_dims = c(2, 3),
  nonlinear_reduction_params = list(),
  force_nonlinear_reduction = TRUE,
  do_cluster_finding = TRUE,
  cluster_algorithm = "louvain",
  cluster_resolution = 0.6,
  cluster_reorder = TRUE,
  SCVI_params = list(),
  num_threads = 8,
  seed = 11
)
}
\arguments{
\item{num_threads}{}
}
\description{
scVI_integrate
}
