% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{show_palettes}
\alias{show_palettes}
\title{Show the palettes collected by SCP}
\usage{
show_palettes(
  palettes = NULL,
  index = NULL,
  palette_names = NULL,
  return_names = TRUE,
  return_palettes = FALSE
)
}
\arguments{
\item{index}{The index of the palette in SCP palette list. Show all palettes in SCP by default.}

\item{palette_names}{Name of the palettes to be shown}

\item{return_names}{Whether to return the palette names.}

\item{return_palettes}{Whether to return the palettes.}
}
\description{
Show the palettes collected by SCP
}
\examples{
all_palettes <- show_palettes(return_palettes = TRUE)
names(all_palettes)
show_palettes(index = 1:10)
show_palettes(palette_names = c("Paired", "nejm", "simspec", "Spectral", "jet"), return_palettes = TRUE)

}
