% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{srt_to_adata}
\alias{srt_to_adata}
\title{Convert a seurat object to an anndata object using reticulate}
\usage{
srt_to_adata(
  srt,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  convert_tools = FALSE,
  convert_misc = FALSE
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{assay_X}{Assays to convert as X(main data matrix) in anndata object.}

\item{assay_layers}{Assays to convert as layers in anndata object.}

\item{slot}{slot use to convert for the slot of X and layer in anndata object.}
}
\value{
A \code{anndata} object.
}
\description{
Convert a seurat object to an anndata object using reticulate
}
\examples{
data("pancreas1k")
adata <- srt_to_adata(pancreas1k)
adata
}
