% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{srt_to_adata}
\alias{srt_to_adata}
\title{Convert a seurat object to an anndata object using reticulate}
\usage{
srt_to_adata(
  srt,
  features = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  convert_tools = FALSE,
  convert_misc = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{srt}{A Seurat object.}

\item{features}{Optional vector of features to include in the anndata object. Defaults to all features in assay_X.}

\item{assay_X}{Assay to convert as the main data matrix (X) in the anndata object.}

\item{slot_X}{Slot name for assay_X in the Seurat object.}

\item{assay_layers}{Assays to convert as layers in the anndata object.}

\item{slot_layers}{Slot names for the assay_layers in the Seurat object.}

\item{convert_tools}{Logical indicating whether to convert the tool-specific data.}

\item{convert_misc}{Logical indicating whether to convert the miscellaneous data.}

\item{verbose}{Logical indicating whether to print verbose messages during the conversion process.}
}
\value{
An anndata object.

A \code{anndata} object.
}
\description{
This function takes a Seurat object and converts it to an anndata object using the reticulate package.
}
\examples{
data("pancreas_sub")
adata <- srt_to_adata(pancreas_sub)
adata

### Or save as an h5ad file or a loom file
# adata$write_h5ad("pancreas_sub.h5ad")
# adata$write_loom("pancreas_sub.loom", write_obsm_varm = TRUE)

}
