% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{SrtAppend}
\alias{SrtAppend}
\title{Append a Seurat object to another}
\usage{
SrtAppend(
  srt_raw,
  srt_append,
  slots = slotNames(srt_append),
  pattern = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{srt_raw}{A Seurat object to be appended.}

\item{srt_append}{New Seurat object to append.}

\item{slots}{slots names.}

\item{pattern}{A character string containing a regular expression. All data with matching names will be considered for appending.}

\item{overwrite}{Whether to overwrite.}

\item{verbose}{Show messages.}
}
\description{
Append a Seurat object to another
}
