% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{SrtReorder}
\alias{SrtReorder}
\title{Reorder idents by the gene expression}
\usage{
SrtReorder(
  srt,
  features = NULL,
  reorder_by = NULL,
  slot = "data",
  assay = NULL,
  log = TRUE,
  distance_metric = "euclidean",
  reorder_FUN = "mean"
)
}
\arguments{
\item{reorder_FUN}{}
}
\description{
Reorder idents by the gene expression
}
