% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{Standard_SCP}
\alias{Standard_SCP}
\title{Standard_SCP}
\usage{
Standard_SCP(
  srt,
  prefix = "Standard",
  do_normalization = NULL,
  normalization_method = "logCPM",
  do_HVF_finding = TRUE,
  HVF_method = "vst",
  nHVF = 2000,
  HVF = NULL,
  do_scaling = TRUE,
  vars_to_regress = NULL,
  regression_model = "linear",
  liner_reduction = "pca",
  liner_reduction_dims = 100,
  liner_reduction_dims_use = NULL,
  liner_reduction_params = list(),
  force_liner_reduction = FALSE,
  nonliner_reduction = "umap",
  nonliner_reduction_dims = c(2, 3),
  nonliner_reduction_params = list(),
  force_nonliner_reduction = TRUE,
  do_cluster_finding = TRUE,
  cluster_algorithm = "louvain",
  cluster_resolution = 0.6,
  cluster_reorder = TRUE,
  seed = 11
)
}
\arguments{
\item{srt}{A \code{Seurat} object.}

\item{prefix}{The prefix used to name the result.}

\item{do_normalization}{Whether to normalize the data. If NULL, will automatically determine.}

\item{normalization_method}{Normalization method.Can be one of "logCPM", "SCT".}

\item{do_HVF_finding}{Whether to find the high variable features(HVF). If NULL, will automatically determine.}

\item{nHVF}{HVF number to use.}

\item{HVF}{Custom high variable features.}

\item{do_scaling}{Whether to scale the data. If NULL, will automatically determine.}

\item{vars_to_regress}{Variables to regress out.}

\item{regression_model}{Use a linear model or generalized linear model (poisson, negative binomial) for the regression. Options are "linear" (default), "poisson", and "negbinom".}

\item{liner_reduction}{Liner reduction method name. Can be one of "pca", "ica", "nmf", "mds", "glmpca".}

\item{liner_reduction_dims}{Dimensions to calculate when performing liner reduction.}

\item{liner_reduction_dims_use}{Which dimensions to use when performing the nonliner reduction.}

\item{nonliner_reduction}{Non-liner reduction method name. Can be one of "umap", "umap-naive", "tsne", "dm", "phate", "pacmap", "trimap", "largevis".}

\item{nonliner_reduction_dims}{Dimensions to calculate when performing non-liner reduction.}

\item{force_nonliner_reduction}{}

\item{cluster_algorithm}{Algorithm for modularity optimization when finding clusters. Can be one of "louvain", "slm", "leiden".}

\item{cluster_resolution}{Cluster resolution parameter.}

\item{cluster_reorder}{Whether to reorder the cluster names using hierarchical clustering.}

\item{seed}{Set a random seed.}
}
\value{
A \code{Seurat} object containing the result.
}
\description{
Single cell pipeline for the single dataset.
}
\examples{
data("pancreas_sub")
pancreas_sub <- Standard_SCP(srt = pancreas_sub)
ClassDimPlot(pancreas_sub, group.by = "CellType")
}
