% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{SummarizeDR}
\alias{SummarizeDR}
\title{Summary of Differential Rhythmicity Tests}
\usage{
SummarizeDR(result, test = "DRF", type = "p-value", val = 0.05, out = "long")
}
\arguments{
\item{result}{List of p-value vectors}

\item{test}{Test names}

\item{type}{Types of results (e.g. "p-value")}

\item{val}{Cutoffs}

\item{out}{"long" or "wide" format}
}
\value{
Summary table
}
\description{
Summary of Differential Rhythmicity Tests
}
