% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{targeted_power}
\alias{targeted_power}
\title{Targeted Power (for biologically meaningful effects)}
\usage{
targeted_power(
  pvals,
  ground_truth,
  effect_sizes,
  min_effect = 0.5,
  alpha = 0.05
)
}
\arguments{
\item{pvals}{P-values}

\item{ground_truth}{Logical vector of true positives}

\item{effect_sizes}{Effect sizes (r = A/σ)}

\item{min_effect}{Minimum effect size to be considered "meaningful"}

\item{alpha}{Significance threshold}
}
\value{
Targeted power (proportion of meaningful effects discovered)
}
\description{
Calculate power only for genes with effect sizes above
a meaningful threshold.
}
