% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{toTOJR}
\alias{toTOJR}
\title{Types of Joint Rhythmicity (TOJR)}
\usage{
toTOJR(
  x,
  method = "Sidak_FS",
  amp.cutoff = 0,
  alpha = 0.05,
  adjustP = TRUE,
  p.adjust.method = "BH",
  parallel.ncores = 1
)
}
\arguments{
\item{x}{Output from DCP_Rhythmicity(x1, x2) with rhythm.joint component}

\item{method}{Algorithm for categorization: "Sidak_FS", "Sidak_BS", "VDA", "AWFisher"}

\item{amp.cutoff}{Amplitude cutoff for considering genes rhythmic}

\item{alpha}{Threshold for rhythmicity p-value}

\item{adjustP}{Use adjusted p-values?}

\item{p.adjust.method}{P-value adjustment method}

\item{parallel.ncores}{Number of cores for parallel processing}
}
\value{
Vector of TOJR categories
}
\description{
Categorize genes into four types of joint rhythmicity:
arrhy, rhyI (rhythmic in group 1 only), rhyII (rhythmic in group 2 only), both
}
\examples{
\dontrun{
# Re-calculate TOJR with q-value cutoff 0.1
# rhythm.res is the output of DCP_Rhythmicity(x1, x2)
TOJR.new = toTOJR(rhythm.res, alpha = 0.1, adjustP = TRUE)
}

}
\keyword{internal}
