% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{trans_dfun}
\alias{trans_dfun}
\title{Transform non-conformity measure}
\usage{
trans_dfun(dfun)
}
\arguments{
\item{dfun}{non-conformity measure with four options.
In which, \code{"residual2"} (default) represents squared residual,
\code{"std_residual2"} represents standardized squared residual,
\code{"abs_residual"} represents absolute residual,
and \code{"std_abs_residual"} represents standardized absolute residual.}
}
\value{

}
\description{
Transform non-conformity measure
}
\keyword{internal}
