% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transferColDataToAssay}
\alias{transferColDataToAssay}
\title{Transfer the \code{colData} to an Assay}
\usage{
transferColDataToAssay(object, i)
}
\arguments{
\item{object}{A \code{QFeatures} object}

\item{i}{A \code{numeric(1)} or \code{character(1)} indicating which assay
to transfer the \code{colData} to.}
}
\value{
A \code{QFeatures} object
}
\description{
The function transfers the \code{colData} from a \code{QFeatures} object to
one of the assays it contains. The transfered data is bound to the
existing \code{colData} of the target assay.
}
\examples{
data("scp1")
colData(scp1[["peptides"]])
scp1 <- transferColDataToAssay(scp1, i = "peptides")
colData(scp1[["peptides"]])

}
