% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{VolcanoPlot}
\usage{
VolcanoPlot(
  srt,
  group_by = NULL,
  test.use = "wilcox",
  DE_threshold = "avg_log2FC > 0 & p_val_adj < 0.05",
  x_metric = "diff_pct",
  palette = "RdBu",
  palcolor = NULL,
  pt.size = 1,
  pt.alpha = 1,
  cols.highlight = "black",
  sizes.highlight = 1,
  alpha.highlight = 1,
  stroke.highlight = 0.5,
  nlabel = 5,
  features_label = NULL,
  label.fg = "black",
  label.bg = "white",
  label.bg.r = 0.1,
  label.size = 4,
  aspect.ratio = NULL,
  xlab = x_metric,
  ylab = "-log10(p-adjust)",
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE
)
}
\arguments{
\item{axis}{}
}
\description{
VolcanoPlot
}
\examples{
data("pancreas_sub")
pancreas_sub <- RunDEtest(pancreas_sub, group_by = "CellType", only.pos = FALSE)
VolcanoPlot(pancreas_sub, group_by = "CellType")
VolcanoPlot(pancreas_sub, group_by = "CellType", DE_threshold = "abs(diff_pct) > 0.3 & p_val_adj < 0.05")
VolcanoPlot(pancreas_sub, group_by = "CellType", x_metric = "avg_log2FC")

}
